Use Vue single file components
