+++ src/public/assets/scripts/annotations.js
@@ -1,2 +1,2 @@
 /*! For license information please see annotations.js.LICENSE.txt */
-!function(t){var e={};function i(n){if(e[n])return e[n].exports;var r=e[n]={i:n,l:!1,exports:{}};return t[n].call(r.exports,r,r.exports,i),r.l=!0,r.exports}i.m=t,i.c=e,i.d=function(t,e,n){i.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:n})},i.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},i.t=function(t,e){if(1&e&&(t=i(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var n=Object.create(null);if(i.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var r in t)i.d(n,r,function(e){return t[e]}.bind(null,r));return n},i.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return i.d(e,"a",e),e},i.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},i.p="/",i(i.s=1)}({1:function(t,e,i){t.exports=i("2R1P")},"2R1P":function(t,e,i){"use strict";function n(){return function(){throw new Error("Unimplemented abstract method.")}()}i.r(e);var r=0;function o(t){return t.ol_uid||(t.ol_uid=String(++r))}var s=function(t){function e(e){var i="Assertion failed. See https://openlayers.org/en/"+("v"+"5.3.0".split("-")[0])+"/doc/errors/#"+e+" for details.";t.call(this,i),this.code=e,this.name="AssertionError",this.message=i}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(Error);function a(t,e){if(!t)throw new s(e)}function h(t,e,i){return Math.min(Math.max(t,e),i)}var u="cosh"in Math?Math.cosh:function(t){var e=Math.exp(t);return(e+1/e)/2};function l(t,e,i,n,r,o){var s=r-i,a=o-n;if(0!==s||0!==a){var h=((t-i)*s+(e-n)*a)/(s*s+a*a);h>1?(i=r,n=o):h>0&&(i+=s*h,n+=a*h)}return c(t,e,i,n)}function c(t,e,i,n){var r=i-t,o=n-e;return r*r+o*o}function p(t,e,i){return t+i*(e-t)}var f=/^#([a-f0-9]{3}|[a-f0-9]{4}(?:[a-f0-9]{2}){0,2})$/i,d=/^([a-z]*)$/i;function _(t){return"string"==typeof t?t:y(t)}!function(){var t={},e=0}();function g(t){return t[0]=h(t[0]+.5|0,0,255),t[1]=h(t[1]+.5|0,0,255),t[2]=h(t[2]+.5|0,0,255),t[3]=h(t[3],0,1),t}function y(t){var e=t[0];e!=(0|e)&&(e=e+.5|0);var i=t[1];i!=(0|i)&&(i=i+.5|0);var n=t[2];return n!=(0|n)&&(n=n+.5|0),"rgba("+e+","+i+","+n+","+(void 0===t[3]?1:t[3])+")"}function v(t){return Array.isArray(t)?y(t):t}function m(t,e){var i=document.createElement("canvas");return t&&(i.width=t),e&&(i.height=e),i.getContext("2d")}var C="undefined"!=typeof navigator?navigator.userAgent.toLowerCase():"",x=(C.indexOf("firefox"),-1!==C.indexOf("safari")&&C.indexOf("chrom"),-1!==C.indexOf("webkit")&&C.indexOf("edge"),C.indexOf("macintosh"),window.devicePixelRatio,function(){var t=!1;try{t=!!document.createElement("canvas").getContext("2d").setLineDash}catch(t){}return t}()),w=(navigator,window,window,navigator.msPointerEnabled,2),E=function(){var t,e={};return function(i){if(t||(t=document.createElement("div").style),!(i in e)){t.font=i;var n=t.fontFamily;if(t.font="",!n)return null;e[i]=n.split(/,\s?/)}return e[i]}}(),F="function"==typeof Object.assign?Object.assign:function(t,e){var i=arguments;if(null==t)throw new TypeError("Cannot convert undefined or null to object");for(var n=Object(t),r=1,o=arguments.length;r<o;++r){var s=i[r];if(null!=s)for(var a in s)s.hasOwnProperty(a)&&(n[a]=s[a])}return n};function S(t){for(var e in t)delete t[e]}function k(t){var e=[];for(var i in t)e.push(t[i]);return e}function M(t){var e;for(e in t)return!1;return!e}var I=function(){this.disposed_=!1};I.prototype.dispose=function(){this.disposed_||(this.disposed_=!0,this.disposeInternal())},I.prototype.disposeInternal=function(){};var P=I;function R(t,e,i,n){for(var r,o=0,s=t.length;o<s;++o)if((r=t[o]).listener===e&&r.bindTo===i)return n&&(r.deleteIndex=o),r}function b(t,e){var i=D(t);return i?i[e]:void 0}function D(t,e){var i=t.ol_lm;return!i&&e&&(i=t.ol_lm={}),i}function A(t,e){var i=b(t,e);if(i){for(var n=0,r=i.length;n<r;++n)t.removeEventListener(e,i[n].boundListener),S(i[n]);i.length=0;var o=D(t);o&&(delete o[e],0===Object.keys(o).length&&function(t){delete t.ol_lm}(t))}}function G(t,e,i,n,r){var o=D(t,!0),s=o[e];s||(s=o[e]=[]);var a=R(s,i,n,!1);return a?r||(a.callOnce=!1):(a={bindTo:n,callOnce:!!r,listener:i,target:t,type:e},t.addEventListener(e,function(t){var e=function(e){var i=t.listener,n=t.bindTo||t.target;return t.callOnce&&T(t),i.call(n,e)};return t.boundListener=e,e}(a)),s.push(a)),a}function O(t,e,i,n){return G(t,e,i,n,!0)}function L(t,e,i,n){var r=b(t,e);if(r){var o=R(r,i,n,!0);o&&T(o)}}function T(t){if(t&&t.target){t.target.removeEventListener(t.type,t.boundListener);var e=b(t.target,t.type);if(e){var i="deleteIndex"in t?t.deleteIndex:e.indexOf(t);-1!==i&&e.splice(i,1),0===e.length&&A(t.target,t.type)}S(t)}}function X(){return!0}function Y(){return!1}function j(){}var q=function(t){this.propagationStopped,this.type=t,this.target=null};q.prototype.preventDefault=function(){this.propagationStopped=!0},q.prototype.stopPropagation=function(){this.propagationStopped=!0};var B=q,N=function(t){function e(){t.call(this),this.pendingRemovals_={},this.dispatching_={},this.listeners_={}}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.addEventListener=function(t,e){var i=this.listeners_[t];i||(i=this.listeners_[t]=[]),-1===i.indexOf(e)&&i.push(e)},e.prototype.dispatchEvent=function(t){var e="string"==typeof t?new B(t):t,i=e.type;e.target=this;var n,r=this.listeners_[i];if(r){i in this.dispatching_||(this.dispatching_[i]=0,this.pendingRemovals_[i]=0),++this.dispatching_[i];for(var o=0,s=r.length;o<s;++o)if(!1===r[o].call(this,e)||e.propagationStopped){n=!1;break}if(--this.dispatching_[i],0===this.dispatching_[i]){var a=this.pendingRemovals_[i];for(delete this.pendingRemovals_[i];a--;)this.removeEventListener(i,j);delete this.dispatching_[i]}return n}},e.prototype.disposeInternal=function(){!function(t){var e=D(t);if(e)for(var i in e)A(t,i)}(this)},e.prototype.getListeners=function(t){return this.listeners_[t]},e.prototype.hasListener=function(t){return t?t in this.listeners_:Object.keys(this.listeners_).length>0},e.prototype.removeEventListener=function(t,e){var i=this.listeners_[t];if(i){var n=i.indexOf(e);t in this.pendingRemovals_?(i[n]=j,++this.pendingRemovals_[t]):(i.splice(n,1),0===i.length&&delete this.listeners_[t])}},e}(P),W="change",V="clear",z="contextmenu",K="dblclick",U=function(t){function e(e){t.call(this),this.highWaterMark=void 0!==e?e:2048,this.count_=0,this.entries_={},this.oldest_=null,this.newest_=null}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.canExpireCache=function(){return this.getCount()>this.highWaterMark},e.prototype.clear=function(){this.count_=0,this.entries_={},this.oldest_=null,this.newest_=null,this.dispatchEvent(V)},e.prototype.containsKey=function(t){return this.entries_.hasOwnProperty(t)},e.prototype.forEach=function(t,e){for(var i=this.oldest_;i;)t.call(e,i.value_,i.key_,this),i=i.newer},e.prototype.get=function(t){var e=this.entries_[t];return a(void 0!==e,15),e===this.newest_||(e===this.oldest_?(this.oldest_=this.oldest_.newer,this.oldest_.older=null):(e.newer.older=e.older,e.older.newer=e.newer),e.newer=null,e.older=this.newest_,this.newest_.newer=e,this.newest_=e),e.value_},e.prototype.remove=function(t){var e=this.entries_[t];return a(void 0!==e,15),e===this.newest_?(this.newest_=e.older,this.newest_&&(this.newest_.newer=null)):e===this.oldest_?(this.oldest_=e.newer,this.oldest_&&(this.oldest_.older=null)):(e.newer.older=e.older,e.older.newer=e.newer),delete this.entries_[t],--this.count_,e.value_},e.prototype.getCount=function(){return this.count_},e.prototype.getKeys=function(){var t,e=new Array(this.count_),i=0;for(t=this.newest_;t;t=t.older)e[i++]=t.key_;return e},e.prototype.getValues=function(){var t,e=new Array(this.count_),i=0;for(t=this.newest_;t;t=t.older)e[i++]=t.value_;return e},e.prototype.peekLast=function(){return this.oldest_.value_},e.prototype.peekLastKey=function(){return this.oldest_.key_},e.prototype.peekFirstKey=function(){return this.newest_.key_},e.prototype.pop=function(){var t=this.oldest_;return delete this.entries_[t.key_],t.newer&&(t.newer.older=null),this.oldest_=t.newer,this.oldest_||(this.newest_=null),--this.count_,t.value_},e.prototype.replace=function(t,e){this.get(t),this.entries_[t].value_=e},e.prototype.set=function(t,e){a(!(t in this.entries_),16);var i={key_:t,newer:null,older:this.newest_,value_:e};this.newest_?this.newest_.newer=i:this.oldest_=i,this.newest_=i,this.entries_[t]=i,++this.count_},e.prototype.setSize=function(t){this.highWaterMark=t},e.prototype.prune=function(){for(;this.canExpireCache();)this.pop()},e}(N);new Array(6);var H=[0,0,0,1],J=[0,0,0,1],Z=new U,$={},Q=null,tt={};!function(){var t,e,i=$,n=["monospace","serif"],r=n.length,o="wmytzilWMYTZIL@#/&?$%10";function s(t){for(var i=et(),s=100;s<=700;s+=300){for(var a=s+" ",h=!0,u=0;u<r;++u){var l=n[u];if(i.font=a+"32px "+l,e=i.measureText(o).width,t!=l){i.font=a+"32px "+t+","+l;var c=i.measureText(o).width;h=h&&c!=e}}if(h)return!0}return!1}function a(){var e=!0;for(var n in i)i[n]<60&&(s(n)?(i[n]=60,S(tt),Q=null,Z.clear()):(++i[n],e=!1));e&&(clearInterval(t),t=void 0)}}();function et(){return Q||(Q=m(1,1)),Q}var it,nt;nt=tt;var rt=function(t){this.opacity_=t.opacity,this.rotateWithView_=t.rotateWithView,this.rotation_=t.rotation,this.scale_=t.scale};rt.prototype.clone=function(){return new rt({opacity:this.getOpacity(),scale:this.getScale(),rotation:this.getRotation(),rotateWithView:this.getRotateWithView()})},rt.prototype.getOpacity=function(){return this.opacity_},rt.prototype.getRotateWithView=function(){return this.rotateWithView_},rt.prototype.getRotation=function(){return this.rotation_},rt.prototype.getScale=function(){return this.scale_},rt.prototype.getSnapToPixel=function(){return!1},rt.prototype.getAnchor=function(){return n()},rt.prototype.getImage=function(t){return n()},rt.prototype.getHitDetectionImage=function(t){return n()},rt.prototype.getImageState=function(){return n()},rt.prototype.getImageSize=function(){return n()},rt.prototype.getHitDetectionImageSize=function(){return n()},rt.prototype.getOrigin=function(){return n()},rt.prototype.getSize=function(){return n()},rt.prototype.setOpacity=function(t){this.opacity_=t},rt.prototype.setRotateWithView=function(t){this.rotateWithView_=t},rt.prototype.setRotation=function(t){this.rotation_=t},rt.prototype.setScale=function(t){this.scale_=t},rt.prototype.setSnapToPixel=function(t){},rt.prototype.listenImageChange=function(t,e){return n()},rt.prototype.load=function(){n()},rt.prototype.unlistenImageChange=function(t,e){n()};var ot=function(t){function e(e){var i=e||{};t.call(this,{points:1/0,fill:i.fill,radius:i.radius,stroke:i.stroke,atlasManager:i.atlasManager})}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){var t=new e({fill:this.getFill()?this.getFill().clone():void 0,stroke:this.getStroke()?this.getStroke().clone():void 0,radius:this.getRadius(),atlasManager:this.atlasManager_});return t.setOpacity(this.getOpacity()),t.setScale(this.getScale()),t},e.prototype.setRadius=function(t){this.radius_=t,this.render_(this.atlasManager_)},e}(function(t){function e(e){var i=void 0!==e.rotateWithView&&e.rotateWithView;t.call(this,{opacity:1,rotateWithView:i,rotation:void 0!==e.rotation?e.rotation:0,scale:1}),this.checksums_=null,this.canvas_=null,this.hitDetectionCanvas_=null,this.fill_=void 0!==e.fill?e.fill:null,this.origin_=[0,0],this.points_=e.points,this.radius_=void 0!==e.radius?e.radius:e.radius1,this.radius2_=e.radius2,this.angle_=void 0!==e.angle?e.angle:0,this.stroke_=void 0!==e.stroke?e.stroke:null,this.anchor_=null,this.size_=null,this.imageSize_=null,this.hitDetectionImageSize_=null,this.atlasManager_=e.atlasManager,this.render_(this.atlasManager_)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){var t=new e({fill:this.getFill()?this.getFill().clone():void 0,points:this.getPoints(),radius:this.getRadius(),radius2:this.getRadius2(),angle:this.getAngle(),stroke:this.getStroke()?this.getStroke().clone():void 0,rotation:this.getRotation(),rotateWithView:this.getRotateWithView(),atlasManager:this.atlasManager_});return t.setOpacity(this.getOpacity()),t.setScale(this.getScale()),t},e.prototype.getAnchor=function(){return this.anchor_},e.prototype.getAngle=function(){return this.angle_},e.prototype.getFill=function(){return this.fill_},e.prototype.getHitDetectionImage=function(t){return this.hitDetectionCanvas_},e.prototype.getImage=function(t){return this.canvas_},e.prototype.getImageSize=function(){return this.imageSize_},e.prototype.getHitDetectionImageSize=function(){return this.hitDetectionImageSize_},e.prototype.getImageState=function(){return w},e.prototype.getOrigin=function(){return this.origin_},e.prototype.getPoints=function(){return this.points_},e.prototype.getRadius=function(){return this.radius_},e.prototype.getRadius2=function(){return this.radius2_},e.prototype.getSize=function(){return this.size_},e.prototype.getStroke=function(){return this.stroke_},e.prototype.listenImageChange=function(t,e){},e.prototype.load=function(){},e.prototype.unlistenImageChange=function(t,e){},e.prototype.render_=function(t){var e,i,n="",r="",o=0,s=null,a=0,h=0;this.stroke_&&(null===(i=this.stroke_.getColor())&&(i=J),i=v(i),void 0===(h=this.stroke_.getWidth())&&(h=1),s=this.stroke_.getLineDash(),a=this.stroke_.getLineDashOffset(),x||(s=null,a=0),void 0===(r=this.stroke_.getLineJoin())&&(r="round"),void 0===(n=this.stroke_.getLineCap())&&(n="round"),void 0===(o=this.stroke_.getMiterLimit())&&(o=10));var u=2*(this.radius_+h)+1,l={strokeStyle:i,strokeWidth:h,size:u,lineCap:n,lineDash:s,lineDashOffset:a,lineJoin:r,miterLimit:o};if(void 0===t){var c=m(u,u);this.canvas_=c.canvas,e=u=this.canvas_.width,this.draw_(l,c,0,0),this.createHitDetectionCanvas_(l)}else{u=Math.round(u);var p,f=!this.fill_;f&&(p=this.drawHitDetectionCanvas_.bind(this,l));var d=this.getChecksum(),_=t.add(d,u,u,this.draw_.bind(this,l),p);this.canvas_=_.image,this.origin_=[_.offsetX,_.offsetY],e=_.image.width,f?(this.hitDetectionCanvas_=_.hitImage,this.hitDetectionImageSize_=[_.hitImage.width,_.hitImage.height]):(this.hitDetectionCanvas_=this.canvas_,this.hitDetectionImageSize_=[e,e])}this.anchor_=[u/2,u/2],this.size_=[u,u],this.imageSize_=[e,e]},e.prototype.draw_=function(t,e,i,n){var r,o,s;e.setTransform(1,0,0,1,0,0),e.translate(i,n),e.beginPath();var a=this.points_;if(a===1/0)e.arc(t.size/2,t.size/2,this.radius_,0,2*Math.PI,!0);else{var h=void 0!==this.radius2_?this.radius2_:this.radius_;for(h!==this.radius_&&(a*=2),r=0;r<=a;r++)o=2*r*Math.PI/a-Math.PI/2+this.angle_,s=r%2==0?this.radius_:h,e.lineTo(t.size/2+s*Math.cos(o),t.size/2+s*Math.sin(o))}if(this.fill_){var u=this.fill_.getColor();null===u&&(u=H),e.fillStyle=v(u),e.fill()}this.stroke_&&(e.strokeStyle=t.strokeStyle,e.lineWidth=t.strokeWidth,t.lineDash&&(e.setLineDash(t.lineDash),e.lineDashOffset=t.lineDashOffset),e.lineCap=t.lineCap,e.lineJoin=t.lineJoin,e.miterLimit=t.miterLimit,e.stroke()),e.closePath()},e.prototype.createHitDetectionCanvas_=function(t){if(this.hitDetectionImageSize_=[t.size,t.size],this.fill_)this.hitDetectionCanvas_=this.canvas_;else{var e=m(t.size,t.size);this.hitDetectionCanvas_=e.canvas,this.drawHitDetectionCanvas_(t,e,0,0)}},e.prototype.drawHitDetectionCanvas_=function(t,e,i,n){e.setTransform(1,0,0,1,0,0),e.translate(i,n),e.beginPath();var r=this.points_;if(r===1/0)e.arc(t.size/2,t.size/2,this.radius_,0,2*Math.PI,!0);else{var o,s,a,h=void 0!==this.radius2_?this.radius2_:this.radius_;for(h!==this.radius_&&(r*=2),o=0;o<=r;o++)a=2*o*Math.PI/r-Math.PI/2+this.angle_,s=o%2==0?this.radius_:h,e.lineTo(t.size/2+s*Math.cos(a),t.size/2+s*Math.sin(a))}e.fillStyle=_(H),e.fill(),this.stroke_&&(e.strokeStyle=t.strokeStyle,e.lineWidth=t.strokeWidth,t.lineDash&&(e.setLineDash(t.lineDash),e.lineDashOffset=t.lineDashOffset),e.stroke()),e.closePath()},e.prototype.getChecksum=function(){var t=this.stroke_?this.stroke_.getChecksum():"-",e=this.fill_?this.fill_.getChecksum():"-";if(!this.checksums_||t!=this.checksums_[1]||e!=this.checksums_[2]||this.radius_!=this.checksums_[3]||this.radius2_!=this.checksums_[4]||this.angle_!=this.checksums_[5]||this.points_!=this.checksums_[6]){var i="r"+t+e+(void 0!==this.radius_?this.radius_.toString():"-")+(void 0!==this.radius2_?this.radius2_.toString():"-")+(void 0!==this.angle_?this.angle_.toString():"-")+(void 0!==this.points_?this.points_.toString():"-");this.checksums_=[i,t,e,this.radius_,this.radius2_,this.angle_,this.points_]}return this.checksums_[0]},e}(rt)),st="add",at="remove",ht="propertychange";var ut=function(t){function e(){t.call(this),this.revision_=0}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.changed=function(){++this.revision_,this.dispatchEvent(W)},e.prototype.getRevision=function(){return this.revision_},e.prototype.on=function(t,e){if(Array.isArray(t)){for(var i=t.length,n=new Array(i),r=0;r<i;++r)n[r]=G(this,t[r],e);return n}return G(this,t,e)},e.prototype.once=function(t,e){if(Array.isArray(t)){for(var i=t.length,n=new Array(i),r=0;r<i;++r)n[r]=O(this,t[r],e);return n}return O(this,t,e)},e.prototype.un=function(t,e){if(Array.isArray(t))for(var i=0,n=t.length;i<n;++i)L(this,t[i],e);else L(this,t,e)},e}(N),lt=function(t){function e(e,i,n){t.call(this,e),this.key=i,this.oldValue=n}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(B),ct={};function pt(t){return ct.hasOwnProperty(t)?ct[t]:ct[t]="change:"+t}var ft=function(t){function e(e){t.call(this),o(this),this.values_={},void 0!==e&&this.setProperties(e)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.get=function(t){var e;return this.values_.hasOwnProperty(t)&&(e=this.values_[t]),e},e.prototype.getKeys=function(){return Object.keys(this.values_)},e.prototype.getProperties=function(){return F({},this.values_)},e.prototype.notify=function(t,e){var i;i=pt(t),this.dispatchEvent(new lt(i,t,e)),i=ht,this.dispatchEvent(new lt(i,t,e))},e.prototype.set=function(t,e,i){if(i)this.values_[t]=e;else{var n=this.values_[t];this.values_[t]=e,n!==e&&this.notify(t,n)}},e.prototype.setProperties=function(t,e){for(var i in t)this.set(i,t[i],e)},e.prototype.unset=function(t,e){if(t in this.values_){var i=this.values_[t];delete this.values_[t],e||this.notify(t,i)}},e}(ut),dt="length",_t=function(t){function e(e,i){t.call(this,e),this.element=i}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(B),gt=function(t){function e(e,i){t.call(this);var n=i||{};if(this.unique_=!!n.unique,this.array_=e||[],this.unique_)for(var r=0,o=this.array_.length;r<o;++r)this.assertUnique_(this.array_[r],r);this.updateLength_()}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clear=function(){for(;this.getLength()>0;)this.pop()},e.prototype.extend=function(t){for(var e=0,i=t.length;e<i;++e)this.push(t[e]);return this},e.prototype.forEach=function(t){for(var e=this.array_,i=0,n=e.length;i<n;++i)t(e[i],i,e)},e.prototype.getArray=function(){return this.array_},e.prototype.item=function(t){return this.array_[t]},e.prototype.getLength=function(){return this.get(dt)},e.prototype.insertAt=function(t,e){this.unique_&&this.assertUnique_(e),this.array_.splice(t,0,e),this.updateLength_(),this.dispatchEvent(new _t(st,e))},e.prototype.pop=function(){return this.removeAt(this.getLength()-1)},e.prototype.push=function(t){this.unique_&&this.assertUnique_(t);var e=this.getLength();return this.insertAt(e,t),this.getLength()},e.prototype.remove=function(t){for(var e=this.array_,i=0,n=e.length;i<n;++i)if(e[i]===t)return this.removeAt(i)},e.prototype.removeAt=function(t){var e=this.array_[t];return this.array_.splice(t,1),this.updateLength_(),this.dispatchEvent(new _t(at,e)),e},e.prototype.setAt=function(t,e){var i=this.getLength();if(t<i){this.unique_&&this.assertUnique_(e,t);var n=this.array_[t];this.array_[t]=e,this.dispatchEvent(new _t(at,n)),this.dispatchEvent(new _t(st,e))}else{for(var r=i;r<t;++r)this.insertAt(r,void 0);this.insertAt(t,e)}},e.prototype.updateLength_=function(){this.set(dt,this.array_.length)},e.prototype.assertUnique_=function(t,e){for(var i=0,n=this.array_.length;i<n;++i)if(this.array_[i]===t&&i!==e)throw new s(58)},e}(ft);var yt=function(t){function e(e){if(t.call(this),this.id_=void 0,this.geometryName_="geometry",this.style_=null,this.styleFunction_=void 0,this.geometryChangeKey_=null,G(this,pt(this.geometryName_),this.handleGeometryChanged_,this),e)if("function"==typeof e.getSimplifiedGeometry){var i=e;this.setGeometry(i)}else{var n=e;this.setProperties(n)}}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){var t=new e(this.getProperties());t.setGeometryName(this.getGeometryName());var i=this.getGeometry();i&&t.setGeometry(i.clone());var n=this.getStyle();return n&&t.setStyle(n),t},e.prototype.getGeometry=function(){return this.get(this.geometryName_)},e.prototype.getId=function(){return this.id_},e.prototype.getGeometryName=function(){return this.geometryName_},e.prototype.getStyle=function(){return this.style_},e.prototype.getStyleFunction=function(){return this.styleFunction_},e.prototype.handleGeometryChange_=function(){this.changed()},e.prototype.handleGeometryChanged_=function(){this.geometryChangeKey_&&(T(this.geometryChangeKey_),this.geometryChangeKey_=null);var t=this.getGeometry();t&&(this.geometryChangeKey_=G(t,W,this.handleGeometryChange_,this)),this.changed()},e.prototype.setGeometry=function(t){this.set(this.geometryName_,t)},e.prototype.setStyle=function(t){this.style_=t,this.styleFunction_=t?function(t){if("function"==typeof t)return t;var e;Array.isArray(t)?e=t:(a("function"==typeof t.getZIndex,41),e=[t]);return function(){return e}}(t):void 0,this.changed()},e.prototype.setId=function(t){this.id_=t,this.changed()},e.prototype.setGeometryName=function(t){L(this,pt(this.geometryName_),this.handleGeometryChanged_,this),this.geometryName_=t,G(this,pt(this.geometryName_),this.handleGeometryChanged_,this),this.handleGeometryChanged_()},e}(ft),vt="singleclick",mt=K,Ct="pointerdrag",xt="pointermove",wt="pointerdown",Et="pointerup",Ft=function(t){function e(e,i,n,r,o){t.call(this,e,i,n.originalEvent,r,o),this.pointerEvent=n}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(function(t){function e(e,i,n,r,o){t.call(this,e,i,o),this.originalEvent=n,this.pixel=i.getEventPixel(n),this.coordinate=i.getCoordinateFromPixel(this.pixel),this.dragging=void 0!==r&&r}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.preventDefault=function(){t.prototype.preventDefault.call(this),this.originalEvent.preventDefault()},e.prototype.stopPropagation=function(){t.prototype.stopPropagation.call(this),this.originalEvent.stopPropagation()},e}(function(t){function e(e,i,n){t.call(this,e),this.map=i,this.frameState=void 0!==n?n:null}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(B)));function St(t,e){var i,n,r=t[0],o=t[1],s=e[0],a=e[1],h=s[0],u=s[1],l=a[0],c=a[1],p=l-h,f=c-u,d=0===p&&0===f?0:(p*(r-h)+f*(o-u))/(p*p+f*f||0);return d<=0?(i=h,n=u):d>=1?(i=l,n=c):(i=h+d*p,n=u+d*f),[i,n]}function kt(t,e){for(var i=!0,n=t.length-1;n>=0;--n)if(t[n]!=e[n]){i=!1;break}return i}function Mt(t,e){var i=t[0]-e[0],n=t[1]-e[1];return i*i+n*n}function It(t,e){return Math.sqrt(Mt(t,e))}var Pt=X,Rt=Y,bt=function(t){var e=t.originalEvent;return!e.altKey&&!(e.metaKey||e.ctrlKey)&&!e.shiftKey},Dt=function(t){var e=t.originalEvent;return!e.altKey&&!(e.metaKey||e.ctrlKey)&&e.shiftKey},At=function(t){var e=t.pointerEvent;return a(void 0!==e,56),e.isPrimary&&0===e.button},Gt=0,Ot=1,Lt=2,Tt=4,Xt=8,Yt=16;function jt(t){for(var e=Vt(),i=0,n=t.length;i<n;++i)Zt(e,t[i]);return e}function qt(t,e,i){var n,r;return(n=e<t[0]?t[0]-e:t[2]<e?e-t[2]:0)*n+(r=i<t[1]?t[1]-i:t[3]<i?i-t[3]:0)*r}function Bt(t,e){return t[0]<=e[0]&&e[2]<=t[2]&&t[1]<=e[1]&&e[3]<=t[3]}function Nt(t,e,i){return t[0]<=e&&e<=t[2]&&t[1]<=i&&i<=t[3]}function Wt(t,e){var i=t[0],n=t[1],r=t[2],o=t[3],s=e[0],a=e[1],h=Gt;return s<i?h|=Yt:s>r&&(h|=Tt),a<n?h|=Xt:a>o&&(h|=Lt),h===Gt&&(h=Ot),h}function Vt(){return[1/0,1/0,-1/0,-1/0]}function zt(t,e,i,n,r){return r?(r[0]=t,r[1]=e,r[2]=i,r[3]=n,r):[t,e,i,n]}function Kt(t){return zt(1/0,1/0,-1/0,-1/0,t)}function Ut(t,e){var i=t[0],n=t[1];return zt(i,n,i,n,e)}function Ht(t,e,i,n,r){return $t(Kt(r),t,e,i,n)}function Jt(t,e){return t[0]==e[0]&&t[2]==e[2]&&t[1]==e[1]&&t[3]==e[3]}function Zt(t,e){e[0]<t[0]&&(t[0]=e[0]),e[0]>t[2]&&(t[2]=e[0]),e[1]<t[1]&&(t[1]=e[1]),e[1]>t[3]&&(t[3]=e[1])}function $t(t,e,i,n,r){for(;i<n;i+=r)Qt(t,e[i],e[i+1]);return t}function Qt(t,e,i){t[0]=Math.min(t[0],e),t[1]=Math.min(t[1],i),t[2]=Math.max(t[2],e),t[3]=Math.max(t[3],i)}function te(t,e,i){var n;return(n=e.call(i,ee(t)))||(n=e.call(i,ie(t)))||(n=e.call(i,se(t)))?n:(n=e.call(i,oe(t)))||!1}function ee(t){return[t[0],t[1]]}function ie(t){return[t[2],t[1]]}function ne(t){return[(t[0]+t[2])/2,(t[1]+t[3])/2]}function re(t){return t[3]-t[1]}function oe(t){return[t[0],t[3]]}function se(t){return[t[2],t[3]]}function ae(t,e){return t[0]<=e[2]&&t[2]>=e[0]&&t[1]<=e[3]&&t[3]>=e[1]}var he="Point",ue="LineString",le="LinearRing",ce="Polygon",pe="MultiPoint",fe="MultiLineString",de="MultiPolygon",_e="GeometryCollection",ge="Circle",ye="Rectangle",ve="Ellipse";function me(t,e,i,n,r,o){for(var s=o||[],a=0,h=e;h<i;h+=n){var u=t[h],l=t[h+1];s[a++]=r[0]*u+r[2]*l+r[4],s[a++]=r[1]*u+r[3]*l+r[5]}return o&&s.length!=a&&(s.length=a),s}var Ce={DEGREES:"degrees",FEET:"ft",METERS:"m",PIXELS:"pixels",TILE_PIXELS:"tile-pixels",USFEET:"us-ft"},xe={};xe[Ce.DEGREES]=2*Math.PI*6370997/360,xe[Ce.FEET]=.3048,xe[Ce.METERS]=1,xe[Ce.USFEET]=1200/3937;var we=Ce,Ee=function(t){this.code_=t.code,this.units_=t.units,this.extent_=void 0!==t.extent?t.extent:null,this.worldExtent_=void 0!==t.worldExtent?t.worldExtent:null,this.axisOrientation_=void 0!==t.axisOrientation?t.axisOrientation:"enu",this.global_=void 0!==t.global&&t.global,this.canWrapX_=!(!this.global_||!this.extent_),this.getPointResolutionFunc_=t.getPointResolution,this.defaultTileGrid_=null,this.metersPerUnit_=t.metersPerUnit};Ee.prototype.canWrapX=function(){return this.canWrapX_},Ee.prototype.getCode=function(){return this.code_},Ee.prototype.getExtent=function(){return this.extent_},Ee.prototype.getUnits=function(){return this.units_},Ee.prototype.getMetersPerUnit=function(){return this.metersPerUnit_||xe[this.units_]},Ee.prototype.getWorldExtent=function(){return this.worldExtent_},Ee.prototype.getAxisOrientation=function(){return this.axisOrientation_},Ee.prototype.isGlobal=function(){return this.global_},Ee.prototype.setGlobal=function(t){this.global_=t,this.canWrapX_=!(!t||!this.extent_)},Ee.prototype.getDefaultTileGrid=function(){return this.defaultTileGrid_},Ee.prototype.setDefaultTileGrid=function(t){this.defaultTileGrid_=t},Ee.prototype.setExtent=function(t){this.extent_=t,this.canWrapX_=!(!this.global_||!t)},Ee.prototype.setWorldExtent=function(t){this.worldExtent_=t},Ee.prototype.setGetPointResolution=function(t){this.getPointResolutionFunc_=t},Ee.prototype.getPointResolutionFunc=function(){return this.getPointResolutionFunc_};var Fe=Ee,Se=6378137*Math.PI,ke=[-Se,-Se,Se,Se],Me=[-180,-85,180,85],Ie=function(t){function e(e){t.call(this,{code:e,units:we.METERS,extent:ke,global:!0,worldExtent:Me,getPointResolution:function(t,e){return t/u(e[1]/6378137)}})}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(Fe),Pe=[new Ie("EPSG:3857"),new Ie("EPSG:102100"),new Ie("EPSG:102113"),new Ie("EPSG:900913"),new Ie("urn:ogc:def:crs:EPSG:6.18:3:3857"),new Ie("urn:ogc:def:crs:EPSG::3857"),new Ie("http://www.opengis.net/gml/srs/epsg.xml#3857")];function Re(t,e,i){var n=t.length,r=i>1?i:2,o=e;void 0===o&&(o=r>2?t.slice():new Array(n));for(var s=Se,a=0;a<n;a+=r){o[a]=s*t[a]/180;var h=6378137*Math.log(Math.tan(Math.PI*(t[a+1]+90)/360));h>s?h=s:h<-s&&(h=-s),o[a+1]=h}return o}function be(t,e,i){var n=t.length,r=i>1?i:2,o=e;void 0===o&&(o=r>2?t.slice():new Array(n));for(var s=0;s<n;s+=r)o[s]=180*t[s]/Se,o[s+1]=360*Math.atan(Math.exp(t[s+1]/6378137))/Math.PI-90;return o}var De=[-180,-90,180,90],Ae=6378137*Math.PI/180,Ge=function(t){function e(e,i){t.call(this,{code:e,units:we.DEGREES,extent:De,axisOrientation:i,global:!0,metersPerUnit:Ae,worldExtent:De})}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(Fe),Oe=[new Ge("CRS:84"),new Ge("EPSG:4326","neu"),new Ge("urn:ogc:def:crs:EPSG::4326","neu"),new Ge("urn:ogc:def:crs:EPSG:6.6:4326","neu"),new Ge("urn:ogc:def:crs:OGC:1.3:CRS84"),new Ge("urn:ogc:def:crs:OGC:2:84"),new Ge("http://www.opengis.net/gml/srs/epsg.xml#4326","neu"),new Ge("urn:x-ogc:def:crs:EPSG:4326","neu")],Le={};var Te,Xe,Ye,je={};function qe(t,e,i){var n=t.getCode(),r=e.getCode();n in je||(je[n]={}),je[n][r]=i}function Be(t,e,i){var n;if(void 0!==e){for(var r=0,o=t.length;r<o;++r)e[r]=t[r];n=e}else n=t.slice();return n}function Ne(t,e,i){if(void 0!==e&&t!==e){for(var n=0,r=t.length;n<r;++n)e[n]=t[n];t=e}return t}function We(t){!function(t,e){Le[t]=e}(t.getCode(),t),qe(t,t,Be)}function Ve(t){return"string"==typeof t?Le[t]||null:t||null}function ze(t){!function(t){t.forEach(We)}(t),t.forEach((function(e){t.forEach((function(t){e!==t&&qe(e,t,Be)}))}))}function Ke(t,e){var i=function(t,e){var i;return t in je&&e in je[t]&&(i=je[t][e]),i}(t.getCode(),e.getCode());return i||(i=Ne),i}function Ue(t,e){return Ke(Ve(t),Ve(e))}ze(Pe),ze(Oe),Te=Pe,Xe=Re,Ye=be,Oe.forEach((function(t){Te.forEach((function(e){qe(t,e,Xe),qe(e,t,Ye)}))}));var He=[1,0,0,1,0,0],Je="XY",Ze="XYZ",$e="XYM",Qe="XYZM";function ti(t){var e;return t==Je?e=2:t==Ze||t==$e?e=3:t==Qe&&(e=4),e}var ei=function(t){function e(){t.call(this),this.layout=Je,this.stride=2,this.flatCoordinates=null}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.computeExtent=function(t){return Ht(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t)},e.prototype.getCoordinates=function(){return n()},e.prototype.getFirstCoordinate=function(){return this.flatCoordinates.slice(0,this.stride)},e.prototype.getFlatCoordinates=function(){return this.flatCoordinates},e.prototype.getLastCoordinate=function(){return this.flatCoordinates.slice(this.flatCoordinates.length-this.stride)},e.prototype.getLayout=function(){return this.layout},e.prototype.getSimplifiedGeometry=function(t){if(this.simplifiedGeometryRevision!=this.getRevision()&&(S(this.simplifiedGeometryCache),this.simplifiedGeometryMaxMinSquaredTolerance=0,this.simplifiedGeometryRevision=this.getRevision()),t<0||0!==this.simplifiedGeometryMaxMinSquaredTolerance&&t<=this.simplifiedGeometryMaxMinSquaredTolerance)return this;var e=t.toString();if(this.simplifiedGeometryCache.hasOwnProperty(e))return this.simplifiedGeometryCache[e];var i=this.getSimplifiedGeometryInternal(t);return i.getFlatCoordinates().length<this.flatCoordinates.length?(this.simplifiedGeometryCache[e]=i,i):(this.simplifiedGeometryMaxMinSquaredTolerance=t,this)},e.prototype.getSimplifiedGeometryInternal=function(t){return this},e.prototype.getStride=function(){return this.stride},e.prototype.setFlatCoordinates=function(t,e){this.stride=ti(t),this.layout=t,this.flatCoordinates=e},e.prototype.setCoordinates=function(t,e){n()},e.prototype.setLayout=function(t,e,i){var n;if(t)n=ti(t);else{for(var r=0;r<i;++r){if(0===e.length)return this.layout=Je,void(this.stride=2);e=e[0]}t=function(t){var e;2==t?e=Je:3==t?e=Ze:4==t&&(e=Qe);return e}(n=e.length)}this.layout=t,this.stride=n},e.prototype.applyTransform=function(t){this.flatCoordinates&&(t(this.flatCoordinates,this.flatCoordinates,this.stride),this.changed())},e.prototype.rotate=function(t,e){var i=this.getFlatCoordinates();if(i){var n=this.getStride();!function(t,e,i,n,r,o,s){for(var a=s||[],h=Math.cos(r),u=Math.sin(r),l=o[0],c=o[1],p=0,f=e;f<i;f+=n){var d=t[f]-l,_=t[f+1]-c;a[p++]=l+d*h-_*u,a[p++]=c+d*u+_*h;for(var g=f+2;g<f+n;++g)a[p++]=t[g]}s&&a.length!=p&&(a.length=p)}(i,0,i.length,n,t,e,i),this.changed()}},e.prototype.scale=function(t,e,i){var n=e;void 0===n&&(n=t);var r=i;r||(r=ne(this.getExtent()));var o=this.getFlatCoordinates();if(o){var s=this.getStride();!function(t,e,i,n,r,o,s,a){for(var h=a||[],u=s[0],l=s[1],c=0,p=e;p<i;p+=n){var f=t[p]-u,d=t[p+1]-l;h[c++]=u+r*f,h[c++]=l+o*d;for(var _=p+2;_<p+n;++_)h[c++]=t[_]}a&&h.length!=c&&(h.length=c)}(o,0,o.length,s,t,n,r,o),this.changed()}},e.prototype.translate=function(t,e){var i=this.getFlatCoordinates();if(i){var n=this.getStride();!function(t,e,i,n,r,o,s){for(var a=s||[],h=0,u=e;u<i;u+=n){a[h++]=t[u]+r,a[h++]=t[u+1]+o;for(var l=u+2;l<u+n;++l)a[h++]=t[l]}s&&a.length!=h&&(a.length=h)}(i,0,i.length,n,t,e,i),this.changed()}},e}(function(t){function e(){t.call(this),this.extent_=[1/0,1/0,-1/0,-1/0],this.extentRevision_=-1,this.simplifiedGeometryCache={},this.simplifiedGeometryMaxMinSquaredTolerance=0,this.simplifiedGeometryRevision=0}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){return n()},e.prototype.closestPointXY=function(t,e,i,r){return n()},e.prototype.containsXY=function(t,e){return!1},e.prototype.getClosestPoint=function(t,e){var i=e||[NaN,NaN];return this.closestPointXY(t[0],t[1],i,1/0),i},e.prototype.intersectsCoordinate=function(t){return this.containsXY(t[0],t[1])},e.prototype.computeExtent=function(t){return n()},e.prototype.getExtent=function(t){return this.extentRevision_!=this.getRevision()&&(this.extent_=this.computeExtent(this.extent_),this.extentRevision_=this.getRevision()),function(t,e){return e?(e[0]=t[0],e[1]=t[1],e[2]=t[2],e[3]=t[3],e):t}(this.extent_,t)},e.prototype.rotate=function(t,e){n()},e.prototype.scale=function(t,e,i){n()},e.prototype.simplify=function(t){return this.getSimplifiedGeometry(t*t)},e.prototype.getSimplifiedGeometry=function(t){return n()},e.prototype.getType=function(){return n()},e.prototype.applyTransform=function(t){n()},e.prototype.intersectsExtent=function(t){return n()},e.prototype.translate=function(t,e){n()},e.prototype.transform=function(t,e){var i=Ve(t),n=i.getUnits()==we.TILE_PIXELS?function(t,n,r){var o=i.getExtent(),s=i.getWorldExtent(),a=re(s)/re(o);return function(t,e,i,n,r,o,s,a){var h=Math.sin(o),u=Math.cos(o);t[0]=n*u,t[1]=r*h,t[2]=-n*h,t[3]=r*u,t[4]=s*n*u-a*n*h+e,t[5]=s*r*h+a*r*u+i}(He,s[0],s[3],a,-a,0,0,0),me(t,0,t.length,r,He,n),Ue(i,e)(t,n,r)}:Ue(i,e);return this.applyTransform(n),this},e}(ft));function ii(t,e,i,n){for(var r=0,o=i.length;r<o;++r)t[e++]=i[r];return e}function ni(t,e,i,n){for(var r=0,o=i.length;r<o;++r)for(var s=i[r],a=0;a<n;++a)t[e++]=s[a];return e}function ri(t,e,i,n,r){for(var o=r||[],s=0,a=0,h=i.length;a<h;++a){var u=ni(t,e,i[a],n);o[s++]=u,e=u}return o.length=s,o}var oi=function(t){function e(e,i,n){if(t.call(this),void 0!==n&&void 0===i)this.setFlatCoordinates(n,e);else{var r=i||0;this.setCenterAndRadius(e,r,n)}}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),void 0,this.layout)},e.prototype.closestPointXY=function(t,e,i,n){var r=this.flatCoordinates,o=t-r[0],s=e-r[1],a=o*o+s*s;if(a<n){if(0===a)for(var h=0;h<this.stride;++h)i[h]=r[h];else{var u=this.getRadius()/Math.sqrt(a);i[0]=r[0]+u*o,i[1]=r[1]+u*s;for(var l=2;l<this.stride;++l)i[l]=r[l]}return i.length=this.stride,a}return n},e.prototype.containsXY=function(t,e){var i=this.flatCoordinates,n=t-i[0],r=e-i[1];return n*n+r*r<=this.getRadiusSquared_()},e.prototype.getCenter=function(){return this.flatCoordinates.slice(0,this.stride)},e.prototype.computeExtent=function(t){var e=this.flatCoordinates,i=e[this.stride]-e[0];return zt(e[0]-i,e[1]-i,e[0]+i,e[1]+i,t)},e.prototype.getRadius=function(){return Math.sqrt(this.getRadiusSquared_())},e.prototype.getRadiusSquared_=function(){var t=this.flatCoordinates[this.stride]-this.flatCoordinates[0],e=this.flatCoordinates[this.stride+1]-this.flatCoordinates[1];return t*t+e*e},e.prototype.getType=function(){return ge},e.prototype.intersectsExtent=function(t){if(ae(t,this.getExtent())){var e=this.getCenter();return t[0]<=e[0]&&t[2]>=e[0]||(t[1]<=e[1]&&t[3]>=e[1]||te(t,this.intersectsCoordinate,this))}return!1},e.prototype.setCenter=function(t){var e=this.stride,i=this.flatCoordinates[e]-this.flatCoordinates[0],n=t.slice();n[e]=n[0]+i;for(var r=1;r<e;++r)n[e+r]=t[r];this.setFlatCoordinates(this.layout,n),this.changed()},e.prototype.setCenterAndRadius=function(t,e,i){this.setLayout(i,t,0),this.flatCoordinates||(this.flatCoordinates=[]);var n=this.flatCoordinates,r=ii(n,0,t,this.stride);n[r++]=n[0]+e;for(var o=1,s=this.stride;o<s;++o)n[r++]=n[o];n.length=r,this.changed()},e.prototype.getCoordinates=function(){return null},e.prototype.setCoordinates=function(t,e){},e.prototype.setRadius=function(t){this.flatCoordinates[this.stride]=this.flatCoordinates[0]+t,this.changed()},e}(ei);oi.prototype.transform;var si=oi;function ai(t,e){return t>e?1:t<e?-1:0}function hi(t,e){for(var i=Array.isArray(e)?e:[e],n=i.length,r=0;r<n;r++)t[t.length]=i[r]}function ui(t,e,i,n,r,o,s){var a,h=t[e],u=t[e+1],l=t[i]-h,c=t[i+1]-u;if(0===l&&0===c)a=e;else{var f=((r-h)*l+(o-u)*c)/(l*l+c*c);if(f>1)a=i;else{if(f>0){for(var d=0;d<n;++d)s[d]=p(t[e+d],t[i+d],f);return void(s.length=n)}a=e}}for(var _=0;_<n;++_)s[_]=t[a+_];s.length=n}function li(t,e,i,n,r){var o=t[e],s=t[e+1];for(e+=n;e<i;e+=n){var a=t[e],h=t[e+1],u=c(o,s,a,h);u>r&&(r=u),o=a,s=h}return r}function ci(t,e,i,n,r){for(var o=0,s=i.length;o<s;++o){var a=i[o];r=li(t,e,a,n,r),e=a}return r}function pi(t,e,i,n,r,o,s,a,h,u,l){if(e==i)return u;var p,f;if(0===r){if((f=c(s,a,t[e],t[e+1]))<u){for(p=0;p<n;++p)h[p]=t[e+p];return h.length=n,f}return u}for(var d=l||[NaN,NaN],_=e+n;_<i;)if(ui(t,_-n,_,n,s,a,d),(f=c(s,a,d[0],d[1]))<u){for(u=f,p=0;p<n;++p)h[p]=d[p];h.length=n,_+=n}else _+=n*Math.max((Math.sqrt(f)-Math.sqrt(u))/r|0,1);if(o&&(ui(t,i-n,e,n,s,a,d),(f=c(s,a,d[0],d[1]))<u)){for(u=f,p=0;p<n;++p)h[p]=d[p];h.length=n}return u}function fi(t,e,i,n,r,o,s,a,h,u,l){for(var c=l||[NaN,NaN],p=0,f=i.length;p<f;++p){var d=i[p];u=pi(t,e,d,n,r,o,s,a,h,u,c),e=d}return u}function di(t,e,i,n,r){for(var o=void 0!==r?r:[],s=0,a=e;a<i;a+=n)o[s++]=t.slice(a,a+n);return o.length=s,o}function _i(t,e,i,n,r){for(var o=void 0!==r?r:[],s=0,a=0,h=i.length;a<h;++a){var u=i[a];o[s++]=di(t,e,u,n,o[s]),e=u}return o.length=s,o}function gi(t,e,i,n,r,o){var s=NaN,a=NaN,h=(i-e)/n;if(1===h)s=t[e],a=t[e+1];else if(2==h)s=(1-r)*t[e]+r*t[e+n],a=(1-r)*t[e+1]+r*t[e+n+1];else if(0!==h){for(var u=t[e],l=t[e+1],c=0,f=[0],d=e+n;d<i;d+=n){var _=t[d],g=t[d+1];c+=Math.sqrt((_-u)*(_-u)+(g-l)*(g-l)),f.push(c),u=_,l=g}var y=r*c,v=function(t,e,i){for(var n,r,o=i||ai,s=0,a=t.length,h=!1;s<a;)(r=+o(t[n=s+(a-s>>1)],e))<0?s=n+1:(a=n,h=!r);return h?s:~s}(f,y);if(v<0){var m=(y-f[-v-2])/(f[-v-1]-f[-v-2]),C=e+(-v-2)*n;s=p(t[C],t[C+n],m),a=p(t[C+1],t[C+n+1],m)}else s=t[e+v*n],a=t[e+v*n+1]}return o?(o[0]=s,o[1]=a,o):[s,a]}function yi(t,e,i,n,r,o){if(i==e)return null;var s;if(r<t[e+n-1])return o?((s=t.slice(e,e+n))[n-1]=r,s):null;if(t[i-1]<r)return o?((s=t.slice(i-n,i))[n-1]=r,s):null;if(r==t[e+n-1])return t.slice(e,e+n);for(var a=e/n,h=i/n;a<h;){var u=a+h>>1;r<t[(u+1)*n-1]?h=u:a=u+1}var l=t[a*n-1];if(r==l)return t.slice((a-1)*n,(a-1)*n+n);var c=(r-l)/(t[(a+1)*n-1]-l);s=[];for(var f=0;f<n-1;++f)s.push(p(t[(a-1)*n+f],t[a*n+f],c));return s.push(r),s}function vi(t,e,i,n,r){return!te(r,(function(r){return!mi(t,e,i,n,r[0],r[1])}))}function mi(t,e,i,n,r,o){for(var s=0,a=t[i-n],h=t[i-n+1];e<i;e+=n){var u=t[e],l=t[e+1];h<=o?l>o&&(u-a)*(o-h)-(r-a)*(l-h)>0&&s++:l<=o&&(u-a)*(o-h)-(r-a)*(l-h)<0&&s--,a=u,h=l}return 0!==s}function Ci(t,e,i,n,r,o){if(0===i.length)return!1;if(!mi(t,e,i[0],n,r,o))return!1;for(var s=1,a=i.length;s<a;++s)if(mi(t,i[s-1],i[s],n,r,o))return!1;return!0}function xi(t,e,i,n,r,o){for(var s,a=[t[e],t[e+1]],h=[];e+n<i;e+=n){if(h[0]=t[e+n],h[1]=t[e+n+1],s=r.call(o,a,h))return s;a[0]=h[0],a[1]=h[1]}return!1}function wi(t,e,i,n,r){var o=$t([1/0,1/0,-1/0,-1/0],t,e,i,n);return!!ae(r,o)&&(!!Bt(r,o)||(o[0]>=r[0]&&o[2]<=r[2]||(o[1]>=r[1]&&o[3]<=r[3]||xi(t,e,i,n,(function(t,e){return function(t,e,i){var n=!1,r=Wt(t,e),o=Wt(t,i);if(r===Ot||o===Ot)n=!0;else{var s,a,h=t[0],u=t[1],l=t[2],c=t[3],p=e[0],f=e[1],d=i[0],_=i[1],g=(_-f)/(d-p);o&Lt&&!(r&Lt)&&(n=(s=d-(_-c)/g)>=h&&s<=l),n||!(o&Tt)||r&Tt||(n=(a=_-(d-l)*g)>=u&&a<=c),n||!(o&Xt)||r&Xt||(n=(s=d-(_-u)/g)>=h&&s<=l),n||!(o&Yt)||r&Yt||(n=(a=_-(d-h)*g)>=u&&a<=c)}return n}(r,t,e)})))))}function Ei(t,e,i,n,r){if(!function(t,e,i,n,r){return!!wi(t,e,i,n,r)||(!!mi(t,e,i,n,r[0],r[1])||(!!mi(t,e,i,n,r[0],r[3])||(!!mi(t,e,i,n,r[2],r[1])||!!mi(t,e,i,n,r[2],r[3]))))}(t,e,i[0],n,r))return!1;if(1===i.length)return!0;for(var o=1,s=i.length;o<s;++o)if(vi(t,i[o-1],i[o],n,r)&&!wi(t,i[o-1],i[o],n,r))return!1;return!0}function Fi(t,e,i,n){for(var r=t[e],o=t[e+1],s=0,a=e+n;a<i;a+=n){var h=t[a],u=t[a+1];s+=Math.sqrt((h-r)*(h-r)+(u-o)*(u-o)),r=h,o=u}return s}function Si(t,e,i,n,r,o,s){var a=(i-e)/n;if(a<3){for(;e<i;e+=n)o[s++]=t[e],o[s++]=t[e+1];return s}var h=new Array(a);h[0]=1,h[a-1]=1;for(var u=[e,i-n],c=0;u.length>0;){for(var p=u.pop(),f=u.pop(),d=0,_=t[f],g=t[f+1],y=t[p],v=t[p+1],m=f+n;m<p;m+=n){var C=l(t[m],t[m+1],_,g,y,v);C>d&&(c=m,d=C)}d>r&&(h[(c-e)/n]=1,f+n<c&&u.push(f,c),c+n<p&&u.push(c,p))}for(var x=0;x<a;++x)h[x]&&(o[s++]=t[e+x*n],o[s++]=t[e+x*n+1]);return s}function ki(t,e,i,n,r,o,s,a){for(var h=0,u=i.length;h<u;++h){var l=i[h];s=Si(t,e,l,n,r,o,s),a.push(s),e=l}return s}function Mi(t,e){return e*Math.round(t/e)}function Ii(t,e,i,n,r,o,s){if(e==i)return s;var a,h,u=Mi(t[e],r),l=Mi(t[e+1],r);e+=n,o[s++]=u,o[s++]=l;do{if(a=Mi(t[e],r),h=Mi(t[e+1],r),(e+=n)==i)return o[s++]=a,o[s++]=h,s}while(a==u&&h==l);for(;e<i;){var c=Mi(t[e],r),p=Mi(t[e+1],r);if(e+=n,c!=a||p!=h){var f=a-u,d=h-l,_=c-u,g=p-l;f*g==d*_&&(f<0&&_<f||f==_||f>0&&_>f)&&(d<0&&g<d||d==g||d>0&&g>d)?(a=c,h=p):(o[s++]=a,o[s++]=h,u=a,l=h,a=c,h=p)}}return o[s++]=a,o[s++]=h,s}function Pi(t,e,i,n,r,o,s,a){for(var h=0,u=i.length;h<u;++h){var l=i[h];s=Ii(t,e,l,n,r,o,s),a.push(s),e=l}return s}var Ri=function(t){function e(e,i){t.call(this),this.flatMidpoint_=null,this.flatMidpointRevision_=-1,this.maxDelta_=-1,this.maxDeltaRevision_=-1,void 0===i||Array.isArray(e[0])?this.setCoordinates(e,i):this.setFlatCoordinates(i,e)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.appendCoordinate=function(t){this.flatCoordinates?hi(this.flatCoordinates,t):this.flatCoordinates=t.slice(),this.changed()},e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),this.layout)},e.prototype.closestPointXY=function(t,e,i,n){return n<qt(this.getExtent(),t,e)?n:(this.maxDeltaRevision_!=this.getRevision()&&(this.maxDelta_=Math.sqrt(li(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,0)),this.maxDeltaRevision_=this.getRevision()),pi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,this.maxDelta_,!1,t,e,i,n))},e.prototype.forEachSegment=function(t){return xi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t)},e.prototype.getCoordinateAtM=function(t,e){if(this.layout!=$e&&this.layout!=Qe)return null;var i=void 0!==e&&e;return yi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t,i)},e.prototype.getCoordinates=function(){return di(this.flatCoordinates,0,this.flatCoordinates.length,this.stride)},e.prototype.getCoordinateAt=function(t,e){return gi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t,e)},e.prototype.getLength=function(){return Fi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride)},e.prototype.getFlatMidpoint=function(){return this.flatMidpointRevision_!=this.getRevision()&&(this.flatMidpoint_=this.getCoordinateAt(.5,this.flatMidpoint_),this.flatMidpointRevision_=this.getRevision()),this.flatMidpoint_},e.prototype.getSimplifiedGeometryInternal=function(t){var i=[];return i.length=Si(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t,i,0),new e(i,Je)},e.prototype.getType=function(){return ue},e.prototype.intersectsExtent=function(t){return wi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t)},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,1),this.flatCoordinates||(this.flatCoordinates=[]),this.flatCoordinates.length=ni(this.flatCoordinates,0,t,this.stride),this.changed()},e}(ei),bi=function(t){function e(e,i,n){if(t.call(this),this.ends_=[],this.maxDelta_=-1,this.maxDeltaRevision_=-1,Array.isArray(e[0]))this.setCoordinates(e,i);else if(void 0!==i&&n)this.setFlatCoordinates(i,e),this.ends_=n;else{for(var r=this.getLayout(),o=e,s=[],a=[],h=0,u=o.length;h<u;++h){var l=o[h];0===h&&(r=l.getLayout()),hi(s,l.getFlatCoordinates()),a.push(s.length)}this.setFlatCoordinates(r,s),this.ends_=a}}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.appendLineString=function(t){this.flatCoordinates?hi(this.flatCoordinates,t.getFlatCoordinates().slice()):this.flatCoordinates=t.getFlatCoordinates().slice(),this.ends_.push(this.flatCoordinates.length),this.changed()},e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),this.layout,this.ends_.slice())},e.prototype.closestPointXY=function(t,e,i,n){return n<qt(this.getExtent(),t,e)?n:(this.maxDeltaRevision_!=this.getRevision()&&(this.maxDelta_=Math.sqrt(ci(this.flatCoordinates,0,this.ends_,this.stride,0)),this.maxDeltaRevision_=this.getRevision()),fi(this.flatCoordinates,0,this.ends_,this.stride,this.maxDelta_,!1,t,e,i,n))},e.prototype.getCoordinateAtM=function(t,e,i){if(this.layout!=$e&&this.layout!=Qe||0===this.flatCoordinates.length)return null;var n=void 0!==e&&e,r=void 0!==i&&i;return function(t,e,i,n,r,o,s){if(s)return yi(t,e,i[i.length-1],n,r,o);var a;if(r<t[n-1])return o?((a=t.slice(0,n))[n-1]=r,a):null;if(t[t.length-1]<r)return o?((a=t.slice(t.length-n))[n-1]=r,a):null;for(var h=0,u=i.length;h<u;++h){var l=i[h];if(e!=l){if(r<t[e+n-1])return null;if(r<=t[l-1])return yi(t,e,l,n,r,!1);e=l}}return null}(this.flatCoordinates,0,this.ends_,this.stride,t,n,r)},e.prototype.getCoordinates=function(){return _i(this.flatCoordinates,0,this.ends_,this.stride)},e.prototype.getEnds=function(){return this.ends_},e.prototype.getLineString=function(t){return t<0||this.ends_.length<=t?null:new Ri(this.flatCoordinates.slice(0===t?0:this.ends_[t-1],this.ends_[t]),this.layout)},e.prototype.getLineStrings=function(){for(var t=this.flatCoordinates,e=this.ends_,i=this.layout,n=[],r=0,o=0,s=e.length;o<s;++o){var a=e[o],h=new Ri(t.slice(r,a),i);n.push(h),r=a}return n},e.prototype.getFlatMidpoints=function(){for(var t=[],e=this.flatCoordinates,i=0,n=this.ends_,r=this.stride,o=0,s=n.length;o<s;++o){var a=n[o];hi(t,gi(e,i,a,r,.5)),i=a}return t},e.prototype.getSimplifiedGeometryInternal=function(t){var i=[],n=[];return i.length=ki(this.flatCoordinates,0,this.ends_,this.stride,t,i,0,n),new e(i,Je,n)},e.prototype.getType=function(){return fe},e.prototype.intersectsExtent=function(t){return function(t,e,i,n,r){for(var o=0,s=i.length;o<s;++o){if(wi(t,e,i[o],n,r))return!0;e=i[o]}return!1}(this.flatCoordinates,0,this.ends_,this.stride,t)},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,2),this.flatCoordinates||(this.flatCoordinates=[]);var i=ri(this.flatCoordinates,0,t,this.stride,this.ends_);this.flatCoordinates.length=0===i.length?0:i[i.length-1],this.changed()},e}(ei),Di=function(t){function e(e,i){t.call(this),this.setCoordinates(e,i)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),this.layout)},e.prototype.closestPointXY=function(t,e,i,n){var r=this.flatCoordinates,o=c(t,e,r[0],r[1]);if(o<n){for(var s=this.stride,a=0;a<s;++a)i[a]=r[a];return i.length=s,o}return n},e.prototype.getCoordinates=function(){return this.flatCoordinates?this.flatCoordinates.slice():[]},e.prototype.computeExtent=function(t){return Ut(this.flatCoordinates,t)},e.prototype.getType=function(){return he},e.prototype.intersectsExtent=function(t){return Nt(t,this.flatCoordinates[0],this.flatCoordinates[1])},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,0),this.flatCoordinates||(this.flatCoordinates=[]),this.flatCoordinates.length=ii(this.flatCoordinates,0,t,this.stride),this.changed()},e}(ei),Ai=function(t){function e(e,i){t.call(this),i&&!Array.isArray(e[0])?this.setFlatCoordinates(i,e):this.setCoordinates(e,i)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.appendPoint=function(t){this.flatCoordinates?hi(this.flatCoordinates,t.getFlatCoordinates()):this.flatCoordinates=t.getFlatCoordinates().slice(),this.changed()},e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),this.layout)},e.prototype.closestPointXY=function(t,e,i,n){if(n<qt(this.getExtent(),t,e))return n;for(var r=this.flatCoordinates,o=this.stride,s=0,a=r.length;s<a;s+=o){var h=c(t,e,r[s],r[s+1]);if(h<n){n=h;for(var u=0;u<o;++u)i[u]=r[s+u];i.length=o}}return n},e.prototype.getCoordinates=function(){return di(this.flatCoordinates,0,this.flatCoordinates.length,this.stride)},e.prototype.getPoint=function(t){var e=this.flatCoordinates?this.flatCoordinates.length/this.stride:0;return t<0||e<=t?null:new Di(this.flatCoordinates.slice(t*this.stride,(t+1)*this.stride),this.layout)},e.prototype.getPoints=function(){for(var t=this.flatCoordinates,e=this.layout,i=this.stride,n=[],r=0,o=t.length;r<o;r+=i){var s=new Di(t.slice(r,r+i),e);n.push(s)}return n},e.prototype.getType=function(){return pe},e.prototype.intersectsExtent=function(t){for(var e=this.flatCoordinates,i=this.stride,n=0,r=e.length;n<r;n+=i){if(Nt(t,e[n],e[n+1]))return!0}return!1},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,1),this.flatCoordinates||(this.flatCoordinates=[]),this.flatCoordinates.length=ni(this.flatCoordinates,0,t,this.stride),this.changed()},e}(ei);function Gi(t,e,i,n){for(var r=0,o=t[i-n],s=t[i-n+1];e<i;e+=n){var a=t[e],h=t[e+1];r+=s*a-o*h,o=a,s=h}return r/2}function Oi(t,e,i,n){for(var r=0,o=0,s=i.length;o<s;++o){var a=i[o];r+=Gi(t,e,a,n),e=a}return r}var Li=function(t){function e(e,i){t.call(this),this.maxDelta_=-1,this.maxDeltaRevision_=-1,void 0===i||Array.isArray(e[0])?this.setCoordinates(e,i):this.setFlatCoordinates(i,e)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),this.layout)},e.prototype.closestPointXY=function(t,e,i,n){return n<qt(this.getExtent(),t,e)?n:(this.maxDeltaRevision_!=this.getRevision()&&(this.maxDelta_=Math.sqrt(li(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,0)),this.maxDeltaRevision_=this.getRevision()),pi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,this.maxDelta_,!0,t,e,i,n))},e.prototype.getArea=function(){return Gi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride)},e.prototype.getCoordinates=function(){return di(this.flatCoordinates,0,this.flatCoordinates.length,this.stride)},e.prototype.getSimplifiedGeometryInternal=function(t){var i=[];return i.length=Si(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t,i,0),new e(i,Je)},e.prototype.getType=function(){return le},e.prototype.intersectsExtent=function(t){return!1},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,1),this.flatCoordinates||(this.flatCoordinates=[]),this.flatCoordinates.length=ni(this.flatCoordinates,0,t,this.stride),this.changed()},e}(ei);function Ti(t,e,i,n,r,o,s){for(var a,h,u,l,c,p,f,d=r[o+1],_=[],g=0,y=i.length;g<y;++g){var v=i[g];for(l=t[v-n],p=t[v-n+1],a=e;a<v;a+=n)c=t[a],f=t[a+1],(d<=p&&f<=d||p<=d&&d<=f)&&(u=(d-p)/(f-p)*(c-l)+l,_.push(u)),l=c,p=f}var m=NaN,C=-1/0;for(_.sort(ai),l=_[0],a=1,h=_.length;a<h;++a){c=_[a];var x=Math.abs(c-l);x>C&&Ci(t,e,i,n,u=(l+c)/2,d)&&(m=u,C=x),l=c}return isNaN(m)&&(m=r[o]),s?(s.push(m,d,C),s):[m,d,C]}function Xi(t,e,i,n){for(;e<i-n;){for(var r=0;r<n;++r){var o=t[e+r];t[e+r]=t[i-n+r],t[i-n+r]=o}e+=n,i-=n}}function Yi(t,e,i,n){for(var r=0,o=t[i-n],s=t[i-n+1];e<i;e+=n){var a=t[e],h=t[e+1];r+=(a-o)*(h+s),o=a,s=h}return r>0}function ji(t,e,i,n,r){for(var o=void 0!==r&&r,s=0,a=i.length;s<a;++s){var h=i[s],u=Yi(t,e,h,n);if(0===s){if(o&&u||!o&&!u)return!1}else if(o&&!u||!o&&u)return!1;e=h}return!0}function qi(t,e,i,n,r){for(var o=void 0!==r&&r,s=0,a=i.length;s<a;++s){var h=i[s],u=Yi(t,e,h,n);(0===s?o&&u||!o&&!u:o&&!u||!o&&u)&&Xi(t,e,h,n),e=h}return e}function Bi(t,e,i,n,r){for(var o=0,s=i.length;o<s;++o)e=qi(t,e,i[o],n,r);return e}var Ni=function(t){function e(e,i,n){t.call(this),this.ends_=[],this.flatInteriorPointRevision_=-1,this.flatInteriorPoint_=null,this.maxDelta_=-1,this.maxDeltaRevision_=-1,this.orientedRevision_=-1,this.orientedFlatCoordinates_=null,void 0!==i&&n?(this.setFlatCoordinates(i,e),this.ends_=n):this.setCoordinates(e,i)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.appendLinearRing=function(t){this.flatCoordinates?hi(this.flatCoordinates,t.getFlatCoordinates()):this.flatCoordinates=t.getFlatCoordinates().slice(),this.ends_.push(this.flatCoordinates.length),this.changed()},e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),this.layout,this.ends_.slice())},e.prototype.closestPointXY=function(t,e,i,n){return n<qt(this.getExtent(),t,e)?n:(this.maxDeltaRevision_!=this.getRevision()&&(this.maxDelta_=Math.sqrt(ci(this.flatCoordinates,0,this.ends_,this.stride,0)),this.maxDeltaRevision_=this.getRevision()),fi(this.flatCoordinates,0,this.ends_,this.stride,this.maxDelta_,!0,t,e,i,n))},e.prototype.containsXY=function(t,e){return Ci(this.getOrientedFlatCoordinates(),0,this.ends_,this.stride,t,e)},e.prototype.getArea=function(){return Oi(this.getOrientedFlatCoordinates(),0,this.ends_,this.stride)},e.prototype.getCoordinates=function(t){var e;return void 0!==t?qi(e=this.getOrientedFlatCoordinates().slice(),0,this.ends_,this.stride,t):e=this.flatCoordinates,_i(e,0,this.ends_,this.stride)},e.prototype.getEnds=function(){return this.ends_},e.prototype.getFlatInteriorPoint=function(){if(this.flatInteriorPointRevision_!=this.getRevision()){var t=ne(this.getExtent());this.flatInteriorPoint_=Ti(this.getOrientedFlatCoordinates(),0,this.ends_,this.stride,t,0),this.flatInteriorPointRevision_=this.getRevision()}return this.flatInteriorPoint_},e.prototype.getInteriorPoint=function(){return new Di(this.getFlatInteriorPoint(),$e)},e.prototype.getLinearRingCount=function(){return this.ends_.length},e.prototype.getLinearRing=function(t){return t<0||this.ends_.length<=t?null:new Li(this.flatCoordinates.slice(0===t?0:this.ends_[t-1],this.ends_[t]),this.layout)},e.prototype.getLinearRings=function(){for(var t=this.layout,e=this.flatCoordinates,i=this.ends_,n=[],r=0,o=0,s=i.length;o<s;++o){var a=i[o],h=new Li(e.slice(r,a),t);n.push(h),r=a}return n},e.prototype.getOrientedFlatCoordinates=function(){if(this.orientedRevision_!=this.getRevision()){var t=this.flatCoordinates;ji(t,0,this.ends_,this.stride)?this.orientedFlatCoordinates_=t:(this.orientedFlatCoordinates_=t.slice(),this.orientedFlatCoordinates_.length=qi(this.orientedFlatCoordinates_,0,this.ends_,this.stride)),this.orientedRevision_=this.getRevision()}return this.orientedFlatCoordinates_},e.prototype.getSimplifiedGeometryInternal=function(t){var i=[],n=[];return i.length=Pi(this.flatCoordinates,0,this.ends_,this.stride,Math.sqrt(t),i,0,n),new e(i,Je,n)},e.prototype.getType=function(){return ce},e.prototype.intersectsExtent=function(t){return Ei(this.getOrientedFlatCoordinates(),0,this.ends_,this.stride,t)},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,2),this.flatCoordinates||(this.flatCoordinates=[]);var i=ri(this.flatCoordinates,0,t,this.stride,this.ends_);this.flatCoordinates.length=0===i.length?0:i[i.length-1],this.changed()},e}(ei),Wi=Ni;var Vi=function(t){function e(e,i,n){if(t.call(this),this.endss_=[],this.flatInteriorPointsRevision_=-1,this.flatInteriorPoints_=null,this.maxDelta_=-1,this.maxDeltaRevision_=-1,this.orientedRevision_=-1,this.orientedFlatCoordinates_=null,!n&&!Array.isArray(e[0])){for(var r=this.getLayout(),o=e,s=[],a=[],h=0,u=o.length;h<u;++h){var l=o[h];0===h&&(r=l.getLayout());for(var c=s.length,p=l.getEnds(),f=0,d=p.length;f<d;++f)p[f]+=c;hi(s,l.getFlatCoordinates()),a.push(p)}i=r,e=s,n=a}void 0!==i&&n?(this.setFlatCoordinates(i,e),this.endss_=n):this.setCoordinates(e,i)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.appendPolygon=function(t){var e;if(this.flatCoordinates){var i=this.flatCoordinates.length;hi(this.flatCoordinates,t.getFlatCoordinates());for(var n=0,r=(e=t.getEnds().slice()).length;n<r;++n)e[n]+=i}else this.flatCoordinates=t.getFlatCoordinates().slice(),e=t.getEnds().slice(),this.endss_.push();this.endss_.push(e),this.changed()},e.prototype.clone=function(){for(var t=this.endss_.length,i=new Array(t),n=0;n<t;++n)i[n]=this.endss_[n].slice();return new e(this.flatCoordinates.slice(),this.layout,i)},e.prototype.closestPointXY=function(t,e,i,n){return n<qt(this.getExtent(),t,e)?n:(this.maxDeltaRevision_!=this.getRevision()&&(this.maxDelta_=Math.sqrt(function(t,e,i,n,r){for(var o=0,s=i.length;o<s;++o){var a=i[o];r=ci(t,e,a,n,r),e=a[a.length-1]}return r}(this.flatCoordinates,0,this.endss_,this.stride,0)),this.maxDeltaRevision_=this.getRevision()),function(t,e,i,n,r,o,s,a,h,u,l){for(var c=l||[NaN,NaN],p=0,f=i.length;p<f;++p){var d=i[p];u=fi(t,e,d,n,r,o,s,a,h,u,c),e=d[d.length-1]}return u}(this.getOrientedFlatCoordinates(),0,this.endss_,this.stride,this.maxDelta_,!0,t,e,i,n))},e.prototype.containsXY=function(t,e){return function(t,e,i,n,r,o){if(0===i.length)return!1;for(var s=0,a=i.length;s<a;++s){var h=i[s];if(Ci(t,e,h,n,r,o))return!0;e=h[h.length-1]}return!1}(this.getOrientedFlatCoordinates(),0,this.endss_,this.stride,t,e)},e.prototype.getArea=function(){return function(t,e,i,n){for(var r=0,o=0,s=i.length;o<s;++o){var a=i[o];r+=Oi(t,e,a,n),e=a[a.length-1]}return r}(this.getOrientedFlatCoordinates(),0,this.endss_,this.stride)},e.prototype.getCoordinates=function(t){var e;return void 0!==t?Bi(e=this.getOrientedFlatCoordinates().slice(),0,this.endss_,this.stride,t):e=this.flatCoordinates,function(t,e,i,n,r){for(var o=void 0!==r?r:[],s=0,a=0,h=i.length;a<h;++a){var u=i[a];o[s++]=_i(t,e,u,n,o[s]),e=u[u.length-1]}return o.length=s,o}(e,0,this.endss_,this.stride)},e.prototype.getEndss=function(){return this.endss_},e.prototype.getFlatInteriorPoints=function(){if(this.flatInteriorPointsRevision_!=this.getRevision()){var t=function(t,e,i,n){for(var r=[],o=[1/0,1/0,-1/0,-1/0],s=0,a=i.length;s<a;++s){var h=i[s];o=Ht(t,e,h[0],n),r.push((o[0]+o[2])/2,(o[1]+o[3])/2),e=h[h.length-1]}return r}(this.flatCoordinates,0,this.endss_,this.stride);this.flatInteriorPoints_=function(t,e,i,n,r){for(var o=[],s=0,a=i.length;s<a;++s){var h=i[s];o=Ti(t,e,h,n,r,2*s,o),e=h[h.length-1]}return o}(this.getOrientedFlatCoordinates(),0,this.endss_,this.stride,t),this.flatInteriorPointsRevision_=this.getRevision()}return this.flatInteriorPoints_},e.prototype.getInteriorPoints=function(){return new Ai(this.getFlatInteriorPoints().slice(),$e)},e.prototype.getOrientedFlatCoordinates=function(){if(this.orientedRevision_!=this.getRevision()){var t=this.flatCoordinates;!function(t,e,i,n,r){for(var o=0,s=i.length;o<s;++o)if(!ji(t,e,i[o],n,r))return!1;return!0}(t,0,this.endss_,this.stride)?(this.orientedFlatCoordinates_=t.slice(),this.orientedFlatCoordinates_.length=Bi(this.orientedFlatCoordinates_,0,this.endss_,this.stride)):this.orientedFlatCoordinates_=t,this.orientedRevision_=this.getRevision()}return this.orientedFlatCoordinates_},e.prototype.getSimplifiedGeometryInternal=function(t){var i=[],n=[];return i.length=function(t,e,i,n,r,o,s,a){for(var h=0,u=i.length;h<u;++h){var l=i[h],c=[];s=Pi(t,e,l,n,r,o,s,c),a.push(c),e=l[l.length-1]}return s}(this.flatCoordinates,0,this.endss_,this.stride,Math.sqrt(t),i,0,n),new e(i,Je,n)},e.prototype.getPolygon=function(t){if(t<0||this.endss_.length<=t)return null;var e;if(0===t)e=0;else{var i=this.endss_[t-1];e=i[i.length-1]}var n=this.endss_[t].slice(),r=n[n.length-1];if(0!==e)for(var o=0,s=n.length;o<s;++o)n[o]-=e;return new Wi(this.flatCoordinates.slice(e,r),this.layout,n)},e.prototype.getPolygons=function(){for(var t=this.layout,e=this.flatCoordinates,i=this.endss_,n=[],r=0,o=0,s=i.length;o<s;++o){var a=i[o].slice(),h=a[a.length-1];if(0!==r)for(var u=0,l=a.length;u<l;++u)a[u]-=r;var c=new Wi(e.slice(r,h),t,a);n.push(c),r=h}return n},e.prototype.getType=function(){return de},e.prototype.intersectsExtent=function(t){return function(t,e,i,n,r){for(var o=0,s=i.length;o<s;++o){var a=i[o];if(Ei(t,e,a,n,r))return!0;e=a[a.length-1]}return!1}(this.getOrientedFlatCoordinates(),0,this.endss_,this.stride,t)},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,3),this.flatCoordinates||(this.flatCoordinates=[]);var i=function(t,e,i,n,r){for(var o=r||[],s=0,a=0,h=i.length;a<h;++a){var u=ri(t,e,i[a],n,o[s]);o[s++]=u,e=u[u.length-1]}return o.length=s,o}(this.flatCoordinates,0,t,this.stride,this.endss_);if(0===i.length)this.flatCoordinates.length=0;else{var n=i[i.length-1];this.flatCoordinates.length=0===n.length?0:n[n.length-1]}this.changed()},e}(ei),zi=function(t,e){this.dispatcher=t,this.mapping_=e};zi.prototype.getEvents=function(){return Object.keys(this.mapping_)},zi.prototype.getHandlerForEvent=function(t){return this.mapping_[t]};var Ki="active";var Ui=function(t){function e(e){var i=e||{};t.call(this,i),i.handleDownEvent&&(this.handleDownEvent=i.handleDownEvent),i.handleDragEvent&&(this.handleDragEvent=i.handleDragEvent),i.handleMoveEvent&&(this.handleMoveEvent=i.handleMoveEvent),i.handleUpEvent&&(this.handleUpEvent=i.handleUpEvent),i.stopDown&&(this.stopDown=i.stopDown),this.handlingDownUpSequence=!1,this.trackedPointers_={},this.targetPointers=[]}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.handleDownEvent=function(t){return!1},e.prototype.handleDragEvent=function(t){},e.prototype.handleEvent=function(t){if(!t.pointerEvent)return!0;var e=!1;if(this.updateTrackedPointers_(t),this.handlingDownUpSequence){if(t.type==Ct)this.handleDragEvent(t);else if(t.type==Et){var i=this.handleUpEvent(t);this.handlingDownUpSequence=i&&this.targetPointers.length>0}}else if(t.type==wt){var n=this.handleDownEvent(t);n&&t.preventDefault(),this.handlingDownUpSequence=n,e=this.stopDown(n)}else t.type==xt&&this.handleMoveEvent(t);return!e},e.prototype.handleMoveEvent=function(t){},e.prototype.handleUpEvent=function(t){return!1},e.prototype.stopDown=function(t){return t},e.prototype.updateTrackedPointers_=function(t){if(function(t){var e=t.type;return e===wt||e===Ct||e===Et}(t)){var e=t.pointerEvent,i=e.pointerId.toString();t.type==Et?delete this.trackedPointers_[i]:(t.type==wt||i in this.trackedPointers_)&&(this.trackedPointers_[i]=e),this.targetPointers=k(this.trackedPointers_)}},e}(function(t){function e(e){t.call(this),e.handleEvent&&(this.handleEvent=e.handleEvent),this.map_=null,this.setActive(!0)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getActive=function(){return this.get(Ki)},e.prototype.getMap=function(){return this.map_},e.prototype.handleEvent=function(t){return!0},e.prototype.setActive=function(t){this.set(Ki,t)},e.prototype.setMap=function(t){this.map_=t},e}(ft)),Hi="VECTOR",Ji="opacity",Zi="visible",$i="extent",Qi="zIndex",tn="maxResolution",en="minResolution",nn="source",rn="precompose",on="undefined",sn="ready";var an=function(t){function e(e){var i=F({},e);delete i.source,t.call(this,i),this.mapPrecomposeKey_=null,this.mapRenderKey_=null,this.sourceChangeKey_=null,e.map&&this.setMap(e.map),G(this,pt(nn),this.handleSourcePropertyChange_,this);var n=e.source?e.source:null;this.setSource(n)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getLayersArray=function(t){var e=t||[];return e.push(this),e},e.prototype.getLayerStatesArray=function(t){var e=t||[];return e.push(this.getLayerState()),e},e.prototype.getSource=function(){return this.get(nn)||null},e.prototype.getSourceState=function(){var t=this.getSource();return t?t.getState():on},e.prototype.handleSourceChange_=function(){this.changed()},e.prototype.handleSourcePropertyChange_=function(){this.sourceChangeKey_&&(T(this.sourceChangeKey_),this.sourceChangeKey_=null);var t=this.getSource();t&&(this.sourceChangeKey_=G(t,W,this.handleSourceChange_,this)),this.changed()},e.prototype.setMap=function(t){this.mapPrecomposeKey_&&(T(this.mapPrecomposeKey_),this.mapPrecomposeKey_=null),t||this.changed(),this.mapRenderKey_&&(T(this.mapRenderKey_),this.mapRenderKey_=null),t&&(this.mapPrecomposeKey_=G(t,rn,(function(t){var e=t,i=this.getLayerState();i.managed=!1,void 0===this.getZIndex()&&(i.zIndex=1/0),e.frameState.layerStatesArray.push(i),e.frameState.layerStates[o(this)]=i}),this),this.mapRenderKey_=G(this,W,t.render,t),this.changed())},e.prototype.setSource=function(t){this.set(nn,t)},e}(function(t){function e(e){t.call(this);var i=F({},e);i[Ji]=void 0!==e.opacity?e.opacity:1,i[Zi]=void 0===e.visible||e.visible,i[Qi]=e.zIndex,i[tn]=void 0!==e.maxResolution?e.maxResolution:1/0,i[en]=void 0!==e.minResolution?e.minResolution:0,this.setProperties(i),this.state_=null,this.type}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getType=function(){return this.type},e.prototype.getLayerState=function(){var t=this.state_||{layer:this,managed:!0};return t.opacity=h(this.getOpacity(),0,1),t.sourceState=this.getSourceState(),t.visible=this.getVisible(),t.extent=this.getExtent(),t.zIndex=this.getZIndex()||0,t.maxResolution=this.getMaxResolution(),t.minResolution=Math.max(this.getMinResolution(),0),this.state_=t,t},e.prototype.getLayersArray=function(t){return n()},e.prototype.getLayerStatesArray=function(t){return n()},e.prototype.getExtent=function(){return this.get($i)},e.prototype.getMaxResolution=function(){return this.get(tn)},e.prototype.getMinResolution=function(){return this.get(en)},e.prototype.getOpacity=function(){return this.get(Ji)},e.prototype.getSourceState=function(){return n()},e.prototype.getVisible=function(){return this.get(Zi)},e.prototype.getZIndex=function(){return this.get(Qi)},e.prototype.setExtent=function(t){this.set($i,t)},e.prototype.setMaxResolution=function(t){this.set(tn,t)},e.prototype.setMinResolution=function(t){this.set(en,t)},e.prototype.setOpacity=function(t){this.set(Ji,t)},e.prototype.setVisible=function(t){this.set(Zi,t)},e.prototype.setZIndex=function(t){this.set(Qi,t)},e}(ft)),hn="vector",un=function(t){var e=t||{};this.color_=void 0!==e.color?e.color:null,this.checksum_=void 0};un.prototype.clone=function(){var t=this.getColor();return new un({color:Array.isArray(t)?t.slice():t||void 0})},un.prototype.getColor=function(){return this.color_},un.prototype.setColor=function(t){this.color_=t,this.checksum_=void 0},un.prototype.getChecksum=function(){if(void 0===this.checksum_){var t=this.color_;t?Array.isArray(t)||"string"==typeof t?this.checksum_="f"+_(t):this.checksum_=o(this.color_):this.checksum_="f-"}return this.checksum_};var ln=un,cn=function(t){var e=t||{};this.color_=void 0!==e.color?e.color:null,this.lineCap_=e.lineCap,this.lineDash_=void 0!==e.lineDash?e.lineDash:null,this.lineDashOffset_=e.lineDashOffset,this.lineJoin_=e.lineJoin,this.miterLimit_=e.miterLimit,this.width_=e.width,this.checksum_=void 0};cn.prototype.clone=function(){var t=this.getColor();return new cn({color:Array.isArray(t)?t.slice():t||void 0,lineCap:this.getLineCap(),lineDash:this.getLineDash()?this.getLineDash().slice():void 0,lineDashOffset:this.getLineDashOffset(),lineJoin:this.getLineJoin(),miterLimit:this.getMiterLimit(),width:this.getWidth()})},cn.prototype.getColor=function(){return this.color_},cn.prototype.getLineCap=function(){return this.lineCap_},cn.prototype.getLineDash=function(){return this.lineDash_},cn.prototype.getLineDashOffset=function(){return this.lineDashOffset_},cn.prototype.getLineJoin=function(){return this.lineJoin_},cn.prototype.getMiterLimit=function(){return this.miterLimit_},cn.prototype.getWidth=function(){return this.width_},cn.prototype.setColor=function(t){this.color_=t,this.checksum_=void 0},cn.prototype.setLineCap=function(t){this.lineCap_=t,this.checksum_=void 0},cn.prototype.setLineDash=function(t){this.lineDash_=t,this.checksum_=void 0},cn.prototype.setLineDashOffset=function(t){this.lineDashOffset_=t,this.checksum_=void 0},cn.prototype.setLineJoin=function(t){this.lineJoin_=t,this.checksum_=void 0},cn.prototype.setMiterLimit=function(t){this.miterLimit_=t,this.checksum_=void 0},cn.prototype.setWidth=function(t){this.width_=t,this.checksum_=void 0},cn.prototype.getChecksum=function(){return void 0===this.checksum_&&(this.checksum_="s",this.color_?"string"==typeof this.color_?this.checksum_+=this.color_:this.checksum_+=o(this.color_):this.checksum_+="-",this.checksum_+=","+(void 0!==this.lineCap_?this.lineCap_.toString():"-")+","+(this.lineDash_?this.lineDash_.toString():"-")+","+(void 0!==this.lineDashOffset_?this.lineDashOffset_:"-")+","+(void 0!==this.lineJoin_?this.lineJoin_:"-")+","+(void 0!==this.miterLimit_?this.miterLimit_.toString():"-")+","+(void 0!==this.width_?this.width_.toString():"-")),this.checksum_};var pn=cn,fn=function(t){var e=t||{};this.geometry_=null,this.geometryFunction_=yn,void 0!==e.geometry&&this.setGeometry(e.geometry),this.fill_=void 0!==e.fill?e.fill:null,this.image_=void 0!==e.image?e.image:null,this.renderer_=void 0!==e.renderer?e.renderer:null,this.stroke_=void 0!==e.stroke?e.stroke:null,this.text_=void 0!==e.text?e.text:null,this.zIndex_=e.zIndex};fn.prototype.clone=function(){var t=this.getGeometry();return t&&"object"==typeof t&&(t=t.clone()),new fn({geometry:t,fill:this.getFill()?this.getFill().clone():void 0,image:this.getImage()?this.getImage().clone():void 0,stroke:this.getStroke()?this.getStroke().clone():void 0,text:this.getText()?this.getText().clone():void 0,zIndex:this.getZIndex()})},fn.prototype.getRenderer=function(){return this.renderer_},fn.prototype.setRenderer=function(t){this.renderer_=t},fn.prototype.getGeometry=function(){return this.geometry_},fn.prototype.getGeometryFunction=function(){return this.geometryFunction_},fn.prototype.getFill=function(){return this.fill_},fn.prototype.setFill=function(t){this.fill_=t},fn.prototype.getImage=function(){return this.image_},fn.prototype.setImage=function(t){this.image_=t},fn.prototype.getStroke=function(){return this.stroke_},fn.prototype.setStroke=function(t){this.stroke_=t},fn.prototype.getText=function(){return this.text_},fn.prototype.setText=function(t){this.text_=t},fn.prototype.getZIndex=function(){return this.zIndex_},fn.prototype.setGeometry=function(t){"function"==typeof t?this.geometryFunction_=t:"string"==typeof t?this.geometryFunction_=function(e){return e.get(t)}:t?void 0!==t&&(this.geometryFunction_=function(){return t}):this.geometryFunction_=yn,this.geometry_=t},fn.prototype.setZIndex=function(t){this.zIndex_=t};var dn=null;function _n(t,e){if(!dn){var i=new ln({color:"rgba(255,255,255,0.4)"}),n=new pn({color:"#3399CC",width:1.25});dn=[new fn({image:new ot({fill:i,stroke:n,radius:5}),fill:i,stroke:n})]}return dn}function gn(){var t={},e=[255,255,255,1],i=[0,153,255,1];return t[ce]=[new fn({fill:new ln({color:[255,255,255,.5]})})],t[de]=t[ce],t[ue]=[new fn({stroke:new pn({color:e,width:5})}),new fn({stroke:new pn({color:i,width:3})})],t[fe]=t[ue],t[ge]=t[ce].concat(t[ue]),t[he]=[new fn({image:new ot({radius:6,fill:new ln({color:i}),stroke:new pn({color:e,width:1.5})}),zIndex:1/0})],t[pe]=t[he],t[_e]=t[ce].concat(t[ue],t[he]),t}function yn(t){return t.getGeometry()}var vn=fn,mn="renderOrder",Cn=function(t){function e(e){var i=e||{},n=F({},i);delete n.style,delete n.renderBuffer,delete n.updateWhileAnimating,delete n.updateWhileInteracting,t.call(this,n),this.declutter_=void 0!==i.declutter&&i.declutter,this.renderBuffer_=void 0!==i.renderBuffer?i.renderBuffer:100,this.style_=null,this.styleFunction_=void 0,this.setStyle(i.style),this.updateWhileAnimating_=void 0!==i.updateWhileAnimating&&i.updateWhileAnimating,this.updateWhileInteracting_=void 0!==i.updateWhileInteracting&&i.updateWhileInteracting,this.renderMode_=i.renderMode||hn,this.type=Hi}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getDeclutter=function(){return this.declutter_},e.prototype.setDeclutter=function(t){this.declutter_=t},e.prototype.getRenderBuffer=function(){return this.renderBuffer_},e.prototype.getRenderOrder=function(){return this.get(mn)},e.prototype.getStyle=function(){return this.style_},e.prototype.getStyleFunction=function(){return this.styleFunction_},e.prototype.getUpdateWhileAnimating=function(){return this.updateWhileAnimating_},e.prototype.getUpdateWhileInteracting=function(){return this.updateWhileInteracting_},e.prototype.setRenderOrder=function(t){this.set(mn,t)},e.prototype.setStyle=function(t){this.style_=void 0!==t?t:_n,this.styleFunction_=null===t?void 0:function(t){var e;if("function"==typeof t)e=t;else{var i;if(Array.isArray(t))i=t;else a("function"==typeof t.getZIndex,41),i=[t];e=function(){return i}}return e}(this.style_),this.changed()},e.prototype.getRenderMode=function(){return this.renderMode_},e}(an);Cn.prototype.getSource;var xn=Cn,wn="arraybuffer",En="json",Fn="text",Sn="xml";function kn(t,e){return function(t,e,i,n){return function(r,o,s){var a=new XMLHttpRequest;a.open("GET","function"==typeof t?t(r,o,s):t,!0),e.getType()==wn&&(a.responseType="arraybuffer"),a.onload=function(t){if(!a.status||a.status>=200&&a.status<300){var r,o=e.getType();o==En||o==Fn?r=a.responseText:o==Sn?(r=a.responseXML)||(r=(new DOMParser).parseFromString(a.responseText,"application/xml")):o==wn&&(r=a.response),r?i.call(this,e.readFeatures(r,{featureProjection:s}),e.readProjection(r),e.getLastExtent()):n.call(this)}else n.call(this)}.bind(this),a.onerror=function(){n.call(this)}.bind(this),a.send()}}(t,e,(function(t,e){"function"==typeof this.addFeatures&&this.addFeatures(t)}),j)}function Mn(t,e){return[[-1/0,-1/0,1/0,1/0]]}function In(t){return t?Array.isArray(t)?function(e){return t}:"function"==typeof t?t:function(e){return[t]}:null}var Pn=function(t){function e(e){t.call(this),this.projection_=Ve(e.projection),this.attributions_=In(e.attributions),this.attributionsCollapsible_=void 0===e.attributionsCollapsible||e.attributionsCollapsible,this.loading=!1,this.state_=void 0!==e.state?e.state:sn,this.wrapX_=void 0!==e.wrapX&&e.wrapX}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getAttributions=function(){return this.attributions_},e.prototype.getAttributionsCollapsible=function(){return this.attributionsCollapsible_},e.prototype.getProjection=function(){return this.projection_},e.prototype.getResolutions=function(){return n()},e.prototype.getState=function(){return this.state_},e.prototype.getWrapX=function(){return this.wrapX_},e.prototype.refresh=function(){this.changed()},e.prototype.setAttributions=function(t){this.attributions_=In(t),this.changed()},e.prototype.setState=function(t){this.state_=t,this.changed()},e}(ft),Rn="addfeature",bn="changefeature",Dn="clear",An="removefeature",Gn=i("A1R+"),On=i.n(Gn),Ln=function(t){this.rbush_=On()(t,void 0),this.items_={}};Ln.prototype.insert=function(t,e){var i={minX:t[0],minY:t[1],maxX:t[2],maxY:t[3],value:e};this.rbush_.insert(i),this.items_[o(e)]=i},Ln.prototype.load=function(t,e){for(var i=new Array(e.length),n=0,r=e.length;n<r;n++){var s=t[n],a=e[n],h={minX:s[0],minY:s[1],maxX:s[2],maxY:s[3],value:a};i[n]=h,this.items_[o(a)]=h}this.rbush_.load(i)},Ln.prototype.remove=function(t){var e=o(t),i=this.items_[e];return delete this.items_[e],null!==this.rbush_.remove(i)},Ln.prototype.update=function(t,e){var i=this.items_[o(e)];Jt([i.minX,i.minY,i.maxX,i.maxY],t)||(this.remove(e),this.insert(t,e))},Ln.prototype.getAll=function(){return this.rbush_.all().map((function(t){return t.value}))},Ln.prototype.getInExtent=function(t){var e={minX:t[0],minY:t[1],maxX:t[2],maxY:t[3]};return this.rbush_.search(e).map((function(t){return t.value}))},Ln.prototype.forEach=function(t,e){return this.forEach_(this.getAll(),t,e)},Ln.prototype.forEachInExtent=function(t,e,i){return this.forEach_(this.getInExtent(t),e,i)},Ln.prototype.forEach_=function(t,e,i){for(var n,r=0,o=t.length;r<o;r++)if(n=e.call(i,t[r]))return n;return n},Ln.prototype.isEmpty=function(){return M(this.items_)},Ln.prototype.clear=function(){this.rbush_.clear(),this.items_={}},Ln.prototype.getExtent=function(t){var e=this.rbush_.toJSON();return zt(e.minX,e.minY,e.maxX,e.maxY,t)},Ln.prototype.concat=function(t){for(var e in this.rbush_.load(t.rbush_.all()),t.items_)this.items_[e]=t.items_[e]};var Tn=Ln,Xn=function(t){function e(e,i){t.call(this,e),this.feature=i}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(B),Yn=function(t){function e(e){var i=e||{};t.call(this,{attributions:i.attributions,projection:void 0,state:sn,wrapX:void 0===i.wrapX||i.wrapX}),this.loader_=j,this.format_=i.format,this.overlaps_=null==i.overlaps||i.overlaps,this.url_=i.url,void 0!==i.loader?this.loader_=i.loader:void 0!==this.url_&&(a(this.format_,7),this.loader_=kn(this.url_,this.format_)),this.strategy_=void 0!==i.strategy?i.strategy:Mn;var n,r,o=void 0===i.useSpatialIndex||i.useSpatialIndex;this.featuresRtree_=o?new Tn:null,this.loadedExtentsRtree_=new Tn,this.nullGeometryFeatures_={},this.idIndex_={},this.undefIdIndex_={},this.featureChangeKeys_={},this.featuresCollection_=null,Array.isArray(i.features)?r=i.features:i.features&&(r=(n=i.features).getArray()),o||void 0!==n||(n=new gt(r)),void 0!==r&&this.addFeaturesInternal(r),void 0!==n&&this.bindFeaturesCollection_(n)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.addFeature=function(t){this.addFeatureInternal(t),this.changed()},e.prototype.addFeatureInternal=function(t){var e=o(t);if(this.addToIndex_(e,t)){this.setupChangeEvents_(e,t);var i=t.getGeometry();if(i){var n=i.getExtent();this.featuresRtree_&&this.featuresRtree_.insert(n,t)}else this.nullGeometryFeatures_[e]=t;this.dispatchEvent(new Xn(Rn,t))}},e.prototype.setupChangeEvents_=function(t,e){this.featureChangeKeys_[t]=[G(e,W,this.handleFeatureChange_,this),G(e,ht,this.handleFeatureChange_,this)]},e.prototype.addToIndex_=function(t,e){var i=!0,n=e.getId();return void 0!==n?n.toString()in this.idIndex_?i=!1:this.idIndex_[n.toString()]=e:(a(!(t in this.undefIdIndex_),30),this.undefIdIndex_[t]=e),i},e.prototype.addFeatures=function(t){this.addFeaturesInternal(t),this.changed()},e.prototype.addFeaturesInternal=function(t){for(var e=[],i=[],n=[],r=0,s=t.length;r<s;r++){var a=t[r],h=o(a);this.addToIndex_(h,a)&&i.push(a)}for(var u=0,l=i.length;u<l;u++){var c=i[u],p=o(c);this.setupChangeEvents_(p,c);var f=c.getGeometry();if(f){var d=f.getExtent();e.push(d),n.push(c)}else this.nullGeometryFeatures_[p]=c}this.featuresRtree_&&this.featuresRtree_.load(e,n);for(var _=0,g=i.length;_<g;_++)this.dispatchEvent(new Xn(Rn,i[_]))},e.prototype.bindFeaturesCollection_=function(t){var e=!1;G(this,Rn,(function(i){e||(e=!0,t.push(i.feature),e=!1)})),G(this,An,(function(i){e||(e=!0,t.remove(i.feature),e=!1)})),G(t,st,(function(t){e||(e=!0,this.addFeature(t.element),e=!1)}),this),G(t,at,(function(t){e||(e=!0,this.removeFeature(t.element),e=!1)}),this),this.featuresCollection_=t},e.prototype.clear=function(t){if(t){for(var e in this.featureChangeKeys_){this.featureChangeKeys_[e].forEach(T)}this.featuresCollection_||(this.featureChangeKeys_={},this.idIndex_={},this.undefIdIndex_={})}else if(this.featuresRtree_)for(var i in this.featuresRtree_.forEach(this.removeFeatureInternal,this),this.nullGeometryFeatures_)this.removeFeatureInternal(this.nullGeometryFeatures_[i]);this.featuresCollection_&&this.featuresCollection_.clear(),this.featuresRtree_&&this.featuresRtree_.clear(),this.loadedExtentsRtree_.clear(),this.nullGeometryFeatures_={};var n=new Xn(Dn);this.dispatchEvent(n),this.changed()},e.prototype.forEachFeature=function(t){if(this.featuresRtree_)return this.featuresRtree_.forEach(t);this.featuresCollection_&&this.featuresCollection_.forEach(t)},e.prototype.forEachFeatureAtCoordinateDirect=function(t,e){var i=[t[0],t[1],t[0],t[1]];return this.forEachFeatureInExtent(i,(function(i){return i.getGeometry().intersectsCoordinate(t)?e(i):void 0}))},e.prototype.forEachFeatureInExtent=function(t,e){if(this.featuresRtree_)return this.featuresRtree_.forEachInExtent(t,e);this.featuresCollection_&&this.featuresCollection_.forEach(e)},e.prototype.forEachFeatureIntersectingExtent=function(t,e){return this.forEachFeatureInExtent(t,(function(i){if(i.getGeometry().intersectsExtent(t)){var n=e(i);if(n)return n}}))},e.prototype.getFeaturesCollection=function(){return this.featuresCollection_},e.prototype.getFeatures=function(){var t;return this.featuresCollection_?t=this.featuresCollection_.getArray():this.featuresRtree_&&(t=this.featuresRtree_.getAll(),M(this.nullGeometryFeatures_)||hi(t,k(this.nullGeometryFeatures_))),t},e.prototype.getFeaturesAtCoordinate=function(t){var e=[];return this.forEachFeatureAtCoordinateDirect(t,(function(t){e.push(t)})),e},e.prototype.getFeaturesInExtent=function(t){return this.featuresRtree_.getInExtent(t)},e.prototype.getClosestFeatureToCoordinate=function(t,e){var i=t[0],n=t[1],r=null,o=[NaN,NaN],s=1/0,a=[-1/0,-1/0,1/0,1/0],h=e||X;return this.featuresRtree_.forEachInExtent(a,(function(t){if(h(t)){var e=t.getGeometry(),u=s;if((s=e.closestPointXY(i,n,o,s))<u){r=t;var l=Math.sqrt(s);a[0]=i-l,a[1]=n-l,a[2]=i+l,a[3]=n+l}}})),r},e.prototype.getExtent=function(t){return this.featuresRtree_.getExtent(t)},e.prototype.getFeatureById=function(t){var e=this.idIndex_[t.toString()];return void 0!==e?e:null},e.prototype.getFormat=function(){return this.format_},e.prototype.getOverlaps=function(){return this.overlaps_},e.prototype.getUrl=function(){return this.url_},e.prototype.handleFeatureChange_=function(t){var e=t.target,i=o(e),n=e.getGeometry();if(n){var r=n.getExtent();i in this.nullGeometryFeatures_?(delete this.nullGeometryFeatures_[i],this.featuresRtree_&&this.featuresRtree_.insert(r,e)):this.featuresRtree_&&this.featuresRtree_.update(r,e)}else i in this.nullGeometryFeatures_||(this.featuresRtree_&&this.featuresRtree_.remove(e),this.nullGeometryFeatures_[i]=e);var s=e.getId();if(void 0!==s){var a=s.toString();i in this.undefIdIndex_?(delete this.undefIdIndex_[i],this.idIndex_[a]=e):this.idIndex_[a]!==e&&(this.removeFromIdIndex_(e),this.idIndex_[a]=e)}else i in this.undefIdIndex_||(this.removeFromIdIndex_(e),this.undefIdIndex_[i]=e);this.changed(),this.dispatchEvent(new Xn(bn,e))},e.prototype.hasFeature=function(t){var e=t.getId();return void 0!==e?e in this.idIndex_:o(t)in this.undefIdIndex_},e.prototype.isEmpty=function(){return this.featuresRtree_.isEmpty()&&M(this.nullGeometryFeatures_)},e.prototype.loadFeatures=function(t,e,i){var n=this,r=this.loadedExtentsRtree_,o=this.strategy_(t,e);this.loading=!1;for(var s=function(t,s){var a=o[t];r.forEachInExtent(a,(function(t){return Bt(t.extent,a)}))||(n.loader_.call(n,a,e,i),r.insert(a,{extent:a.slice()}),n.loading=n.loader_!==j)},a=0,h=o.length;a<h;++a)s(a)},e.prototype.removeLoadedExtent=function(t){var e,i=this.loadedExtentsRtree_;i.forEachInExtent(t,(function(i){if(Jt(i.extent,t))return e=i,!0})),e&&i.remove(e)},e.prototype.removeFeature=function(t){var e=o(t);e in this.nullGeometryFeatures_?delete this.nullGeometryFeatures_[e]:this.featuresRtree_&&this.featuresRtree_.remove(t),this.removeFeatureInternal(t),this.changed()},e.prototype.removeFeatureInternal=function(t){var e=o(t);this.featureChangeKeys_[e].forEach(T),delete this.featureChangeKeys_[e];var i=t.getId();void 0!==i?delete this.idIndex_[i.toString()]:delete this.undefIdIndex_[e],this.dispatchEvent(new Xn(An,t))},e.prototype.removeFromIdIndex_=function(t){var e=!1;for(var i in this.idIndex_)if(this.idIndex_[i]===t){delete this.idIndex_[i],e=!0;break}return e},e.prototype.setLoader=function(t){this.loader_=t},e}(Pn),jn=function(t){function e(){t.apply(this,arguments)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getType=function(){return ye},e.prototype.closestPointXY=function(t,e,i,n){var r,o,s,a=this.flatCoordinates,h=n;i[0]=a[0],i[1]=a[1];for(var u=0,l=a.length;u<l;u+=2)(r=(o=t-a[u])*o+(s=e-a[u+1])*s)<h&&(h=r,i[0]=a[u],i[1]=a[u+1]);return h},e}(Wi),qn=function(t){function e(){t.apply(this,arguments)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getType=function(){return ve},e.prototype.closestPointXY=function(t,e,i,n){var r,o,s,a=this.flatCoordinates,h=n;i[0]=a[0],i[1]=a[1];for(var u=0,l=a.length;u<l;u+=2)(r=(o=t-a[u])*o+(s=e-a[u+1])*s)<h&&(h=r,i[0]=a[u],i[1]=a[u+1]);return h},e.prototype.getArea=function(){var t=this.flatCoordinates,e=Math.sqrt(Math.pow(t[0]-t[4],2)+Math.pow(t[1]-t[5],2)),i=Math.sqrt(Math.pow(t[2]-t[6],2)+Math.pow(t[3]-t[7],2));return Math.PI*e*i*.25},e}(Wi),Bn="Point",Nn="LineString",Wn="Polygon",Vn="Circle",zn="Ellipse",Kn="drawstart",Un="drawend",Hn=function(t){function e(e,i){t.call(this,e),this.feature=i}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(B);var Jn=function(t){function e(e){var i=e;i.stopDown||(i.stopDown=Y),t.call(this,i),this.shouldHandle_=!1,this.downPx_=null,this.downTimeout_,this.lastDragTime_,this.freehand_=!1,this.source_=e.source?e.source:null,this.features_=e.features?e.features:null,this.snapTolerance_=e.snapTolerance?e.snapTolerance:12,this.type_=e.type,this.mode_=function(t){var e;t===he||t===pe?e=Bn:t===ue||t===fe||t===ye?e=Nn:t===ce||t===de?e=Wn:t===ge?e=Vn:t===ve&&(e=zn);return e}(this.type_),this.stopClick_=!!e.stopClick,this.minPoints_=e.minPoints?e.minPoints:this.mode_===Wn?3:2,this.maxPoints_=e.maxPoints?e.maxPoints:1/0,this.finishCondition_=e.finishCondition?e.finishCondition:X;var n,r=e.geometryFunction;if(!r)if(this.type_===ge)r=function(t,e){var i=e||new si([NaN,NaN]),n=Mt(t[0],t[1]);return i.setCenterAndRadius(t[0],Math.sqrt(n)),i};else if(this.type_===ye)this.minPoints_=3,this.maxPoints_=3,r=function(t,e){if(t.length>2){var i=t[0],n=t[1],r=t[2],o=[n[0]-i[0],n[1]-i[1]];if(0===o[1])t=[[i,i,i,i]];else{var s=[-1*o[1],o[0]],a=o[0]/o[1],h=(r[0]+a*(n[1]-r[1])-n[0])/(s[0]-s[1]*a),u=[h*s[0],h*s[1]];t=[[[i[0]-u[0],i[1]-u[1]],[n[0]-u[0],n[1]-u[1]],[n[0]+u[0],n[1]+u[1]],[i[0]+u[0],i[1]+u[1]]]]}}else t=[t];var l=e;return l?l.setCoordinates(t):l=new jn(t),l};else if(this.type_===ve)this.minPoints_=3,this.maxPoints_=3,r=function(t,e){if(t.length>2){var i=t[0],n=t[1],r=t[2],o=[n[0]-i[0],n[1]-i[1]],s=[i[0]+.5*o[0],i[1]+.5*o[1]];if(0===o[1])t=[[i,i,i,i]];else{var a=[-1*o[1],o[0]],h=o[0]/o[1],u=(r[0]+h*(n[1]-r[1])-n[0])/(a[0]-a[1]*h),l=[u*a[0],u*a[1]];t=[[[i[0],i[1]],[s[0]-l[0],s[1]-l[1]],[n[0],n[1]],[s[0]+l[0],s[1]+l[1]]]]}}else t=[t];var c=e;return c?c.setCoordinates(t):c=new qn(t),c};else{var o,s=this.mode_;s===Bn?o=Di:s===Nn?o=Ri:s===Wn&&(o=Wi),r=function(t,e){var i=e;return i?s===Wn?t[0].length?i.setCoordinates([t[0].concat([t[0][0]])]):i.setCoordinates([]):i.setCoordinates(t):i=new o(t),i}}this.geometryFunction_=r,this.dragVertexDelay_=void 0!==e.dragVertexDelay?e.dragVertexDelay:500,this.finishCoordinate_=null,this.sketchFeature_=null,this.sketchPoint_=null,this.sketchCoords_=null,this.sketchLine_=null,this.sketchLineCoords_=null,this.squaredClickTolerance_=e.clickTolerance?e.clickTolerance*e.clickTolerance:36,this.overlay_=new xn({source:new Yn({useSpatialIndex:!1,wrapX:!!e.wrapX&&e.wrapX}),style:e.style?e.style:(n=gn(),function(t,e){return n[t.getGeometry().getType()]}),updateWhileInteracting:!0}),this.geometryName_=e.geometryName,this.condition_=e.condition?e.condition:bt,this.freehandCondition_,e.freehand?this.freehandCondition_=Pt:this.freehandCondition_=e.freehandCondition?e.freehandCondition:Dt,G(this,pt(Ki),this.updateState_,this)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.setMap=function(e){t.prototype.setMap.call(this,e),this.updateState_()},e.prototype.getOverlay=function(){return this.overlay_},e.prototype.handleEvent=function(e){e.originalEvent.type===z&&e.preventDefault(),this.freehand_=this.mode_!==Bn&&this.freehandCondition_(e);var i=e.type===xt,n=!0;!this.freehand_&&this.lastDragTime_&&e.type===Ct&&(Date.now()-this.lastDragTime_>=this.dragVertexDelay_?(this.downPx_=e.pixel,this.shouldHandle_=!this.freehand_,i=!0):this.lastDragTime_=void 0,this.shouldHandle_&&void 0!==this.downTimeout_&&(clearTimeout(this.downTimeout_),this.downTimeout_=void 0));return this.freehand_&&e.type===Ct&&null!==this.sketchFeature_?(this.addToDrawing_(e),n=!1):this.freehand_&&e.type===wt?n=!1:i?(n=e.type===xt)&&this.freehand_?n=this.handlePointerMove_(e):("mouse"==e.pointerEvent.pointerType||e.type===Ct&&void 0===this.downTimeout_)&&this.handlePointerMove_(e):e.type===mt&&(n=!1),t.prototype.handleEvent.call(this,e)&&n},e.prototype.handleDownEvent=function(t){return this.shouldHandle_=!this.freehand_,this.freehand_?(this.downPx_=t.pixel,this.finishCoordinate_||this.startDrawing_(t),!0):!!this.condition_(t)&&(this.lastDragTime_=Date.now(),this.downTimeout_=setTimeout(function(){this.handlePointerMove_(new Ft(xt,t.map,t.pointerEvent,!1,t.frameState))}.bind(this),this.dragVertexDelay_),this.downPx_=t.pixel,!0)},e.prototype.handleUpEvent=function(t){var e=!0;this.downTimeout_&&(clearTimeout(this.downTimeout_),this.downTimeout_=void 0),this.handlePointerMove_(t);var i=this.mode_===Vn;return this.shouldHandle_?(this.finishCoordinate_?this.freehand_||i?this.finishDrawing():this.atFinish_(t)?this.finishCondition_(t)&&this.finishDrawing():this.addToDrawing_(t):(this.startDrawing_(t),this.mode_===Bn&&this.finishDrawing()),e=!1):this.freehand_&&(this.finishCoordinate_=null,this.abortDrawing_()),!e&&this.stopClick_&&t.stopPropagation(),e},e.prototype.handlePointerMove_=function(t){if(this.downPx_&&(!this.freehand_&&this.shouldHandle_||this.freehand_&&!this.shouldHandle_)){var e=this.downPx_,i=t.pixel,n=e[0]-i[0],r=e[1]-i[1],o=n*n+r*r;if(this.shouldHandle_=this.freehand_?o>this.squaredClickTolerance_:o<=this.squaredClickTolerance_,!this.shouldHandle_)return!0}return this.finishCoordinate_?this.modifyDrawing_(t):this.createOrUpdateSketchPoint_(t),!0},e.prototype.atFinish_=function(t){var e=!1;if(this.sketchFeature_){var i=!1,n=[this.finishCoordinate_];if(this.mode_===Nn)i=this.sketchCoords_.length>this.minPoints_;else if(this.mode_===Wn){var r=this.sketchCoords_;i=r[0].length>this.minPoints_,n=[r[0][0],r[0][r[0].length-2]]}if(i)for(var o=t.map,s=0,a=n.length;s<a;s++){var h=n[s],u=o.getPixelFromCoordinate(h),l=t.pixel,c=l[0]-u[0],p=l[1]-u[1],f=this.freehand_?1:this.snapTolerance_;if(e=Math.sqrt(c*c+p*p)<=f){this.finishCoordinate_=h;break}}}return e},e.prototype.createOrUpdateSketchPoint_=function(t){var e=t.coordinate.slice();this.sketchPoint_?this.sketchPoint_.getGeometry().setCoordinates(e):(this.sketchPoint_=new yt(new Di(e)),this.updateSketchFeatures_())},e.prototype.startDrawing_=function(t){var e=t.coordinate;this.finishCoordinate_=e,this.mode_===Bn?this.sketchCoords_=e.slice():this.mode_===Wn?(this.sketchCoords_=[[e.slice(),e.slice()]],this.sketchLineCoords_=this.sketchCoords_[0]):this.sketchCoords_=[e.slice(),e.slice()],this.sketchLineCoords_&&(this.sketchLine_=new yt(new Ri(this.sketchLineCoords_)));var i=this.geometryFunction_(this.sketchCoords_);this.sketchFeature_=new yt,this.geometryName_&&this.sketchFeature_.setGeometryName(this.geometryName_),this.sketchFeature_.setGeometry(i),this.updateSketchFeatures_(),this.dispatchEvent(new Hn(Kn,this.sketchFeature_))},e.prototype.modifyDrawing_=function(t){var e,i,n,r=t.coordinate,o=this.sketchFeature_.getGeometry();(this.mode_===Bn?i=this.sketchCoords_:this.mode_===Wn?(i=(e=this.sketchCoords_[0])[e.length-1],this.atFinish_(t)&&(r=this.finishCoordinate_.slice())):i=(e=this.sketchCoords_)[e.length-1],i[0]=r[0],i[1]=r[1],this.geometryFunction_(this.sketchCoords_,o),this.sketchPoint_)&&this.sketchPoint_.getGeometry().setCoordinates(r);if(this.mode_===zn)if(this.sketchLine_||(this.sketchLine_=new yt(new Ri([0,0]))),n=this.sketchLine_.getGeometry(),this.sketchCoords_.length<3)n.setCoordinates(this.sketchCoords_);else{var s=[((e=o.getCoordinates()[0])[0][0]+e[2][0])/2,(e[0][1]+e[2][1])/2];n.setCoordinates([e[0],s,e[3]])}else if(o.getType()==ce&&this.mode_!==Wn){this.sketchLine_||(this.sketchLine_=new yt);var a=o.getLinearRing(0);(n=this.sketchLine_.getGeometry())?(n.setFlatCoordinates(a.getLayout(),a.getFlatCoordinates()),n.changed()):(n=new Ri(a.getFlatCoordinates(),a.getLayout()),this.sketchLine_.setGeometry(n))}else this.sketchLineCoords_&&(n=this.sketchLine_.getGeometry()).setCoordinates(this.sketchLineCoords_);this.updateSketchFeatures_()},e.prototype.addToDrawing_=function(t){var e,i,n=t.coordinate,r=this.sketchFeature_.getGeometry();this.mode_===Nn||this.mode_===zn?(this.finishCoordinate_=n.slice(),(i=this.sketchCoords_).length>=this.maxPoints_&&(this.freehand_?i.pop():e=!0),i.push(n.slice()),this.geometryFunction_(i,r)):this.mode_===Wn&&((i=this.sketchCoords_[0]).length>=this.maxPoints_&&(this.freehand_?i.pop():e=!0),i.push(n.slice()),e&&(this.finishCoordinate_=i[0]),this.geometryFunction_(this.sketchCoords_,r)),this.updateSketchFeatures_(),e&&this.finishDrawing()},e.prototype.removeLastPoint=function(){if(this.sketchFeature_){var t,e=this.sketchFeature_.getGeometry();this.mode_===Nn?((t=this.sketchCoords_).splice(-2,1),this.geometryFunction_(t,e),t.length>=2&&(this.finishCoordinate_=t[t.length-2].slice())):this.mode_===Wn&&((t=this.sketchCoords_[0]).splice(-2,1),this.sketchLine_.getGeometry().setCoordinates(t),this.geometryFunction_(this.sketchCoords_,e)),0===t.length&&(this.finishCoordinate_=null),this.updateSketchFeatures_()}},e.prototype.finishDrawing=function(){var t=this.abortDrawing_();if(t){var e=this.sketchCoords_,i=t.getGeometry();this.mode_===Nn?(e.pop(),this.geometryFunction_(e,i)):this.mode_===Wn&&(e[0].pop(),this.geometryFunction_(e,i),e=i.getCoordinates()),this.type_===pe?t.setGeometry(new Ai([e])):this.type_===fe?t.setGeometry(new bi([e])):this.type_===de&&t.setGeometry(new Vi([e])),this.dispatchEvent(new Hn(Un,t)),this.features_&&this.features_.push(t),this.source_&&this.source_.addFeature(t)}},e.prototype.abortDrawing_=function(){this.finishCoordinate_=null;var t=this.sketchFeature_;return t&&(this.sketchFeature_=null,this.sketchPoint_=null,this.sketchLine_=null,this.overlay_.getSource().clear(!0)),t},e.prototype.extend=function(t){var e=t.getGeometry();this.sketchFeature_=t,this.sketchCoords_=e.getCoordinates();var i=this.sketchCoords_[this.sketchCoords_.length-1];this.finishCoordinate_=i.slice(),this.sketchCoords_.push(i.slice()),this.updateSketchFeatures_(),this.dispatchEvent(new Hn(Kn,this.sketchFeature_))},e.prototype.updateSketchFeatures_=function(){var t=[];this.sketchFeature_&&t.push(this.sketchFeature_),this.sketchLine_&&t.push(this.sketchLine_),this.sketchPoint_&&t.push(this.sketchPoint_);var e=this.overlay_.getSource();e.clear(!0),e.addFeatures(t)},e.prototype.updateState_=function(){var t=this.getMap(),e=this.getActive();t&&e||this.abortDrawing_(),this.overlay_.setMap(e?t:null)},e}(Ui),Zn=Vue.resource("/api/v1/volumes{/id}/export-area"),$n="modifystart",Qn="modifyend",tr=function(t){function e(e,i,n){t.call(this,e),this.features=i,this.mapBrowserEvent=n}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(B);function er(t,e){return t.index-e.index}function ir(t,e){var i,n,r=e.geometry;if(r.getType()===ge){var o=r;if(1===e.index){var s=Mt(o.getCenter(),t),a=Math.sqrt(s)-o.getRadius();return a*a}}return i=t,n=e.segment,Mt(i,St(i,n))}function nr(t,e){var i=e.geometry;return i.getType()===ge&&1===e.index||i.getType()===ve||i.getType()===ye?i.getClosestPoint(t):St(t,e.segment)}var rr=function(t){function e(e){var i,n;if(t.call(this,e),this.condition_=e.condition?e.condition:At,this.defaultDeleteCondition_=function(t){return function(t){var e=t.originalEvent;return e.altKey&&!(e.metaKey||e.ctrlKey)&&!e.shiftKey}(t)&&function(t){return t.type==vt}(t)},this.deleteCondition_=e.deleteCondition?e.deleteCondition:this.defaultDeleteCondition_,this.insertVertexCondition_=e.insertVertexCondition?e.insertVertexCondition:Pt,this.vertexFeature_=null,this.vertexSegments_=null,this.lastPixel_=[0,0],this.ignoreNextSingleClick_=!1,this.modified_=!1,this.rBush_=new Tn,this.pixelTolerance_=void 0!==e.pixelTolerance?e.pixelTolerance:10,this.snappedToVertex_=!1,this.changingFeature_=!1,this.dragSegments_=[],this.overlay_=new xn({source:new Yn({useSpatialIndex:!1,wrapX:!!e.wrapX}),style:e.style?e.style:(i=gn(),function(t,e){return i[he]}),updateWhileAnimating:!0,updateWhileInteracting:!0}),this.SEGMENT_WRITERS_={Point:this.writePointGeometry_,LineString:this.writeLineStringGeometry_,LinearRing:this.writeLineStringGeometry_,Polygon:this.writePolygonGeometry_,MultiPoint:this.writeMultiPointGeometry_,MultiLineString:this.writeMultiLineStringGeometry_,MultiPolygon:this.writeMultiPolygonGeometry_,Rectangle:this.writePolygonGeometry_,Ellipse:this.writePolygonGeometry_,Circle:this.writeCircleGeometry_,GeometryCollection:this.writeGeometryCollectionGeometry_},this.source_=null,e.source?(this.source_=e.source,n=new gt(this.source_.getFeatures()),G(this.source_,Rn,this.handleSourceAdd_,this),G(this.source_,An,this.handleSourceRemove_,this)):n=e.features,!n)throw new Error("The modify interaction requires features or a source");this.features_=n,this.features_.forEach(this.addFeature_.bind(this)),G(this.features_,st,this.handleFeatureAdd_,this),G(this.features_,at,this.handleFeatureRemove_,this),this.lastPointerEvent_=null}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.addFeature_=function(t){var e=t.getGeometry();e&&e.getType()in this.SEGMENT_WRITERS_&&this.SEGMENT_WRITERS_[e.getType()].call(this,t,e);var i=this.getMap();i&&i.isRendered()&&this.getActive()&&this.handlePointerAtPixel_(this.lastPixel_,i),G(t,W,this.handleFeatureChange_,this)},e.prototype.willModifyFeatures_=function(t){this.modified_||(this.modified_=!0,this.dispatchEvent(new tr($n,this.features_,t)))},e.prototype.removeFeature_=function(t){this.removeFeatureSegmentData_(t),this.vertexFeature_&&0===this.features_.getLength()&&(this.overlay_.getSource().removeFeature(this.vertexFeature_),this.vertexFeature_=null),L(t,W,this.handleFeatureChange_,this)},e.prototype.removeFeatureSegmentData_=function(t){var e=this.rBush_,i=[];e.forEach((function(e){t===e.feature&&i.push(e)}));for(var n=i.length-1;n>=0;--n)e.remove(i[n])},e.prototype.setActive=function(e){this.vertexFeature_&&!e&&(this.overlay_.getSource().removeFeature(this.vertexFeature_),this.vertexFeature_=null),t.prototype.setActive.call(this,e)},e.prototype.setMap=function(e){this.overlay_.setMap(e),t.prototype.setMap.call(this,e)},e.prototype.getOverlay=function(){return this.overlay_},e.prototype.handleSourceAdd_=function(t){t.feature&&this.features_.push(t.feature)},e.prototype.handleSourceRemove_=function(t){t.feature&&this.features_.remove(t.feature)},e.prototype.handleFeatureAdd_=function(t){this.addFeature_(t.element)},e.prototype.handleFeatureChange_=function(t){if(!this.changingFeature_){var e=t.target;this.removeFeature_(e),this.addFeature_(e)}},e.prototype.handleFeatureRemove_=function(t){var e=t.element;this.removeFeature_(e)},e.prototype.writePointGeometry_=function(t,e){var i=e.getCoordinates(),n={feature:t,geometry:e,segment:[i,i]};this.rBush_.insert(e.getExtent(),n)},e.prototype.writeMultiPointGeometry_=function(t,e){for(var i=e.getCoordinates(),n=0,r=i.length;n<r;++n){var o=i[n],s={feature:t,geometry:e,depth:[n],index:n,segment:[o,o]};this.rBush_.insert(e.getExtent(),s)}},e.prototype.writeLineStringGeometry_=function(t,e){for(var i=e.getCoordinates(),n=0,r=i.length-1;n<r;++n){var o=i.slice(n,n+2),s={feature:t,geometry:e,index:n,segment:o};this.rBush_.insert(jt(o),s)}},e.prototype.writeMultiLineStringGeometry_=function(t,e){for(var i=e.getCoordinates(),n=0,r=i.length;n<r;++n)for(var o=i[n],s=0,a=o.length-1;s<a;++s){var h=o.slice(s,s+2),u={feature:t,geometry:e,depth:[n],index:s,segment:h};this.rBush_.insert(jt(h),u)}},e.prototype.writePolygonGeometry_=function(t,e){for(var i=e.getCoordinates(),n=0,r=i.length;n<r;++n)for(var o=i[n],s=0,a=o.length-1;s<a;++s){var h=o.slice(s,s+2),u={feature:t,geometry:e,depth:[n],index:s,segment:h};this.rBush_.insert(jt(h),u)}},e.prototype.writeMultiPolygonGeometry_=function(t,e){for(var i=e.getCoordinates(),n=0,r=i.length;n<r;++n)for(var o=i[n],s=0,a=o.length;s<a;++s)for(var h=o[s],u=0,l=h.length-1;u<l;++u){var c=h.slice(u,u+2),p={feature:t,geometry:e,depth:[s,n],index:u,segment:c};this.rBush_.insert(jt(c),p)}},e.prototype.writeCircleGeometry_=function(t,e){var i=e.getCenter(),n={feature:t,geometry:e,index:0,segment:[i,i]},r={feature:t,geometry:e,index:1,segment:[i,i]},o=[n,r];n.featureSegments=r.featureSegments=o,this.rBush_.insert(Ut(i),n),this.rBush_.insert(e.getExtent(),r)},e.prototype.writeGeometryCollectionGeometry_=function(t,e){for(var i=e.getGeometriesArray(),n=0;n<i.length;++n)this.SEGMENT_WRITERS_[i[n].getType()].call(this,t,i[n])},e.prototype.createOrUpdateVertexFeature_=function(t){var e=this.vertexFeature_;e?e.getGeometry().setCoordinates(t):(e=new yt(new Di(t)),this.vertexFeature_=e,this.overlay_.getSource().addFeature(e));return e},e.prototype.handleEvent=function(e){return!e.pointerEvent||(this.lastPointerEvent_=e,e.map.getView().getInteracting()||e.type!=xt||this.handlingDownUpSequence||this.handlePointerMove_(e),this.vertexFeature_&&this.deleteCondition_(e)&&(i=!(e.type!=vt||!this.ignoreNextSingleClick_)||this.removePoint()),e.type==vt&&(this.ignoreNextSingleClick_=!1),t.prototype.handleEvent.call(this,e)&&!i);var i},e.prototype.handleDragEvent=function(t){this.ignoreNextSingleClick_=!1,this.willModifyFeatures_(t);for(var e=t.coordinate,i=0,n=this.dragSegments_.length;i<n;++i){for(var r=this.dragSegments_[i],o=r[0],s=o.depth,a=o.geometry,h=void 0,u=o.segment,l=r[1];e.length<a.getStride();)e.push(u[l][e.length]);switch(a.getType()){case he:h=e,u[0]=u[1]=e;break;case pe:(h=a.getCoordinates())[o.index]=e,u[0]=u[1]=e;break;case ue:(h=a.getCoordinates())[o.index+l]=e,u[l]=e;break;case fe:case ce:(h=a.getCoordinates())[s[0]][o.index+l]=e,u[l]=e;break;case de:(h=a.getCoordinates())[s[1]][s[0]][o.index+l]=e,u[l]=e;break;case ye:var c=(y=(h=a.getCoordinates())[s[0]])[(o.index+l+2)%4],p=[e[0]-c[0],e[1]-c[1]],f=[(v=y[(o.index+l+1)%4])[0]-c[0],v[1]-c[1]],d=Math.sqrt(f[0]*f[0]+f[1]*f[1]);f[0]=f[0]/d,f[1]=f[1]/d;var _=p[0]*f[0]+p[1]*f[1];y[(o.index+l+1)%4]=[c[0]+_*f[0],c[1]+_*f[1]];var g=[(m=y[(o.index+l+3)%4])[0]-c[0],m[1]-c[1]];d=Math.sqrt(g[0]*g[0]+g[1]*g[1]),g[0]=g[0]/d,g[1]=g[1]/d,_=p[0]*g[0]+p[1]*g[1],y[(o.index+l+3)%4]=[c[0]+_*g[0],c[1]+_*g[1]],y[o.index+l]=e;break;case ve:var y,v=(y=(h=a.getCoordinates())[s[0]])[(o.index+l+1)%4],m=(c=y[(o.index+l+2)%4],y[(o.index+l+3)%4]),C=Math.sqrt(Math.pow(v[0]-m[0],2)+Math.pow(v[1]-m[1],2))/2,x=[e[0]-c[0],e[1]-c[1]],w=[-1*x[1],x[0]];d=Math.sqrt(w[0]*w[0]+w[1]*w[1]);w[0]=w[0]/d,w[1]=w[1]/d;var E=[c[0]+x[0]/2,c[1]+x[1]/2];y[(o.index+l+1)%4]=[E[0]+w[0]*C,E[1]+w[1]*C],y[(o.index+l+3)%4]=[E[0]-w[0]*C,E[1]-w[1]*C],y[o.index+l]=e;break;case ge:u[0]=u[1]=e,0===o.index?(this.changingFeature_=!0,a.setCenter(e),this.changingFeature_=!1):(this.changingFeature_=!0,a.setRadius(It(a.getCenter(),e)),this.changingFeature_=!1)}h&&this.setGeometryCoordinates_(a,h)}this.createOrUpdateVertexFeature_(e)},e.prototype.handleDownEvent=function(t){if(!this.condition_(t))return!1;this.handlePointerAtPixel_(t.pixel,t.map);var e=t.map.getCoordinateFromPixel(t.pixel);this.dragSegments_.length=0,this.modified_=!1;var i=this.vertexFeature_;if(i){var n=[],r=i.getGeometry().getCoordinates(),s=jt([r]),a=this.rBush_.getInExtent(s),h={};a.sort(er);for(var u=0,l=a.length;u<l;++u){var c=a[u],p=c.segment,f=o(c.feature),d=c.depth;if(d&&(f+="-"+d.join("-")),h[f]||(h[f]=new Array(2)),c.geometry.getType()===ge&&1===c.index)kt(nr(e,c),r)&&!h[f][0]&&(this.dragSegments_.push([c,0]),h[f][0]=c);else if(kt(p[0],r)&&!h[f][0])this.dragSegments_.push([c,0]),h[f][0]=c;else if(kt(p[1],r)&&!h[f][1]){if((c.geometry.getType()===ue||c.geometry.getType()===fe)&&h[f][0]&&0===h[f][0].index)continue;this.dragSegments_.push([c,1]),h[f][1]=c}else this.insertVertexCondition_(t)&&o(p)in this.vertexSegments_&&!h[f][0]&&!h[f][1]&&n.push([c,r])}n.length&&this.willModifyFeatures_(t);for(var _=n.length-1;_>=0;--_)this.insertVertex_.apply(this,n[_])}return!!this.vertexFeature_},e.prototype.handleUpEvent=function(t){for(var e=this.dragSegments_.length-1;e>=0;--e){var i=this.dragSegments_[e][0],n=i.geometry;if(n.getType()===ye||n.getType()===ve)this.features_.remove(i.feature),this.features_.push(i.feature);else if(n.getType()===ge){var r=n.getCenter(),o=i.featureSegments[0],s=i.featureSegments[1];o.segment[0]=o.segment[1]=r,s.segment[0]=s.segment[1]=r,this.rBush_.update(Ut(r),o),this.rBush_.update(n.getExtent(),s)}else this.rBush_.update(jt(i.segment),i)}return this.modified_&&(this.dispatchEvent(new tr(Qn,this.features_,t)),this.modified_=!1),!1},e.prototype.handlePointerMove_=function(t){this.lastPixel_=t.pixel,this.handlePointerAtPixel_(t.pixel,t.map)},e.prototype.handlePointerAtPixel_=function(t,e){var i,n,r,s=e.getCoordinateFromPixel(t),a=(i=Ut(s),n=e.getView().getResolution()*this.pixelTolerance_,r?(r[0]=i[0]-n,r[1]=i[1]-n,r[2]=i[2]+n,r[3]=i[3]+n,r):[i[0]-n,i[1]-n,i[2]+n,i[3]+n]),h=this.rBush_.getInExtent(a);if(h.length>0){h.sort((function(t,e){return ir(s,t)-ir(s,e)}));var u=h[0],l=u.segment,c=nr(s,u),p=e.getPixelFromCoordinate(c),f=It(t,p);if(f<=this.pixelTolerance_){var d={};if(u.geometry.getType()===ge&&1===u.index)this.snappedToVertex_=!0,this.createOrUpdateVertexFeature_(c);else{var _=e.getPixelFromCoordinate(l[0]),g=e.getPixelFromCoordinate(l[1]),y=Mt(p,_),v=Mt(p,g);f=Math.sqrt(Math.min(y,v)),this.snappedToVertex_=f<=this.pixelTolerance_,this.snappedToVertex_&&(c=y>v?l[1]:l[0]),this.createOrUpdateVertexFeature_(c);for(var m=1,C=h.length;m<C;++m){var x=h[m].segment;if(!(kt(l[0],x[0])&&kt(l[1],x[1])||kt(l[0],x[1])&&kt(l[1],x[0])))break;d[o(x)]=!0}}return d[o(l)]=!0,void(this.vertexSegments_=d)}}this.vertexFeature_&&(this.overlay_.getSource().removeFeature(this.vertexFeature_),this.vertexFeature_=null)},e.prototype.insertVertex_=function(t,e){for(var i,n=t.segment,r=t.feature,o=t.geometry,s=t.depth,a=t.index;e.length<o.getStride();)e.push(0);switch(o.getType()){case fe:case ce:(i=o.getCoordinates())[s[0]].splice(a+1,0,e);break;case de:(i=o.getCoordinates())[s[1]][s[0]].splice(a+1,0,e);break;case ue:(i=o.getCoordinates()).splice(a+1,0,e);break;default:return}this.setGeometryCoordinates_(o,i);var h=this.rBush_;h.remove(t),this.updateSegmentIndices_(o,a,s,1);var u={segment:[n[0],e],feature:r,geometry:o,depth:s,index:a};h.insert(jt(u.segment),u),this.dragSegments_.push([u,1]);var l={segment:[e,n[1]],feature:r,geometry:o,depth:s,index:a+1};h.insert(jt(l.segment),l),this.dragSegments_.push([l,0]),this.ignoreNextSingleClick_=!0},e.prototype.removePoint=function(){if(this.lastPointerEvent_&&this.lastPointerEvent_.type!=Ct){var t=this.lastPointerEvent_;return this.willModifyFeatures_(t),this.removeVertex_(),this.dispatchEvent(new tr(Qn,this.features_,t)),this.modified_=!1,!0}return!1},e.prototype.removeVertex_=function(){var t,e,i,n,r,s,a,h,u,l,c,p=this.dragSegments_,f={},d=!1;for(r=p.length-1;r>=0;--r)c=o((l=(i=p[r])[0]).feature),l.depth&&(c+="-"+l.depth.join("-")),c in f||(f[c]={}),0===i[1]?(f[c].right=l,f[c].index=l.index):1==i[1]&&(f[c].left=l,f[c].index=l.index+1);for(c in f){switch(u=f[c].right,a=f[c].left,(h=(s=f[c].index)-1)<0&&(h=0),t=e=(n=(l=void 0!==a?a:u).geometry).getCoordinates(),d=!1,n.getType()){case fe:e[l.depth[0]].length>2&&(e[l.depth[0]].splice(s,1),d=!0);break;case ue:e.length>2&&(e.splice(s,1),d=!0);break;case de:t=t[l.depth[1]];case ce:(t=t[l.depth[0]]).length>4&&(s==t.length-1&&(s=0),t.splice(s,1),d=!0,0===s&&(t.pop(),t.push(t[0]),h=t.length-1))}if(d){this.setGeometryCoordinates_(n,e);var _=[];if(void 0!==a&&(this.rBush_.remove(a),_.push(a.segment[0])),void 0!==u&&(this.rBush_.remove(u),_.push(u.segment[1])),void 0!==a&&void 0!==u){var g={depth:l.depth,feature:l.feature,geometry:l.geometry,index:h,segment:_};this.rBush_.insert(jt(g.segment),g)}this.updateSegmentIndices_(n,s,l.depth,-1),this.vertexFeature_&&(this.overlay_.getSource().removeFeature(this.vertexFeature_),this.vertexFeature_=null),p.length=0}}return d},e.prototype.setGeometryCoordinates_=function(t,e){this.changingFeature_=!0,t.setCoordinates(e),this.changingFeature_=!1},e.prototype.updateSegmentIndices_=function(t,e,i,n){this.rBush_.forEachInExtent(t.getExtent(),(function(r){r.geometry===t&&(void 0===i||void 0===r.depth||function(t,e){var i=t.length;if(i!==e.length)return!1;for(var n=0;n<i;n++)if(t[n]!==e[n])return!1;return!0}(r.depth,i))&&r.index>e&&(r.index+=n)}))},e}(Ui),or=biigle.$require("events"),sr=biigle.$require("messages").handleErrorResponse,ar=biigle.$require("annotations.components.settingsTabPlugins");ar&&(ar.exportArea={props:{settings:{type:Object,required:!0}},data:function(){return{opacityValue:"1",currentImage:null,isEditing:!1,exportArea:null,volumeId:null}},computed:{opacity:function(){return parseFloat(this.opacityValue)},shown:function(){return this.opacity>0},height:function(){return this.currentImage?this.currentImage.height:0},hasExportArea:function(){return null!==this.exportArea},layer:function(){return new xn({source:new Yn({features:new gt}),style:[new vn({stroke:new pn({color:"white",width:4}),image:new ot({radius:6,fill:new ln({color:"#666666"}),stroke:new pn({color:"white",width:2,lineDash:[2]})})}),new vn({stroke:new pn({color:"#666666",width:1,lineDash:[2]})})],zIndex:4,updateWhileAnimating:!0,updateWhileInteracting:!0})},drawInteraction:function(){return new Jn({source:this.layer.getSource(),type:"Rectangle",style:this.layer.getStyle(),minPoints:2,maxPoints:2,geometryFunction:function(t,e){t.length>1&&(t=[t[0],[t[0][0],t[1][1]],t[1],[t[1][0],t[0][1]]]);var i=e;return i?i.setCoordinates([t]):i=new jn([t]),i}})},modifyInteraction:function(){return new rr({features:this.layer.getSource().getFeaturesCollection(),style:this.layer.getStyle(),deleteCondition:Rt})}},methods:{toggleEditing:function(){this.isEditing=!this.isEditing,this.isEditing?(this.drawInteraction.setActive(!0),this.modifyInteraction.setActive(!0)):(this.drawInteraction.setActive(!1),this.modifyInteraction.setActive(!1))},deleteArea:function(){var t=this;if(this.hasExportArea&&confirm("Do you really want to delete the export area?")){var e=this.layer.getSource(),i=e.getFeatures()[0];e.clear(),Zn.delete({id:this.volumeId}).then((function(){return t.exportArea=null})).catch((function(t){e.addFeature(i),sr(t)}))}},updateCurrentImage:function(t,e){this.currentImage=e},maybeDrawArea:function(){if(this.clearSource(),this.exportArea&&this.height>0){var t=this.currentImage.tiled?0:this.height,e=new jn([[[this.exportArea[0],t-this.exportArea[1]],[this.exportArea[0],t-this.exportArea[3]],[this.exportArea[2],t-this.exportArea[3]],[this.exportArea[2],t-this.exportArea[1]]]]);this.layer.getSource().addFeature(new yt({geometry:e}))}},handleModifyend:function(t){this.updateExportArea(t.features.item(0))},clearSource:function(){this.layer.getSource().clear()},handleDrawend:function(t){var e=this.layer.getSource(),i=e.getFeatures()[0];e.clear(),this.updateExportArea(t.feature).catch((function(){e.clear(),i&&e.addFeature(i)}))},updateExportArea:function(t){var e=this,i=t.getGeometry().getCoordinates()[0],n=this.currentImage.tiled?0:this.height;i=[i[0][0],n-i[0][1],i[2][0],n-i[2][1]].map(Math.round);var r=Zn.save({id:this.volumeId},{coordinates:i}).then((function(){return e.exportArea=i}));return r.catch(sr),r},extendMap:function(t){t.addLayer(this.layer),t.addInteraction(this.drawInteraction),t.addInteraction(this.modifyInteraction)}},watch:{opacity:function(t){t<1?this.settings.set("exportAreaOpacity",t):this.settings.delete("exportAreaOpacity"),this.layer.setOpacity(t)},exportArea:function(){this.maybeDrawArea()},height:function(){this.maybeDrawArea()}},created:function(){this.volumeId=biigle.$require("annotations.volumeId"),this.settings.has("exportAreaOpacity")&&(this.opacityValue=this.settings.get("exportAreaOpacity")),this.exportArea=biigle.$require("annotations.exportArea"),this.drawInteraction.setActive(!1),this.modifyInteraction.setActive(!1),this.drawInteraction.on("drawend",this.handleDrawend),this.modifyInteraction.on("modifyend",this.handleModifyend),or.$on("images.change",this.updateCurrentImage),or.$on("annotations.map.init",this.extendMap)}})},"A1R+":function(t,e,i){"use strict";t.exports=r,t.exports.default=r;var n=i("YcpW");function r(t,e){if(!(this instanceof r))return new r(t,e);this._maxEntries=Math.max(4,t||9),this._minEntries=Math.max(2,Math.ceil(.4*this._maxEntries)),e&&this._initFormat(e),this.clear()}function o(t,e,i){if(!i)return e.indexOf(t);for(var n=0;n<e.length;n++)if(i(t,e[n]))return n;return-1}function s(t,e){a(t,0,t.children.length,e,t)}function a(t,e,i,n,r){r||(r=_(null)),r.minX=1/0,r.minY=1/0,r.maxX=-1/0,r.maxY=-1/0;for(var o,s=e;s<i;s++)o=t.children[s],h(r,t.leaf?n(o):o);return r}function h(t,e){return t.minX=Math.min(t.minX,e.minX),t.minY=Math.min(t.minY,e.minY),t.maxX=Math.max(t.maxX,e.maxX),t.maxY=Math.max(t.maxY,e.maxY),t}function u(t,e){return t.minX-e.minX}function l(t,e){return t.minY-e.minY}function c(t){return(t.maxX-t.minX)*(t.maxY-t.minY)}function p(t){return t.maxX-t.minX+(t.maxY-t.minY)}function f(t,e){return t.minX<=e.minX&&t.minY<=e.minY&&e.maxX<=t.maxX&&e.maxY<=t.maxY}function d(t,e){return e.minX<=t.maxX&&e.minY<=t.maxY&&e.maxX>=t.minX&&e.maxY>=t.minY}function _(t){return{children:t,height:1,leaf:!0,minX:1/0,minY:1/0,maxX:-1/0,maxY:-1/0}}function g(t,e,i,r,o){for(var s,a=[e,i];a.length;)(i=a.pop())-(e=a.pop())<=r||(s=e+Math.ceil((i-e)/r/2)*r,n(t,s,e,i,o),a.push(e,s,s,i))}r.prototype={all:function(){return this._all(this.data,[])},search:function(t){var e=this.data,i=[],n=this.toBBox;if(!d(t,e))return i;for(var r,o,s,a,h=[];e;){for(r=0,o=e.children.length;r<o;r++)s=e.children[r],d(t,a=e.leaf?n(s):s)&&(e.leaf?i.push(s):f(t,a)?this._all(s,i):h.push(s));e=h.pop()}return i},collides:function(t){var e=this.data,i=this.toBBox;if(!d(t,e))return!1;for(var n,r,o,s,a=[];e;){for(n=0,r=e.children.length;n<r;n++)if(o=e.children[n],d(t,s=e.leaf?i(o):o)){if(e.leaf||f(t,s))return!0;a.push(o)}e=a.pop()}return!1},load:function(t){if(!t||!t.length)return this;if(t.length<this._minEntries){for(var e=0,i=t.length;e<i;e++)this.insert(t[e]);return this}var n=this._build(t.slice(),0,t.length-1,0);if(this.data.children.length)if(this.data.height===n.height)this._splitRoot(this.data,n);else{if(this.data.height<n.height){var r=this.data;this.data=n,n=r}this._insert(n,this.data.height-n.height-1,!0)}else this.data=n;return this},insert:function(t){return t&&this._insert(t,this.data.height-1),this},clear:function(){return this.data=_([]),this},remove:function(t,e){if(!t)return this;for(var i,n,r,s,a=this.data,h=this.toBBox(t),u=[],l=[];a||u.length;){if(a||(a=u.pop(),n=u[u.length-1],i=l.pop(),s=!0),a.leaf&&-1!==(r=o(t,a.children,e)))return a.children.splice(r,1),u.push(a),this._condense(u),this;s||a.leaf||!f(a,h)?n?(i++,a=n.children[i],s=!1):a=null:(u.push(a),l.push(i),i=0,n=a,a=a.children[0])}return this},toBBox:function(t){return t},compareMinX:u,compareMinY:l,toJSON:function(){return this.data},fromJSON:function(t){return this.data=t,this},_all:function(t,e){for(var i=[];t;)t.leaf?e.push.apply(e,t.children):i.push.apply(i,t.children),t=i.pop();return e},_build:function(t,e,i,n){var r,o=i-e+1,a=this._maxEntries;if(o<=a)return s(r=_(t.slice(e,i+1)),this.toBBox),r;n||(n=Math.ceil(Math.log(o)/Math.log(a)),a=Math.ceil(o/Math.pow(a,n-1))),(r=_([])).leaf=!1,r.height=n;var h,u,l,c,p=Math.ceil(o/a),f=p*Math.ceil(Math.sqrt(a));for(g(t,e,i,f,this.compareMinX),h=e;h<=i;h+=f)for(g(t,h,l=Math.min(h+f-1,i),p,this.compareMinY),u=h;u<=l;u+=p)c=Math.min(u+p-1,l),r.children.push(this._build(t,u,c,n-1));return s(r,this.toBBox),r},_chooseSubtree:function(t,e,i,n){for(var r,o,s,a,h,u,l,p,f,d;n.push(e),!e.leaf&&n.length-1!==i;){for(l=p=1/0,r=0,o=e.children.length;r<o;r++)h=c(s=e.children[r]),f=t,d=s,(u=(Math.max(d.maxX,f.maxX)-Math.min(d.minX,f.minX))*(Math.max(d.maxY,f.maxY)-Math.min(d.minY,f.minY))-h)<p?(p=u,l=h<l?h:l,a=s):u===p&&h<l&&(l=h,a=s);e=a||e.children[0]}return e},_insert:function(t,e,i){var n=this.toBBox,r=i?t:n(t),o=[],s=this._chooseSubtree(r,this.data,e,o);for(s.children.push(t),h(s,r);e>=0&&o[e].children.length>this._maxEntries;)this._split(o,e),e--;this._adjustParentBBoxes(r,o,e)},_split:function(t,e){var i=t[e],n=i.children.length,r=this._minEntries;this._chooseSplitAxis(i,r,n);var o=this._chooseSplitIndex(i,r,n),a=_(i.children.splice(o,i.children.length-o));a.height=i.height,a.leaf=i.leaf,s(i,this.toBBox),s(a,this.toBBox),e?t[e-1].children.push(a):this._splitRoot(i,a)},_splitRoot:function(t,e){this.data=_([t,e]),this.data.height=t.height+1,this.data.leaf=!1,s(this.data,this.toBBox)},_chooseSplitIndex:function(t,e,i){var n,r,o,s,h,u,l,p,f,d,_,g,y,v;for(u=l=1/0,n=e;n<=i-e;n++)r=a(t,0,n,this.toBBox),o=a(t,n,i,this.toBBox),f=r,d=o,_=void 0,g=void 0,y=void 0,v=void 0,_=Math.max(f.minX,d.minX),g=Math.max(f.minY,d.minY),y=Math.min(f.maxX,d.maxX),v=Math.min(f.maxY,d.maxY),s=Math.max(0,y-_)*Math.max(0,v-g),h=c(r)+c(o),s<u?(u=s,p=n,l=h<l?h:l):s===u&&h<l&&(l=h,p=n);return p},_chooseSplitAxis:function(t,e,i){var n=t.leaf?this.compareMinX:u,r=t.leaf?this.compareMinY:l;this._allDistMargin(t,e,i,n)<this._allDistMargin(t,e,i,r)&&t.children.sort(n)},_allDistMargin:function(t,e,i,n){t.children.sort(n);var r,o,s=this.toBBox,u=a(t,0,e,s),l=a(t,i-e,i,s),c=p(u)+p(l);for(r=e;r<i-e;r++)o=t.children[r],h(u,t.leaf?s(o):o),c+=p(u);for(r=i-e-1;r>=e;r--)o=t.children[r],h(l,t.leaf?s(o):o),c+=p(l);return c},_adjustParentBBoxes:function(t,e,i){for(var n=i;n>=0;n--)h(e[n],t)},_condense:function(t){for(var e,i=t.length-1;i>=0;i--)0===t[i].children.length?i>0?(e=t[i-1].children).splice(e.indexOf(t[i]),1):this.clear():s(t[i],this.toBBox)},_initFormat:function(t){var e=["return a"," - b",";"];this.compareMinX=new Function("a","b",e.join(t[0])),this.compareMinY=new Function("a","b",e.join(t[1])),this.toBBox=new Function("a","return {minX: a"+t[0]+", minY: a"+t[1]+", maxX: a"+t[2]+", maxY: a"+t[3]+"};")}}},YcpW:function(t,e,i){t.exports=function(){"use strict";function t(t,e,i){var n=t[e];t[e]=t[i],t[i]=n}function e(t,e){return t<e?-1:t>e?1:0}return function(i,n,r,o,s){!function e(i,n,r,o,s){for(;o>r;){if(o-r>600){var a=o-r+1,h=n-r+1,u=Math.log(a),l=.5*Math.exp(2*u/3),c=.5*Math.sqrt(u*l*(a-l)/a)*(h-a/2<0?-1:1),p=Math.max(r,Math.floor(n-h*l/a+c)),f=Math.min(o,Math.floor(n+(a-h)*l/a+c));e(i,n,p,f,s)}var d=i[n],_=r,g=o;for(t(i,r,n),s(i[o],d)>0&&t(i,r,o);_<g;){for(t(i,_,g),_++,g--;s(i[_],d)<0;)_++;for(;s(i[g],d)>0;)g--}0===s(i[r],d)?t(i,r,g):(g++,t(i,g,o)),g<=n&&(r=g+1),n<=g&&(o=g-1)}}(i,n,r||0,o||i.length-1,s||e)}}()}});
+!function(t){var e={};function i(n){if(e[n])return e[n].exports;var r=e[n]={i:n,l:!1,exports:{}};return t[n].call(r.exports,r,r.exports,i),r.l=!0,r.exports}i.m=t,i.c=e,i.d=function(t,e,n){i.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:n})},i.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},i.t=function(t,e){if(1&e&&(t=i(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var n=Object.create(null);if(i.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var r in t)i.d(n,r,function(e){return t[e]}.bind(null,r));return n},i.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return i.d(e,"a",e),e},i.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},i.p="/",i(i.s=1)}({1:function(t,e,i){t.exports=i("2R1P")},"2R1P":function(t,e,i){"use strict";function n(){return function(){throw new Error("Unimplemented abstract method.")}()}i.r(e);var r=0;function o(t){return t.ol_uid||(t.ol_uid=String(++r))}var s=function(t){function e(e){var i="Assertion failed. See https://openlayers.org/en/"+("v"+"5.3.0".split("-")[0])+"/doc/errors/#"+e+" for details.";t.call(this,i),this.code=e,this.name="AssertionError",this.message=i}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(Error);function a(t,e){if(!t)throw new s(e)}function h(t,e,i){return Math.min(Math.max(t,e),i)}var u="cosh"in Math?Math.cosh:function(t){var e=Math.exp(t);return(e+1/e)/2};function l(t,e,i,n,r,o){var s=r-i,a=o-n;if(0!==s||0!==a){var h=((t-i)*s+(e-n)*a)/(s*s+a*a);h>1?(i=r,n=o):h>0&&(i+=s*h,n+=a*h)}return c(t,e,i,n)}function c(t,e,i,n){var r=i-t,o=n-e;return r*r+o*o}function p(t,e,i){return t+i*(e-t)}var f=/^#([a-f0-9]{3}|[a-f0-9]{4}(?:[a-f0-9]{2}){0,2})$/i,d=/^([a-z]*)$/i;function _(t){return"string"==typeof t?t:y(t)}!function(){var t={},e=0}();function g(t){return t[0]=h(t[0]+.5|0,0,255),t[1]=h(t[1]+.5|0,0,255),t[2]=h(t[2]+.5|0,0,255),t[3]=h(t[3],0,1),t}function y(t){var e=t[0];e!=(0|e)&&(e=e+.5|0);var i=t[1];i!=(0|i)&&(i=i+.5|0);var n=t[2];return n!=(0|n)&&(n=n+.5|0),"rgba("+e+","+i+","+n+","+(void 0===t[3]?1:t[3])+")"}function v(t){return Array.isArray(t)?y(t):t}function m(t,e){var i=document.createElement("canvas");return t&&(i.width=t),e&&(i.height=e),i.getContext("2d")}var C="undefined"!=typeof navigator?navigator.userAgent.toLowerCase():"",x=(C.indexOf("firefox"),-1!==C.indexOf("safari")&&C.indexOf("chrom"),-1!==C.indexOf("webkit")&&C.indexOf("edge"),C.indexOf("macintosh"),window.devicePixelRatio,function(){var t=!1;try{t=!!document.createElement("canvas").getContext("2d").setLineDash}catch(t){}return t}()),w=(navigator,window,window,navigator.msPointerEnabled,2),E=function(){var t,e={};return function(i){if(t||(t=document.createElement("div").style),!(i in e)){t.font=i;var n=t.fontFamily;if(t.font="",!n)return null;e[i]=n.split(/,\s?/)}return e[i]}}(),F="function"==typeof Object.assign?Object.assign:function(t,e){var i=arguments;if(null==t)throw new TypeError("Cannot convert undefined or null to object");for(var n=Object(t),r=1,o=arguments.length;r<o;++r){var s=i[r];if(null!=s)for(var a in s)s.hasOwnProperty(a)&&(n[a]=s[a])}return n};function S(t){for(var e in t)delete t[e]}function k(t){var e=[];for(var i in t)e.push(t[i]);return e}function M(t){var e;for(e in t)return!1;return!e}var I=function(){this.disposed_=!1};I.prototype.dispose=function(){this.disposed_||(this.disposed_=!0,this.disposeInternal())},I.prototype.disposeInternal=function(){};var P=I;function R(t,e,i,n){for(var r,o=0,s=t.length;o<s;++o)if((r=t[o]).listener===e&&r.bindTo===i)return n&&(r.deleteIndex=o),r}function b(t,e){var i=D(t);return i?i[e]:void 0}function D(t,e){var i=t.ol_lm;return!i&&e&&(i=t.ol_lm={}),i}function A(t,e){var i=b(t,e);if(i){for(var n=0,r=i.length;n<r;++n)t.removeEventListener(e,i[n].boundListener),S(i[n]);i.length=0;var o=D(t);o&&(delete o[e],0===Object.keys(o).length&&function(t){delete t.ol_lm}(t))}}function O(t,e,i,n,r){var o=D(t,!0),s=o[e];s||(s=o[e]=[]);var a=R(s,i,n,!1);return a?r||(a.callOnce=!1):(a={bindTo:n,callOnce:!!r,listener:i,target:t,type:e},t.addEventListener(e,function(t){var e=function(e){var i=t.listener,n=t.bindTo||t.target;return t.callOnce&&L(t),i.call(n,e)};return t.boundListener=e,e}(a)),s.push(a)),a}function G(t,e,i,n){return O(t,e,i,n,!0)}function T(t,e,i,n){var r=b(t,e);if(r){var o=R(r,i,n,!0);o&&L(o)}}function L(t){if(t&&t.target){t.target.removeEventListener(t.type,t.boundListener);var e=b(t.target,t.type);if(e){var i="deleteIndex"in t?t.deleteIndex:e.indexOf(t);-1!==i&&e.splice(i,1),0===e.length&&A(t.target,t.type)}S(t)}}function X(){return!0}function Y(){return!1}function j(){}var N=function(t){this.propagationStopped,this.type=t,this.target=null};N.prototype.preventDefault=function(){this.propagationStopped=!0},N.prototype.stopPropagation=function(){this.propagationStopped=!0};var q=N,B=function(t){function e(){t.call(this),this.pendingRemovals_={},this.dispatching_={},this.listeners_={}}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.addEventListener=function(t,e){var i=this.listeners_[t];i||(i=this.listeners_[t]=[]),-1===i.indexOf(e)&&i.push(e)},e.prototype.dispatchEvent=function(t){var e="string"==typeof t?new q(t):t,i=e.type;e.target=this;var n,r=this.listeners_[i];if(r){i in this.dispatching_||(this.dispatching_[i]=0,this.pendingRemovals_[i]=0),++this.dispatching_[i];for(var o=0,s=r.length;o<s;++o)if(!1===r[o].call(this,e)||e.propagationStopped){n=!1;break}if(--this.dispatching_[i],0===this.dispatching_[i]){var a=this.pendingRemovals_[i];for(delete this.pendingRemovals_[i];a--;)this.removeEventListener(i,j);delete this.dispatching_[i]}return n}},e.prototype.disposeInternal=function(){!function(t){var e=D(t);if(e)for(var i in e)A(t,i)}(this)},e.prototype.getListeners=function(t){return this.listeners_[t]},e.prototype.hasListener=function(t){return t?t in this.listeners_:Object.keys(this.listeners_).length>0},e.prototype.removeEventListener=function(t,e){var i=this.listeners_[t];if(i){var n=i.indexOf(e);t in this.pendingRemovals_?(i[n]=j,++this.pendingRemovals_[t]):(i.splice(n,1),0===i.length&&delete this.listeners_[t])}},e}(P),W="change",V="clear",K="contextmenu",z="dblclick",U=function(t){function e(e){t.call(this),this.highWaterMark=void 0!==e?e:2048,this.count_=0,this.entries_={},this.oldest_=null,this.newest_=null}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.canExpireCache=function(){return this.getCount()>this.highWaterMark},e.prototype.clear=function(){this.count_=0,this.entries_={},this.oldest_=null,this.newest_=null,this.dispatchEvent(V)},e.prototype.containsKey=function(t){return this.entries_.hasOwnProperty(t)},e.prototype.forEach=function(t,e){for(var i=this.oldest_;i;)t.call(e,i.value_,i.key_,this),i=i.newer},e.prototype.get=function(t){var e=this.entries_[t];return a(void 0!==e,15),e===this.newest_||(e===this.oldest_?(this.oldest_=this.oldest_.newer,this.oldest_.older=null):(e.newer.older=e.older,e.older.newer=e.newer),e.newer=null,e.older=this.newest_,this.newest_.newer=e,this.newest_=e),e.value_},e.prototype.remove=function(t){var e=this.entries_[t];return a(void 0!==e,15),e===this.newest_?(this.newest_=e.older,this.newest_&&(this.newest_.newer=null)):e===this.oldest_?(this.oldest_=e.newer,this.oldest_&&(this.oldest_.older=null)):(e.newer.older=e.older,e.older.newer=e.newer),delete this.entries_[t],--this.count_,e.value_},e.prototype.getCount=function(){return this.count_},e.prototype.getKeys=function(){var t,e=new Array(this.count_),i=0;for(t=this.newest_;t;t=t.older)e[i++]=t.key_;return e},e.prototype.getValues=function(){var t,e=new Array(this.count_),i=0;for(t=this.newest_;t;t=t.older)e[i++]=t.value_;return e},e.prototype.peekLast=function(){return this.oldest_.value_},e.prototype.peekLastKey=function(){return this.oldest_.key_},e.prototype.peekFirstKey=function(){return this.newest_.key_},e.prototype.pop=function(){var t=this.oldest_;return delete this.entries_[t.key_],t.newer&&(t.newer.older=null),this.oldest_=t.newer,this.oldest_||(this.newest_=null),--this.count_,t.value_},e.prototype.replace=function(t,e){this.get(t),this.entries_[t].value_=e},e.prototype.set=function(t,e){a(!(t in this.entries_),16);var i={key_:t,newer:null,older:this.newest_,value_:e};this.newest_?this.newest_.newer=i:this.oldest_=i,this.newest_=i,this.entries_[t]=i,++this.count_},e.prototype.setSize=function(t){this.highWaterMark=t},e.prototype.prune=function(){for(;this.canExpireCache();)this.pop()},e}(B);new Array(6);var H=[0,0,0,1],J=[0,0,0,1],$=new U,Z={},Q=null,tt={};!function(){var t,e,i=Z,n=["monospace","serif"],r=n.length,o="wmytzilWMYTZIL@#/&?$%10";function s(t){for(var i=et(),s=100;s<=700;s+=300){for(var a=s+" ",h=!0,u=0;u<r;++u){var l=n[u];if(i.font=a+"32px "+l,e=i.measureText(o).width,t!=l){i.font=a+"32px "+t+","+l;var c=i.measureText(o).width;h=h&&c!=e}}if(h)return!0}return!1}function a(){var e=!0;for(var n in i)i[n]<60&&(s(n)?(i[n]=60,S(tt),Q=null,$.clear()):(++i[n],e=!1));e&&(clearInterval(t),t=void 0)}}();function et(){return Q||(Q=m(1,1)),Q}var it,nt;nt=tt;var rt=function(t){this.opacity_=t.opacity,this.rotateWithView_=t.rotateWithView,this.rotation_=t.rotation,this.scale_=t.scale};rt.prototype.clone=function(){return new rt({opacity:this.getOpacity(),scale:this.getScale(),rotation:this.getRotation(),rotateWithView:this.getRotateWithView()})},rt.prototype.getOpacity=function(){return this.opacity_},rt.prototype.getRotateWithView=function(){return this.rotateWithView_},rt.prototype.getRotation=function(){return this.rotation_},rt.prototype.getScale=function(){return this.scale_},rt.prototype.getSnapToPixel=function(){return!1},rt.prototype.getAnchor=function(){return n()},rt.prototype.getImage=function(t){return n()},rt.prototype.getHitDetectionImage=function(t){return n()},rt.prototype.getImageState=function(){return n()},rt.prototype.getImageSize=function(){return n()},rt.prototype.getHitDetectionImageSize=function(){return n()},rt.prototype.getOrigin=function(){return n()},rt.prototype.getSize=function(){return n()},rt.prototype.setOpacity=function(t){this.opacity_=t},rt.prototype.setRotateWithView=function(t){this.rotateWithView_=t},rt.prototype.setRotation=function(t){this.rotation_=t},rt.prototype.setScale=function(t){this.scale_=t},rt.prototype.setSnapToPixel=function(t){},rt.prototype.listenImageChange=function(t,e){return n()},rt.prototype.load=function(){n()},rt.prototype.unlistenImageChange=function(t,e){n()};var ot=function(t){function e(e){var i=e||{};t.call(this,{points:1/0,fill:i.fill,radius:i.radius,stroke:i.stroke,atlasManager:i.atlasManager})}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){var t=new e({fill:this.getFill()?this.getFill().clone():void 0,stroke:this.getStroke()?this.getStroke().clone():void 0,radius:this.getRadius(),atlasManager:this.atlasManager_});return t.setOpacity(this.getOpacity()),t.setScale(this.getScale()),t},e.prototype.setRadius=function(t){this.radius_=t,this.render_(this.atlasManager_)},e}(function(t){function e(e){var i=void 0!==e.rotateWithView&&e.rotateWithView;t.call(this,{opacity:1,rotateWithView:i,rotation:void 0!==e.rotation?e.rotation:0,scale:1}),this.checksums_=null,this.canvas_=null,this.hitDetectionCanvas_=null,this.fill_=void 0!==e.fill?e.fill:null,this.origin_=[0,0],this.points_=e.points,this.radius_=void 0!==e.radius?e.radius:e.radius1,this.radius2_=e.radius2,this.angle_=void 0!==e.angle?e.angle:0,this.stroke_=void 0!==e.stroke?e.stroke:null,this.anchor_=null,this.size_=null,this.imageSize_=null,this.hitDetectionImageSize_=null,this.atlasManager_=e.atlasManager,this.render_(this.atlasManager_)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){var t=new e({fill:this.getFill()?this.getFill().clone():void 0,points:this.getPoints(),radius:this.getRadius(),radius2:this.getRadius2(),angle:this.getAngle(),stroke:this.getStroke()?this.getStroke().clone():void 0,rotation:this.getRotation(),rotateWithView:this.getRotateWithView(),atlasManager:this.atlasManager_});return t.setOpacity(this.getOpacity()),t.setScale(this.getScale()),t},e.prototype.getAnchor=function(){return this.anchor_},e.prototype.getAngle=function(){return this.angle_},e.prototype.getFill=function(){return this.fill_},e.prototype.getHitDetectionImage=function(t){return this.hitDetectionCanvas_},e.prototype.getImage=function(t){return this.canvas_},e.prototype.getImageSize=function(){return this.imageSize_},e.prototype.getHitDetectionImageSize=function(){return this.hitDetectionImageSize_},e.prototype.getImageState=function(){return w},e.prototype.getOrigin=function(){return this.origin_},e.prototype.getPoints=function(){return this.points_},e.prototype.getRadius=function(){return this.radius_},e.prototype.getRadius2=function(){return this.radius2_},e.prototype.getSize=function(){return this.size_},e.prototype.getStroke=function(){return this.stroke_},e.prototype.listenImageChange=function(t,e){},e.prototype.load=function(){},e.prototype.unlistenImageChange=function(t,e){},e.prototype.render_=function(t){var e,i,n="",r="",o=0,s=null,a=0,h=0;this.stroke_&&(null===(i=this.stroke_.getColor())&&(i=J),i=v(i),void 0===(h=this.stroke_.getWidth())&&(h=1),s=this.stroke_.getLineDash(),a=this.stroke_.getLineDashOffset(),x||(s=null,a=0),void 0===(r=this.stroke_.getLineJoin())&&(r="round"),void 0===(n=this.stroke_.getLineCap())&&(n="round"),void 0===(o=this.stroke_.getMiterLimit())&&(o=10));var u=2*(this.radius_+h)+1,l={strokeStyle:i,strokeWidth:h,size:u,lineCap:n,lineDash:s,lineDashOffset:a,lineJoin:r,miterLimit:o};if(void 0===t){var c=m(u,u);this.canvas_=c.canvas,e=u=this.canvas_.width,this.draw_(l,c,0,0),this.createHitDetectionCanvas_(l)}else{u=Math.round(u);var p,f=!this.fill_;f&&(p=this.drawHitDetectionCanvas_.bind(this,l));var d=this.getChecksum(),_=t.add(d,u,u,this.draw_.bind(this,l),p);this.canvas_=_.image,this.origin_=[_.offsetX,_.offsetY],e=_.image.width,f?(this.hitDetectionCanvas_=_.hitImage,this.hitDetectionImageSize_=[_.hitImage.width,_.hitImage.height]):(this.hitDetectionCanvas_=this.canvas_,this.hitDetectionImageSize_=[e,e])}this.anchor_=[u/2,u/2],this.size_=[u,u],this.imageSize_=[e,e]},e.prototype.draw_=function(t,e,i,n){var r,o,s;e.setTransform(1,0,0,1,0,0),e.translate(i,n),e.beginPath();var a=this.points_;if(a===1/0)e.arc(t.size/2,t.size/2,this.radius_,0,2*Math.PI,!0);else{var h=void 0!==this.radius2_?this.radius2_:this.radius_;for(h!==this.radius_&&(a*=2),r=0;r<=a;r++)o=2*r*Math.PI/a-Math.PI/2+this.angle_,s=r%2==0?this.radius_:h,e.lineTo(t.size/2+s*Math.cos(o),t.size/2+s*Math.sin(o))}if(this.fill_){var u=this.fill_.getColor();null===u&&(u=H),e.fillStyle=v(u),e.fill()}this.stroke_&&(e.strokeStyle=t.strokeStyle,e.lineWidth=t.strokeWidth,t.lineDash&&(e.setLineDash(t.lineDash),e.lineDashOffset=t.lineDashOffset),e.lineCap=t.lineCap,e.lineJoin=t.lineJoin,e.miterLimit=t.miterLimit,e.stroke()),e.closePath()},e.prototype.createHitDetectionCanvas_=function(t){if(this.hitDetectionImageSize_=[t.size,t.size],this.fill_)this.hitDetectionCanvas_=this.canvas_;else{var e=m(t.size,t.size);this.hitDetectionCanvas_=e.canvas,this.drawHitDetectionCanvas_(t,e,0,0)}},e.prototype.drawHitDetectionCanvas_=function(t,e,i,n){e.setTransform(1,0,0,1,0,0),e.translate(i,n),e.beginPath();var r=this.points_;if(r===1/0)e.arc(t.size/2,t.size/2,this.radius_,0,2*Math.PI,!0);else{var o,s,a,h=void 0!==this.radius2_?this.radius2_:this.radius_;for(h!==this.radius_&&(r*=2),o=0;o<=r;o++)a=2*o*Math.PI/r-Math.PI/2+this.angle_,s=o%2==0?this.radius_:h,e.lineTo(t.size/2+s*Math.cos(a),t.size/2+s*Math.sin(a))}e.fillStyle=_(H),e.fill(),this.stroke_&&(e.strokeStyle=t.strokeStyle,e.lineWidth=t.strokeWidth,t.lineDash&&(e.setLineDash(t.lineDash),e.lineDashOffset=t.lineDashOffset),e.stroke()),e.closePath()},e.prototype.getChecksum=function(){var t=this.stroke_?this.stroke_.getChecksum():"-",e=this.fill_?this.fill_.getChecksum():"-";if(!this.checksums_||t!=this.checksums_[1]||e!=this.checksums_[2]||this.radius_!=this.checksums_[3]||this.radius2_!=this.checksums_[4]||this.angle_!=this.checksums_[5]||this.points_!=this.checksums_[6]){var i="r"+t+e+(void 0!==this.radius_?this.radius_.toString():"-")+(void 0!==this.radius2_?this.radius2_.toString():"-")+(void 0!==this.angle_?this.angle_.toString():"-")+(void 0!==this.points_?this.points_.toString():"-");this.checksums_=[i,t,e,this.radius_,this.radius2_,this.angle_,this.points_]}return this.checksums_[0]},e}(rt)),st="add",at="remove",ht="propertychange";var ut=function(t){function e(){t.call(this),this.revision_=0}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.changed=function(){++this.revision_,this.dispatchEvent(W)},e.prototype.getRevision=function(){return this.revision_},e.prototype.on=function(t,e){if(Array.isArray(t)){for(var i=t.length,n=new Array(i),r=0;r<i;++r)n[r]=O(this,t[r],e);return n}return O(this,t,e)},e.prototype.once=function(t,e){if(Array.isArray(t)){for(var i=t.length,n=new Array(i),r=0;r<i;++r)n[r]=G(this,t[r],e);return n}return G(this,t,e)},e.prototype.un=function(t,e){if(Array.isArray(t))for(var i=0,n=t.length;i<n;++i)T(this,t[i],e);else T(this,t,e)},e}(B),lt=function(t){function e(e,i,n){t.call(this,e),this.key=i,this.oldValue=n}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(q),ct={};function pt(t){return ct.hasOwnProperty(t)?ct[t]:ct[t]="change:"+t}var ft=function(t){function e(e){t.call(this),o(this),this.values_={},void 0!==e&&this.setProperties(e)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.get=function(t){var e;return this.values_.hasOwnProperty(t)&&(e=this.values_[t]),e},e.prototype.getKeys=function(){return Object.keys(this.values_)},e.prototype.getProperties=function(){return F({},this.values_)},e.prototype.notify=function(t,e){var i;i=pt(t),this.dispatchEvent(new lt(i,t,e)),i=ht,this.dispatchEvent(new lt(i,t,e))},e.prototype.set=function(t,e,i){if(i)this.values_[t]=e;else{var n=this.values_[t];this.values_[t]=e,n!==e&&this.notify(t,n)}},e.prototype.setProperties=function(t,e){for(var i in t)this.set(i,t[i],e)},e.prototype.unset=function(t,e){if(t in this.values_){var i=this.values_[t];delete this.values_[t],e||this.notify(t,i)}},e}(ut),dt="length",_t=function(t){function e(e,i){t.call(this,e),this.element=i}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(q),gt=function(t){function e(e,i){t.call(this);var n=i||{};if(this.unique_=!!n.unique,this.array_=e||[],this.unique_)for(var r=0,o=this.array_.length;r<o;++r)this.assertUnique_(this.array_[r],r);this.updateLength_()}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clear=function(){for(;this.getLength()>0;)this.pop()},e.prototype.extend=function(t){for(var e=0,i=t.length;e<i;++e)this.push(t[e]);return this},e.prototype.forEach=function(t){for(var e=this.array_,i=0,n=e.length;i<n;++i)t(e[i],i,e)},e.prototype.getArray=function(){return this.array_},e.prototype.item=function(t){return this.array_[t]},e.prototype.getLength=function(){return this.get(dt)},e.prototype.insertAt=function(t,e){this.unique_&&this.assertUnique_(e),this.array_.splice(t,0,e),this.updateLength_(),this.dispatchEvent(new _t(st,e))},e.prototype.pop=function(){return this.removeAt(this.getLength()-1)},e.prototype.push=function(t){this.unique_&&this.assertUnique_(t);var e=this.getLength();return this.insertAt(e,t),this.getLength()},e.prototype.remove=function(t){for(var e=this.array_,i=0,n=e.length;i<n;++i)if(e[i]===t)return this.removeAt(i)},e.prototype.removeAt=function(t){var e=this.array_[t];return this.array_.splice(t,1),this.updateLength_(),this.dispatchEvent(new _t(at,e)),e},e.prototype.setAt=function(t,e){var i=this.getLength();if(t<i){this.unique_&&this.assertUnique_(e,t);var n=this.array_[t];this.array_[t]=e,this.dispatchEvent(new _t(at,n)),this.dispatchEvent(new _t(st,e))}else{for(var r=i;r<t;++r)this.insertAt(r,void 0);this.insertAt(t,e)}},e.prototype.updateLength_=function(){this.set(dt,this.array_.length)},e.prototype.assertUnique_=function(t,e){for(var i=0,n=this.array_.length;i<n;++i)if(this.array_[i]===t&&i!==e)throw new s(58)},e}(ft);var yt=function(t){function e(e){if(t.call(this),this.id_=void 0,this.geometryName_="geometry",this.style_=null,this.styleFunction_=void 0,this.geometryChangeKey_=null,O(this,pt(this.geometryName_),this.handleGeometryChanged_,this),e)if("function"==typeof e.getSimplifiedGeometry){var i=e;this.setGeometry(i)}else{var n=e;this.setProperties(n)}}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){var t=new e(this.getProperties());t.setGeometryName(this.getGeometryName());var i=this.getGeometry();i&&t.setGeometry(i.clone());var n=this.getStyle();return n&&t.setStyle(n),t},e.prototype.getGeometry=function(){return this.get(this.geometryName_)},e.prototype.getId=function(){return this.id_},e.prototype.getGeometryName=function(){return this.geometryName_},e.prototype.getStyle=function(){return this.style_},e.prototype.getStyleFunction=function(){return this.styleFunction_},e.prototype.handleGeometryChange_=function(){this.changed()},e.prototype.handleGeometryChanged_=function(){this.geometryChangeKey_&&(L(this.geometryChangeKey_),this.geometryChangeKey_=null);var t=this.getGeometry();t&&(this.geometryChangeKey_=O(t,W,this.handleGeometryChange_,this)),this.changed()},e.prototype.setGeometry=function(t){this.set(this.geometryName_,t)},e.prototype.setStyle=function(t){this.style_=t,this.styleFunction_=t?function(t){if("function"==typeof t)return t;var e;Array.isArray(t)?e=t:(a("function"==typeof t.getZIndex,41),e=[t]);return function(){return e}}(t):void 0,this.changed()},e.prototype.setId=function(t){this.id_=t,this.changed()},e.prototype.setGeometryName=function(t){T(this,pt(this.geometryName_),this.handleGeometryChanged_,this),this.geometryName_=t,O(this,pt(this.geometryName_),this.handleGeometryChanged_,this),this.handleGeometryChanged_()},e}(ft),vt="singleclick",mt=z,Ct="pointerdrag",xt="pointermove",wt="pointerdown",Et="pointerup",Ft=function(t){function e(e,i,n,r,o){t.call(this,e,i,n.originalEvent,r,o),this.pointerEvent=n}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(function(t){function e(e,i,n,r,o){t.call(this,e,i,o),this.originalEvent=n,this.pixel=i.getEventPixel(n),this.coordinate=i.getCoordinateFromPixel(this.pixel),this.dragging=void 0!==r&&r}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.preventDefault=function(){t.prototype.preventDefault.call(this),this.originalEvent.preventDefault()},e.prototype.stopPropagation=function(){t.prototype.stopPropagation.call(this),this.originalEvent.stopPropagation()},e}(function(t){function e(e,i,n){t.call(this,e),this.map=i,this.frameState=void 0!==n?n:null}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(q)));function St(t,e){var i,n,r=t[0],o=t[1],s=e[0],a=e[1],h=s[0],u=s[1],l=a[0],c=a[1],p=l-h,f=c-u,d=0===p&&0===f?0:(p*(r-h)+f*(o-u))/(p*p+f*f||0);return d<=0?(i=h,n=u):d>=1?(i=l,n=c):(i=h+d*p,n=u+d*f),[i,n]}function kt(t,e){for(var i=!0,n=t.length-1;n>=0;--n)if(t[n]!=e[n]){i=!1;break}return i}function Mt(t,e){var i=t[0]-e[0],n=t[1]-e[1];return i*i+n*n}function It(t,e){return Math.sqrt(Mt(t,e))}var Pt=X,Rt=Y,bt=function(t){var e=t.originalEvent;return!e.altKey&&!(e.metaKey||e.ctrlKey)&&!e.shiftKey},Dt=function(t){var e=t.originalEvent;return!e.altKey&&!(e.metaKey||e.ctrlKey)&&e.shiftKey},At=function(t){var e=t.pointerEvent;return a(void 0!==e,56),e.isPrimary&&0===e.button},Ot=0,Gt=1,Tt=2,Lt=4,Xt=8,Yt=16;function jt(t){for(var e=Vt(),i=0,n=t.length;i<n;++i)$t(e,t[i]);return e}function Nt(t,e,i){var n,r;return(n=e<t[0]?t[0]-e:t[2]<e?e-t[2]:0)*n+(r=i<t[1]?t[1]-i:t[3]<i?i-t[3]:0)*r}function qt(t,e){return t[0]<=e[0]&&e[2]<=t[2]&&t[1]<=e[1]&&e[3]<=t[3]}function Bt(t,e,i){return t[0]<=e&&e<=t[2]&&t[1]<=i&&i<=t[3]}function Wt(t,e){var i=t[0],n=t[1],r=t[2],o=t[3],s=e[0],a=e[1],h=Ot;return s<i?h|=Yt:s>r&&(h|=Lt),a<n?h|=Xt:a>o&&(h|=Tt),h===Ot&&(h=Gt),h}function Vt(){return[1/0,1/0,-1/0,-1/0]}function Kt(t,e,i,n,r){return r?(r[0]=t,r[1]=e,r[2]=i,r[3]=n,r):[t,e,i,n]}function zt(t){return Kt(1/0,1/0,-1/0,-1/0,t)}function Ut(t,e){var i=t[0],n=t[1];return Kt(i,n,i,n,e)}function Ht(t,e,i,n,r){return Zt(zt(r),t,e,i,n)}function Jt(t,e){return t[0]==e[0]&&t[2]==e[2]&&t[1]==e[1]&&t[3]==e[3]}function $t(t,e){e[0]<t[0]&&(t[0]=e[0]),e[0]>t[2]&&(t[2]=e[0]),e[1]<t[1]&&(t[1]=e[1]),e[1]>t[3]&&(t[3]=e[1])}function Zt(t,e,i,n,r){for(;i<n;i+=r)Qt(t,e[i],e[i+1]);return t}function Qt(t,e,i){t[0]=Math.min(t[0],e),t[1]=Math.min(t[1],i),t[2]=Math.max(t[2],e),t[3]=Math.max(t[3],i)}function te(t,e,i){var n;return(n=e.call(i,ee(t)))||(n=e.call(i,ie(t)))||(n=e.call(i,se(t)))?n:(n=e.call(i,oe(t)))||!1}function ee(t){return[t[0],t[1]]}function ie(t){return[t[2],t[1]]}function ne(t){return[(t[0]+t[2])/2,(t[1]+t[3])/2]}function re(t){return t[3]-t[1]}function oe(t){return[t[0],t[3]]}function se(t){return[t[2],t[3]]}function ae(t,e){return t[0]<=e[2]&&t[2]>=e[0]&&t[1]<=e[3]&&t[3]>=e[1]}var he="Point",ue="LineString",le="LinearRing",ce="Polygon",pe="MultiPoint",fe="MultiLineString",de="MultiPolygon",_e="GeometryCollection",ge="Circle",ye="Rectangle",ve="Ellipse";function me(t,e,i,n,r,o){for(var s=o||[],a=0,h=e;h<i;h+=n){var u=t[h],l=t[h+1];s[a++]=r[0]*u+r[2]*l+r[4],s[a++]=r[1]*u+r[3]*l+r[5]}return o&&s.length!=a&&(s.length=a),s}var Ce={DEGREES:"degrees",FEET:"ft",METERS:"m",PIXELS:"pixels",TILE_PIXELS:"tile-pixels",USFEET:"us-ft"},xe={};xe[Ce.DEGREES]=2*Math.PI*6370997/360,xe[Ce.FEET]=.3048,xe[Ce.METERS]=1,xe[Ce.USFEET]=1200/3937;var we=Ce,Ee=function(t){this.code_=t.code,this.units_=t.units,this.extent_=void 0!==t.extent?t.extent:null,this.worldExtent_=void 0!==t.worldExtent?t.worldExtent:null,this.axisOrientation_=void 0!==t.axisOrientation?t.axisOrientation:"enu",this.global_=void 0!==t.global&&t.global,this.canWrapX_=!(!this.global_||!this.extent_),this.getPointResolutionFunc_=t.getPointResolution,this.defaultTileGrid_=null,this.metersPerUnit_=t.metersPerUnit};Ee.prototype.canWrapX=function(){return this.canWrapX_},Ee.prototype.getCode=function(){return this.code_},Ee.prototype.getExtent=function(){return this.extent_},Ee.prototype.getUnits=function(){return this.units_},Ee.prototype.getMetersPerUnit=function(){return this.metersPerUnit_||xe[this.units_]},Ee.prototype.getWorldExtent=function(){return this.worldExtent_},Ee.prototype.getAxisOrientation=function(){return this.axisOrientation_},Ee.prototype.isGlobal=function(){return this.global_},Ee.prototype.setGlobal=function(t){this.global_=t,this.canWrapX_=!(!t||!this.extent_)},Ee.prototype.getDefaultTileGrid=function(){return this.defaultTileGrid_},Ee.prototype.setDefaultTileGrid=function(t){this.defaultTileGrid_=t},Ee.prototype.setExtent=function(t){this.extent_=t,this.canWrapX_=!(!this.global_||!t)},Ee.prototype.setWorldExtent=function(t){this.worldExtent_=t},Ee.prototype.setGetPointResolution=function(t){this.getPointResolutionFunc_=t},Ee.prototype.getPointResolutionFunc=function(){return this.getPointResolutionFunc_};var Fe=Ee,Se=6378137*Math.PI,ke=[-Se,-Se,Se,Se],Me=[-180,-85,180,85],Ie=function(t){function e(e){t.call(this,{code:e,units:we.METERS,extent:ke,global:!0,worldExtent:Me,getPointResolution:function(t,e){return t/u(e[1]/6378137)}})}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(Fe),Pe=[new Ie("EPSG:3857"),new Ie("EPSG:102100"),new Ie("EPSG:102113"),new Ie("EPSG:900913"),new Ie("urn:ogc:def:crs:EPSG:6.18:3:3857"),new Ie("urn:ogc:def:crs:EPSG::3857"),new Ie("http://www.opengis.net/gml/srs/epsg.xml#3857")];function Re(t,e,i){var n=t.length,r=i>1?i:2,o=e;void 0===o&&(o=r>2?t.slice():new Array(n));for(var s=Se,a=0;a<n;a+=r){o[a]=s*t[a]/180;var h=6378137*Math.log(Math.tan(Math.PI*(t[a+1]+90)/360));h>s?h=s:h<-s&&(h=-s),o[a+1]=h}return o}function be(t,e,i){var n=t.length,r=i>1?i:2,o=e;void 0===o&&(o=r>2?t.slice():new Array(n));for(var s=0;s<n;s+=r)o[s]=180*t[s]/Se,o[s+1]=360*Math.atan(Math.exp(t[s+1]/6378137))/Math.PI-90;return o}var De=[-180,-90,180,90],Ae=6378137*Math.PI/180,Oe=function(t){function e(e,i){t.call(this,{code:e,units:we.DEGREES,extent:De,axisOrientation:i,global:!0,metersPerUnit:Ae,worldExtent:De})}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(Fe),Ge=[new Oe("CRS:84"),new Oe("EPSG:4326","neu"),new Oe("urn:ogc:def:crs:EPSG::4326","neu"),new Oe("urn:ogc:def:crs:EPSG:6.6:4326","neu"),new Oe("urn:ogc:def:crs:OGC:1.3:CRS84"),new Oe("urn:ogc:def:crs:OGC:2:84"),new Oe("http://www.opengis.net/gml/srs/epsg.xml#4326","neu"),new Oe("urn:x-ogc:def:crs:EPSG:4326","neu")],Te={};var Le,Xe,Ye,je={};function Ne(t,e,i){var n=t.getCode(),r=e.getCode();n in je||(je[n]={}),je[n][r]=i}function qe(t,e,i){var n;if(void 0!==e){for(var r=0,o=t.length;r<o;++r)e[r]=t[r];n=e}else n=t.slice();return n}function Be(t,e,i){if(void 0!==e&&t!==e){for(var n=0,r=t.length;n<r;++n)e[n]=t[n];t=e}return t}function We(t){!function(t,e){Te[t]=e}(t.getCode(),t),Ne(t,t,qe)}function Ve(t){return"string"==typeof t?Te[t]||null:t||null}function Ke(t){!function(t){t.forEach(We)}(t),t.forEach((function(e){t.forEach((function(t){e!==t&&Ne(e,t,qe)}))}))}function ze(t,e){var i=function(t,e){var i;return t in je&&e in je[t]&&(i=je[t][e]),i}(t.getCode(),e.getCode());return i||(i=Be),i}function Ue(t,e){return ze(Ve(t),Ve(e))}Ke(Pe),Ke(Ge),Le=Pe,Xe=Re,Ye=be,Ge.forEach((function(t){Le.forEach((function(e){Ne(t,e,Xe),Ne(e,t,Ye)}))}));var He=[1,0,0,1,0,0],Je="XY",$e="XYZ",Ze="XYM",Qe="XYZM";function ti(t){var e;return t==Je?e=2:t==$e||t==Ze?e=3:t==Qe&&(e=4),e}var ei=function(t){function e(){t.call(this),this.layout=Je,this.stride=2,this.flatCoordinates=null}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.computeExtent=function(t){return Ht(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t)},e.prototype.getCoordinates=function(){return n()},e.prototype.getFirstCoordinate=function(){return this.flatCoordinates.slice(0,this.stride)},e.prototype.getFlatCoordinates=function(){return this.flatCoordinates},e.prototype.getLastCoordinate=function(){return this.flatCoordinates.slice(this.flatCoordinates.length-this.stride)},e.prototype.getLayout=function(){return this.layout},e.prototype.getSimplifiedGeometry=function(t){if(this.simplifiedGeometryRevision!=this.getRevision()&&(S(this.simplifiedGeometryCache),this.simplifiedGeometryMaxMinSquaredTolerance=0,this.simplifiedGeometryRevision=this.getRevision()),t<0||0!==this.simplifiedGeometryMaxMinSquaredTolerance&&t<=this.simplifiedGeometryMaxMinSquaredTolerance)return this;var e=t.toString();if(this.simplifiedGeometryCache.hasOwnProperty(e))return this.simplifiedGeometryCache[e];var i=this.getSimplifiedGeometryInternal(t);return i.getFlatCoordinates().length<this.flatCoordinates.length?(this.simplifiedGeometryCache[e]=i,i):(this.simplifiedGeometryMaxMinSquaredTolerance=t,this)},e.prototype.getSimplifiedGeometryInternal=function(t){return this},e.prototype.getStride=function(){return this.stride},e.prototype.setFlatCoordinates=function(t,e){this.stride=ti(t),this.layout=t,this.flatCoordinates=e},e.prototype.setCoordinates=function(t,e){n()},e.prototype.setLayout=function(t,e,i){var n;if(t)n=ti(t);else{for(var r=0;r<i;++r){if(0===e.length)return this.layout=Je,void(this.stride=2);e=e[0]}t=function(t){var e;2==t?e=Je:3==t?e=$e:4==t&&(e=Qe);return e}(n=e.length)}this.layout=t,this.stride=n},e.prototype.applyTransform=function(t){this.flatCoordinates&&(t(this.flatCoordinates,this.flatCoordinates,this.stride),this.changed())},e.prototype.rotate=function(t,e){var i=this.getFlatCoordinates();if(i){var n=this.getStride();!function(t,e,i,n,r,o,s){for(var a=s||[],h=Math.cos(r),u=Math.sin(r),l=o[0],c=o[1],p=0,f=e;f<i;f+=n){var d=t[f]-l,_=t[f+1]-c;a[p++]=l+d*h-_*u,a[p++]=c+d*u+_*h;for(var g=f+2;g<f+n;++g)a[p++]=t[g]}s&&a.length!=p&&(a.length=p)}(i,0,i.length,n,t,e,i),this.changed()}},e.prototype.scale=function(t,e,i){var n=e;void 0===n&&(n=t);var r=i;r||(r=ne(this.getExtent()));var o=this.getFlatCoordinates();if(o){var s=this.getStride();!function(t,e,i,n,r,o,s,a){for(var h=a||[],u=s[0],l=s[1],c=0,p=e;p<i;p+=n){var f=t[p]-u,d=t[p+1]-l;h[c++]=u+r*f,h[c++]=l+o*d;for(var _=p+2;_<p+n;++_)h[c++]=t[_]}a&&h.length!=c&&(h.length=c)}(o,0,o.length,s,t,n,r,o),this.changed()}},e.prototype.translate=function(t,e){var i=this.getFlatCoordinates();if(i){var n=this.getStride();!function(t,e,i,n,r,o,s){for(var a=s||[],h=0,u=e;u<i;u+=n){a[h++]=t[u]+r,a[h++]=t[u+1]+o;for(var l=u+2;l<u+n;++l)a[h++]=t[l]}s&&a.length!=h&&(a.length=h)}(i,0,i.length,n,t,e,i),this.changed()}},e}(function(t){function e(){t.call(this),this.extent_=[1/0,1/0,-1/0,-1/0],this.extentRevision_=-1,this.simplifiedGeometryCache={},this.simplifiedGeometryMaxMinSquaredTolerance=0,this.simplifiedGeometryRevision=0}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){return n()},e.prototype.closestPointXY=function(t,e,i,r){return n()},e.prototype.containsXY=function(t,e){return!1},e.prototype.getClosestPoint=function(t,e){var i=e||[NaN,NaN];return this.closestPointXY(t[0],t[1],i,1/0),i},e.prototype.intersectsCoordinate=function(t){return this.containsXY(t[0],t[1])},e.prototype.computeExtent=function(t){return n()},e.prototype.getExtent=function(t){return this.extentRevision_!=this.getRevision()&&(this.extent_=this.computeExtent(this.extent_),this.extentRevision_=this.getRevision()),function(t,e){return e?(e[0]=t[0],e[1]=t[1],e[2]=t[2],e[3]=t[3],e):t}(this.extent_,t)},e.prototype.rotate=function(t,e){n()},e.prototype.scale=function(t,e,i){n()},e.prototype.simplify=function(t){return this.getSimplifiedGeometry(t*t)},e.prototype.getSimplifiedGeometry=function(t){return n()},e.prototype.getType=function(){return n()},e.prototype.applyTransform=function(t){n()},e.prototype.intersectsExtent=function(t){return n()},e.prototype.translate=function(t,e){n()},e.prototype.transform=function(t,e){var i=Ve(t),n=i.getUnits()==we.TILE_PIXELS?function(t,n,r){var o=i.getExtent(),s=i.getWorldExtent(),a=re(s)/re(o);return function(t,e,i,n,r,o,s,a){var h=Math.sin(o),u=Math.cos(o);t[0]=n*u,t[1]=r*h,t[2]=-n*h,t[3]=r*u,t[4]=s*n*u-a*n*h+e,t[5]=s*r*h+a*r*u+i}(He,s[0],s[3],a,-a,0,0,0),me(t,0,t.length,r,He,n),Ue(i,e)(t,n,r)}:Ue(i,e);return this.applyTransform(n),this},e}(ft));function ii(t,e,i,n){for(var r=0,o=i.length;r<o;++r)t[e++]=i[r];return e}function ni(t,e,i,n){for(var r=0,o=i.length;r<o;++r)for(var s=i[r],a=0;a<n;++a)t[e++]=s[a];return e}function ri(t,e,i,n,r){for(var o=r||[],s=0,a=0,h=i.length;a<h;++a){var u=ni(t,e,i[a],n);o[s++]=u,e=u}return o.length=s,o}var oi=function(t){function e(e,i,n){if(t.call(this),void 0!==n&&void 0===i)this.setFlatCoordinates(n,e);else{var r=i||0;this.setCenterAndRadius(e,r,n)}}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),void 0,this.layout)},e.prototype.closestPointXY=function(t,e,i,n){var r=this.flatCoordinates,o=t-r[0],s=e-r[1],a=o*o+s*s;if(a<n){if(0===a)for(var h=0;h<this.stride;++h)i[h]=r[h];else{var u=this.getRadius()/Math.sqrt(a);i[0]=r[0]+u*o,i[1]=r[1]+u*s;for(var l=2;l<this.stride;++l)i[l]=r[l]}return i.length=this.stride,a}return n},e.prototype.containsXY=function(t,e){var i=this.flatCoordinates,n=t-i[0],r=e-i[1];return n*n+r*r<=this.getRadiusSquared_()},e.prototype.getCenter=function(){return this.flatCoordinates.slice(0,this.stride)},e.prototype.computeExtent=function(t){var e=this.flatCoordinates,i=e[this.stride]-e[0];return Kt(e[0]-i,e[1]-i,e[0]+i,e[1]+i,t)},e.prototype.getRadius=function(){return Math.sqrt(this.getRadiusSquared_())},e.prototype.getRadiusSquared_=function(){var t=this.flatCoordinates[this.stride]-this.flatCoordinates[0],e=this.flatCoordinates[this.stride+1]-this.flatCoordinates[1];return t*t+e*e},e.prototype.getType=function(){return ge},e.prototype.intersectsExtent=function(t){if(ae(t,this.getExtent())){var e=this.getCenter();return t[0]<=e[0]&&t[2]>=e[0]||(t[1]<=e[1]&&t[3]>=e[1]||te(t,this.intersectsCoordinate,this))}return!1},e.prototype.setCenter=function(t){var e=this.stride,i=this.flatCoordinates[e]-this.flatCoordinates[0],n=t.slice();n[e]=n[0]+i;for(var r=1;r<e;++r)n[e+r]=t[r];this.setFlatCoordinates(this.layout,n),this.changed()},e.prototype.setCenterAndRadius=function(t,e,i){this.setLayout(i,t,0),this.flatCoordinates||(this.flatCoordinates=[]);var n=this.flatCoordinates,r=ii(n,0,t,this.stride);n[r++]=n[0]+e;for(var o=1,s=this.stride;o<s;++o)n[r++]=n[o];n.length=r,this.changed()},e.prototype.getCoordinates=function(){return null},e.prototype.setCoordinates=function(t,e){},e.prototype.setRadius=function(t){this.flatCoordinates[this.stride]=this.flatCoordinates[0]+t,this.changed()},e}(ei);oi.prototype.transform;var si=oi;function ai(t,e){return t>e?1:t<e?-1:0}function hi(t,e){for(var i=Array.isArray(e)?e:[e],n=i.length,r=0;r<n;r++)t[t.length]=i[r]}function ui(t,e,i,n,r,o,s){var a,h=t[e],u=t[e+1],l=t[i]-h,c=t[i+1]-u;if(0===l&&0===c)a=e;else{var f=((r-h)*l+(o-u)*c)/(l*l+c*c);if(f>1)a=i;else{if(f>0){for(var d=0;d<n;++d)s[d]=p(t[e+d],t[i+d],f);return void(s.length=n)}a=e}}for(var _=0;_<n;++_)s[_]=t[a+_];s.length=n}function li(t,e,i,n,r){var o=t[e],s=t[e+1];for(e+=n;e<i;e+=n){var a=t[e],h=t[e+1],u=c(o,s,a,h);u>r&&(r=u),o=a,s=h}return r}function ci(t,e,i,n,r){for(var o=0,s=i.length;o<s;++o){var a=i[o];r=li(t,e,a,n,r),e=a}return r}function pi(t,e,i,n,r,o,s,a,h,u,l){if(e==i)return u;var p,f;if(0===r){if((f=c(s,a,t[e],t[e+1]))<u){for(p=0;p<n;++p)h[p]=t[e+p];return h.length=n,f}return u}for(var d=l||[NaN,NaN],_=e+n;_<i;)if(ui(t,_-n,_,n,s,a,d),(f=c(s,a,d[0],d[1]))<u){for(u=f,p=0;p<n;++p)h[p]=d[p];h.length=n,_+=n}else _+=n*Math.max((Math.sqrt(f)-Math.sqrt(u))/r|0,1);if(o&&(ui(t,i-n,e,n,s,a,d),(f=c(s,a,d[0],d[1]))<u)){for(u=f,p=0;p<n;++p)h[p]=d[p];h.length=n}return u}function fi(t,e,i,n,r,o,s,a,h,u,l){for(var c=l||[NaN,NaN],p=0,f=i.length;p<f;++p){var d=i[p];u=pi(t,e,d,n,r,o,s,a,h,u,c),e=d}return u}function di(t,e,i,n,r){for(var o=void 0!==r?r:[],s=0,a=e;a<i;a+=n)o[s++]=t.slice(a,a+n);return o.length=s,o}function _i(t,e,i,n,r){for(var o=void 0!==r?r:[],s=0,a=0,h=i.length;a<h;++a){var u=i[a];o[s++]=di(t,e,u,n,o[s]),e=u}return o.length=s,o}function gi(t,e,i,n,r,o){var s=NaN,a=NaN,h=(i-e)/n;if(1===h)s=t[e],a=t[e+1];else if(2==h)s=(1-r)*t[e]+r*t[e+n],a=(1-r)*t[e+1]+r*t[e+n+1];else if(0!==h){for(var u=t[e],l=t[e+1],c=0,f=[0],d=e+n;d<i;d+=n){var _=t[d],g=t[d+1];c+=Math.sqrt((_-u)*(_-u)+(g-l)*(g-l)),f.push(c),u=_,l=g}var y=r*c,v=function(t,e,i){for(var n,r,o=i||ai,s=0,a=t.length,h=!1;s<a;)(r=+o(t[n=s+(a-s>>1)],e))<0?s=n+1:(a=n,h=!r);return h?s:~s}(f,y);if(v<0){var m=(y-f[-v-2])/(f[-v-1]-f[-v-2]),C=e+(-v-2)*n;s=p(t[C],t[C+n],m),a=p(t[C+1],t[C+n+1],m)}else s=t[e+v*n],a=t[e+v*n+1]}return o?(o[0]=s,o[1]=a,o):[s,a]}function yi(t,e,i,n,r,o){if(i==e)return null;var s;if(r<t[e+n-1])return o?((s=t.slice(e,e+n))[n-1]=r,s):null;if(t[i-1]<r)return o?((s=t.slice(i-n,i))[n-1]=r,s):null;if(r==t[e+n-1])return t.slice(e,e+n);for(var a=e/n,h=i/n;a<h;){var u=a+h>>1;r<t[(u+1)*n-1]?h=u:a=u+1}var l=t[a*n-1];if(r==l)return t.slice((a-1)*n,(a-1)*n+n);var c=(r-l)/(t[(a+1)*n-1]-l);s=[];for(var f=0;f<n-1;++f)s.push(p(t[(a-1)*n+f],t[a*n+f],c));return s.push(r),s}function vi(t,e,i,n,r){return!te(r,(function(r){return!mi(t,e,i,n,r[0],r[1])}))}function mi(t,e,i,n,r,o){for(var s=0,a=t[i-n],h=t[i-n+1];e<i;e+=n){var u=t[e],l=t[e+1];h<=o?l>o&&(u-a)*(o-h)-(r-a)*(l-h)>0&&s++:l<=o&&(u-a)*(o-h)-(r-a)*(l-h)<0&&s--,a=u,h=l}return 0!==s}function Ci(t,e,i,n,r,o){if(0===i.length)return!1;if(!mi(t,e,i[0],n,r,o))return!1;for(var s=1,a=i.length;s<a;++s)if(mi(t,i[s-1],i[s],n,r,o))return!1;return!0}function xi(t,e,i,n,r,o){for(var s,a=[t[e],t[e+1]],h=[];e+n<i;e+=n){if(h[0]=t[e+n],h[1]=t[e+n+1],s=r.call(o,a,h))return s;a[0]=h[0],a[1]=h[1]}return!1}function wi(t,e,i,n,r){var o=Zt([1/0,1/0,-1/0,-1/0],t,e,i,n);return!!ae(r,o)&&(!!qt(r,o)||(o[0]>=r[0]&&o[2]<=r[2]||(o[1]>=r[1]&&o[3]<=r[3]||xi(t,e,i,n,(function(t,e){return function(t,e,i){var n=!1,r=Wt(t,e),o=Wt(t,i);if(r===Gt||o===Gt)n=!0;else{var s,a,h=t[0],u=t[1],l=t[2],c=t[3],p=e[0],f=e[1],d=i[0],_=i[1],g=(_-f)/(d-p);o&Tt&&!(r&Tt)&&(n=(s=d-(_-c)/g)>=h&&s<=l),n||!(o&Lt)||r&Lt||(n=(a=_-(d-l)*g)>=u&&a<=c),n||!(o&Xt)||r&Xt||(n=(s=d-(_-u)/g)>=h&&s<=l),n||!(o&Yt)||r&Yt||(n=(a=_-(d-h)*g)>=u&&a<=c)}return n}(r,t,e)})))))}function Ei(t,e,i,n,r){if(!function(t,e,i,n,r){return!!wi(t,e,i,n,r)||(!!mi(t,e,i,n,r[0],r[1])||(!!mi(t,e,i,n,r[0],r[3])||(!!mi(t,e,i,n,r[2],r[1])||!!mi(t,e,i,n,r[2],r[3]))))}(t,e,i[0],n,r))return!1;if(1===i.length)return!0;for(var o=1,s=i.length;o<s;++o)if(vi(t,i[o-1],i[o],n,r)&&!wi(t,i[o-1],i[o],n,r))return!1;return!0}function Fi(t,e,i,n){for(var r=t[e],o=t[e+1],s=0,a=e+n;a<i;a+=n){var h=t[a],u=t[a+1];s+=Math.sqrt((h-r)*(h-r)+(u-o)*(u-o)),r=h,o=u}return s}function Si(t,e,i,n,r,o,s){var a=(i-e)/n;if(a<3){for(;e<i;e+=n)o[s++]=t[e],o[s++]=t[e+1];return s}var h=new Array(a);h[0]=1,h[a-1]=1;for(var u=[e,i-n],c=0;u.length>0;){for(var p=u.pop(),f=u.pop(),d=0,_=t[f],g=t[f+1],y=t[p],v=t[p+1],m=f+n;m<p;m+=n){var C=l(t[m],t[m+1],_,g,y,v);C>d&&(c=m,d=C)}d>r&&(h[(c-e)/n]=1,f+n<c&&u.push(f,c),c+n<p&&u.push(c,p))}for(var x=0;x<a;++x)h[x]&&(o[s++]=t[e+x*n],o[s++]=t[e+x*n+1]);return s}function ki(t,e,i,n,r,o,s,a){for(var h=0,u=i.length;h<u;++h){var l=i[h];s=Si(t,e,l,n,r,o,s),a.push(s),e=l}return s}function Mi(t,e){return e*Math.round(t/e)}function Ii(t,e,i,n,r,o,s){if(e==i)return s;var a,h,u=Mi(t[e],r),l=Mi(t[e+1],r);e+=n,o[s++]=u,o[s++]=l;do{if(a=Mi(t[e],r),h=Mi(t[e+1],r),(e+=n)==i)return o[s++]=a,o[s++]=h,s}while(a==u&&h==l);for(;e<i;){var c=Mi(t[e],r),p=Mi(t[e+1],r);if(e+=n,c!=a||p!=h){var f=a-u,d=h-l,_=c-u,g=p-l;f*g==d*_&&(f<0&&_<f||f==_||f>0&&_>f)&&(d<0&&g<d||d==g||d>0&&g>d)?(a=c,h=p):(o[s++]=a,o[s++]=h,u=a,l=h,a=c,h=p)}}return o[s++]=a,o[s++]=h,s}function Pi(t,e,i,n,r,o,s,a){for(var h=0,u=i.length;h<u;++h){var l=i[h];s=Ii(t,e,l,n,r,o,s),a.push(s),e=l}return s}var Ri=function(t){function e(e,i){t.call(this),this.flatMidpoint_=null,this.flatMidpointRevision_=-1,this.maxDelta_=-1,this.maxDeltaRevision_=-1,void 0===i||Array.isArray(e[0])?this.setCoordinates(e,i):this.setFlatCoordinates(i,e)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.appendCoordinate=function(t){this.flatCoordinates?hi(this.flatCoordinates,t):this.flatCoordinates=t.slice(),this.changed()},e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),this.layout)},e.prototype.closestPointXY=function(t,e,i,n){return n<Nt(this.getExtent(),t,e)?n:(this.maxDeltaRevision_!=this.getRevision()&&(this.maxDelta_=Math.sqrt(li(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,0)),this.maxDeltaRevision_=this.getRevision()),pi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,this.maxDelta_,!1,t,e,i,n))},e.prototype.forEachSegment=function(t){return xi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t)},e.prototype.getCoordinateAtM=function(t,e){if(this.layout!=Ze&&this.layout!=Qe)return null;var i=void 0!==e&&e;return yi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t,i)},e.prototype.getCoordinates=function(){return di(this.flatCoordinates,0,this.flatCoordinates.length,this.stride)},e.prototype.getCoordinateAt=function(t,e){return gi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t,e)},e.prototype.getLength=function(){return Fi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride)},e.prototype.getFlatMidpoint=function(){return this.flatMidpointRevision_!=this.getRevision()&&(this.flatMidpoint_=this.getCoordinateAt(.5,this.flatMidpoint_),this.flatMidpointRevision_=this.getRevision()),this.flatMidpoint_},e.prototype.getSimplifiedGeometryInternal=function(t){var i=[];return i.length=Si(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t,i,0),new e(i,Je)},e.prototype.getType=function(){return ue},e.prototype.intersectsExtent=function(t){return wi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t)},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,1),this.flatCoordinates||(this.flatCoordinates=[]),this.flatCoordinates.length=ni(this.flatCoordinates,0,t,this.stride),this.changed()},e}(ei),bi=function(t){function e(e,i,n){if(t.call(this),this.ends_=[],this.maxDelta_=-1,this.maxDeltaRevision_=-1,Array.isArray(e[0]))this.setCoordinates(e,i);else if(void 0!==i&&n)this.setFlatCoordinates(i,e),this.ends_=n;else{for(var r=this.getLayout(),o=e,s=[],a=[],h=0,u=o.length;h<u;++h){var l=o[h];0===h&&(r=l.getLayout()),hi(s,l.getFlatCoordinates()),a.push(s.length)}this.setFlatCoordinates(r,s),this.ends_=a}}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.appendLineString=function(t){this.flatCoordinates?hi(this.flatCoordinates,t.getFlatCoordinates().slice()):this.flatCoordinates=t.getFlatCoordinates().slice(),this.ends_.push(this.flatCoordinates.length),this.changed()},e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),this.layout,this.ends_.slice())},e.prototype.closestPointXY=function(t,e,i,n){return n<Nt(this.getExtent(),t,e)?n:(this.maxDeltaRevision_!=this.getRevision()&&(this.maxDelta_=Math.sqrt(ci(this.flatCoordinates,0,this.ends_,this.stride,0)),this.maxDeltaRevision_=this.getRevision()),fi(this.flatCoordinates,0,this.ends_,this.stride,this.maxDelta_,!1,t,e,i,n))},e.prototype.getCoordinateAtM=function(t,e,i){if(this.layout!=Ze&&this.layout!=Qe||0===this.flatCoordinates.length)return null;var n=void 0!==e&&e,r=void 0!==i&&i;return function(t,e,i,n,r,o,s){if(s)return yi(t,e,i[i.length-1],n,r,o);var a;if(r<t[n-1])return o?((a=t.slice(0,n))[n-1]=r,a):null;if(t[t.length-1]<r)return o?((a=t.slice(t.length-n))[n-1]=r,a):null;for(var h=0,u=i.length;h<u;++h){var l=i[h];if(e!=l){if(r<t[e+n-1])return null;if(r<=t[l-1])return yi(t,e,l,n,r,!1);e=l}}return null}(this.flatCoordinates,0,this.ends_,this.stride,t,n,r)},e.prototype.getCoordinates=function(){return _i(this.flatCoordinates,0,this.ends_,this.stride)},e.prototype.getEnds=function(){return this.ends_},e.prototype.getLineString=function(t){return t<0||this.ends_.length<=t?null:new Ri(this.flatCoordinates.slice(0===t?0:this.ends_[t-1],this.ends_[t]),this.layout)},e.prototype.getLineStrings=function(){for(var t=this.flatCoordinates,e=this.ends_,i=this.layout,n=[],r=0,o=0,s=e.length;o<s;++o){var a=e[o],h=new Ri(t.slice(r,a),i);n.push(h),r=a}return n},e.prototype.getFlatMidpoints=function(){for(var t=[],e=this.flatCoordinates,i=0,n=this.ends_,r=this.stride,o=0,s=n.length;o<s;++o){var a=n[o];hi(t,gi(e,i,a,r,.5)),i=a}return t},e.prototype.getSimplifiedGeometryInternal=function(t){var i=[],n=[];return i.length=ki(this.flatCoordinates,0,this.ends_,this.stride,t,i,0,n),new e(i,Je,n)},e.prototype.getType=function(){return fe},e.prototype.intersectsExtent=function(t){return function(t,e,i,n,r){for(var o=0,s=i.length;o<s;++o){if(wi(t,e,i[o],n,r))return!0;e=i[o]}return!1}(this.flatCoordinates,0,this.ends_,this.stride,t)},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,2),this.flatCoordinates||(this.flatCoordinates=[]);var i=ri(this.flatCoordinates,0,t,this.stride,this.ends_);this.flatCoordinates.length=0===i.length?0:i[i.length-1],this.changed()},e}(ei),Di=function(t){function e(e,i){t.call(this),this.setCoordinates(e,i)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),this.layout)},e.prototype.closestPointXY=function(t,e,i,n){var r=this.flatCoordinates,o=c(t,e,r[0],r[1]);if(o<n){for(var s=this.stride,a=0;a<s;++a)i[a]=r[a];return i.length=s,o}return n},e.prototype.getCoordinates=function(){return this.flatCoordinates?this.flatCoordinates.slice():[]},e.prototype.computeExtent=function(t){return Ut(this.flatCoordinates,t)},e.prototype.getType=function(){return he},e.prototype.intersectsExtent=function(t){return Bt(t,this.flatCoordinates[0],this.flatCoordinates[1])},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,0),this.flatCoordinates||(this.flatCoordinates=[]),this.flatCoordinates.length=ii(this.flatCoordinates,0,t,this.stride),this.changed()},e}(ei),Ai=function(t){function e(e,i){t.call(this),i&&!Array.isArray(e[0])?this.setFlatCoordinates(i,e):this.setCoordinates(e,i)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.appendPoint=function(t){this.flatCoordinates?hi(this.flatCoordinates,t.getFlatCoordinates()):this.flatCoordinates=t.getFlatCoordinates().slice(),this.changed()},e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),this.layout)},e.prototype.closestPointXY=function(t,e,i,n){if(n<Nt(this.getExtent(),t,e))return n;for(var r=this.flatCoordinates,o=this.stride,s=0,a=r.length;s<a;s+=o){var h=c(t,e,r[s],r[s+1]);if(h<n){n=h;for(var u=0;u<o;++u)i[u]=r[s+u];i.length=o}}return n},e.prototype.getCoordinates=function(){return di(this.flatCoordinates,0,this.flatCoordinates.length,this.stride)},e.prototype.getPoint=function(t){var e=this.flatCoordinates?this.flatCoordinates.length/this.stride:0;return t<0||e<=t?null:new Di(this.flatCoordinates.slice(t*this.stride,(t+1)*this.stride),this.layout)},e.prototype.getPoints=function(){for(var t=this.flatCoordinates,e=this.layout,i=this.stride,n=[],r=0,o=t.length;r<o;r+=i){var s=new Di(t.slice(r,r+i),e);n.push(s)}return n},e.prototype.getType=function(){return pe},e.prototype.intersectsExtent=function(t){for(var e=this.flatCoordinates,i=this.stride,n=0,r=e.length;n<r;n+=i){if(Bt(t,e[n],e[n+1]))return!0}return!1},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,1),this.flatCoordinates||(this.flatCoordinates=[]),this.flatCoordinates.length=ni(this.flatCoordinates,0,t,this.stride),this.changed()},e}(ei);function Oi(t,e,i,n){for(var r=0,o=t[i-n],s=t[i-n+1];e<i;e+=n){var a=t[e],h=t[e+1];r+=s*a-o*h,o=a,s=h}return r/2}function Gi(t,e,i,n){for(var r=0,o=0,s=i.length;o<s;++o){var a=i[o];r+=Oi(t,e,a,n),e=a}return r}var Ti=function(t){function e(e,i){t.call(this),this.maxDelta_=-1,this.maxDeltaRevision_=-1,void 0===i||Array.isArray(e[0])?this.setCoordinates(e,i):this.setFlatCoordinates(i,e)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),this.layout)},e.prototype.closestPointXY=function(t,e,i,n){return n<Nt(this.getExtent(),t,e)?n:(this.maxDeltaRevision_!=this.getRevision()&&(this.maxDelta_=Math.sqrt(li(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,0)),this.maxDeltaRevision_=this.getRevision()),pi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,this.maxDelta_,!0,t,e,i,n))},e.prototype.getArea=function(){return Oi(this.flatCoordinates,0,this.flatCoordinates.length,this.stride)},e.prototype.getCoordinates=function(){return di(this.flatCoordinates,0,this.flatCoordinates.length,this.stride)},e.prototype.getSimplifiedGeometryInternal=function(t){var i=[];return i.length=Si(this.flatCoordinates,0,this.flatCoordinates.length,this.stride,t,i,0),new e(i,Je)},e.prototype.getType=function(){return le},e.prototype.intersectsExtent=function(t){return!1},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,1),this.flatCoordinates||(this.flatCoordinates=[]),this.flatCoordinates.length=ni(this.flatCoordinates,0,t,this.stride),this.changed()},e}(ei);function Li(t,e,i,n,r,o,s){for(var a,h,u,l,c,p,f,d=r[o+1],_=[],g=0,y=i.length;g<y;++g){var v=i[g];for(l=t[v-n],p=t[v-n+1],a=e;a<v;a+=n)c=t[a],f=t[a+1],(d<=p&&f<=d||p<=d&&d<=f)&&(u=(d-p)/(f-p)*(c-l)+l,_.push(u)),l=c,p=f}var m=NaN,C=-1/0;for(_.sort(ai),l=_[0],a=1,h=_.length;a<h;++a){c=_[a];var x=Math.abs(c-l);x>C&&Ci(t,e,i,n,u=(l+c)/2,d)&&(m=u,C=x),l=c}return isNaN(m)&&(m=r[o]),s?(s.push(m,d,C),s):[m,d,C]}function Xi(t,e,i,n){for(;e<i-n;){for(var r=0;r<n;++r){var o=t[e+r];t[e+r]=t[i-n+r],t[i-n+r]=o}e+=n,i-=n}}function Yi(t,e,i,n){for(var r=0,o=t[i-n],s=t[i-n+1];e<i;e+=n){var a=t[e],h=t[e+1];r+=(a-o)*(h+s),o=a,s=h}return r>0}function ji(t,e,i,n,r){for(var o=void 0!==r&&r,s=0,a=i.length;s<a;++s){var h=i[s],u=Yi(t,e,h,n);if(0===s){if(o&&u||!o&&!u)return!1}else if(o&&!u||!o&&u)return!1;e=h}return!0}function Ni(t,e,i,n,r){for(var o=void 0!==r&&r,s=0,a=i.length;s<a;++s){var h=i[s],u=Yi(t,e,h,n);(0===s?o&&u||!o&&!u:o&&!u||!o&&u)&&Xi(t,e,h,n),e=h}return e}function qi(t,e,i,n,r){for(var o=0,s=i.length;o<s;++o)e=Ni(t,e,i[o],n,r);return e}var Bi=function(t){function e(e,i,n){t.call(this),this.ends_=[],this.flatInteriorPointRevision_=-1,this.flatInteriorPoint_=null,this.maxDelta_=-1,this.maxDeltaRevision_=-1,this.orientedRevision_=-1,this.orientedFlatCoordinates_=null,void 0!==i&&n?(this.setFlatCoordinates(i,e),this.ends_=n):this.setCoordinates(e,i)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.appendLinearRing=function(t){this.flatCoordinates?hi(this.flatCoordinates,t.getFlatCoordinates()):this.flatCoordinates=t.getFlatCoordinates().slice(),this.ends_.push(this.flatCoordinates.length),this.changed()},e.prototype.clone=function(){return new e(this.flatCoordinates.slice(),this.layout,this.ends_.slice())},e.prototype.closestPointXY=function(t,e,i,n){return n<Nt(this.getExtent(),t,e)?n:(this.maxDeltaRevision_!=this.getRevision()&&(this.maxDelta_=Math.sqrt(ci(this.flatCoordinates,0,this.ends_,this.stride,0)),this.maxDeltaRevision_=this.getRevision()),fi(this.flatCoordinates,0,this.ends_,this.stride,this.maxDelta_,!0,t,e,i,n))},e.prototype.containsXY=function(t,e){return Ci(this.getOrientedFlatCoordinates(),0,this.ends_,this.stride,t,e)},e.prototype.getArea=function(){return Gi(this.getOrientedFlatCoordinates(),0,this.ends_,this.stride)},e.prototype.getCoordinates=function(t){var e;return void 0!==t?Ni(e=this.getOrientedFlatCoordinates().slice(),0,this.ends_,this.stride,t):e=this.flatCoordinates,_i(e,0,this.ends_,this.stride)},e.prototype.getEnds=function(){return this.ends_},e.prototype.getFlatInteriorPoint=function(){if(this.flatInteriorPointRevision_!=this.getRevision()){var t=ne(this.getExtent());this.flatInteriorPoint_=Li(this.getOrientedFlatCoordinates(),0,this.ends_,this.stride,t,0),this.flatInteriorPointRevision_=this.getRevision()}return this.flatInteriorPoint_},e.prototype.getInteriorPoint=function(){return new Di(this.getFlatInteriorPoint(),Ze)},e.prototype.getLinearRingCount=function(){return this.ends_.length},e.prototype.getLinearRing=function(t){return t<0||this.ends_.length<=t?null:new Ti(this.flatCoordinates.slice(0===t?0:this.ends_[t-1],this.ends_[t]),this.layout)},e.prototype.getLinearRings=function(){for(var t=this.layout,e=this.flatCoordinates,i=this.ends_,n=[],r=0,o=0,s=i.length;o<s;++o){var a=i[o],h=new Ti(e.slice(r,a),t);n.push(h),r=a}return n},e.prototype.getOrientedFlatCoordinates=function(){if(this.orientedRevision_!=this.getRevision()){var t=this.flatCoordinates;ji(t,0,this.ends_,this.stride)?this.orientedFlatCoordinates_=t:(this.orientedFlatCoordinates_=t.slice(),this.orientedFlatCoordinates_.length=Ni(this.orientedFlatCoordinates_,0,this.ends_,this.stride)),this.orientedRevision_=this.getRevision()}return this.orientedFlatCoordinates_},e.prototype.getSimplifiedGeometryInternal=function(t){var i=[],n=[];return i.length=Pi(this.flatCoordinates,0,this.ends_,this.stride,Math.sqrt(t),i,0,n),new e(i,Je,n)},e.prototype.getType=function(){return ce},e.prototype.intersectsExtent=function(t){return Ei(this.getOrientedFlatCoordinates(),0,this.ends_,this.stride,t)},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,2),this.flatCoordinates||(this.flatCoordinates=[]);var i=ri(this.flatCoordinates,0,t,this.stride,this.ends_);this.flatCoordinates.length=0===i.length?0:i[i.length-1],this.changed()},e}(ei),Wi=Bi;var Vi=function(t){function e(e,i,n){if(t.call(this),this.endss_=[],this.flatInteriorPointsRevision_=-1,this.flatInteriorPoints_=null,this.maxDelta_=-1,this.maxDeltaRevision_=-1,this.orientedRevision_=-1,this.orientedFlatCoordinates_=null,!n&&!Array.isArray(e[0])){for(var r=this.getLayout(),o=e,s=[],a=[],h=0,u=o.length;h<u;++h){var l=o[h];0===h&&(r=l.getLayout());for(var c=s.length,p=l.getEnds(),f=0,d=p.length;f<d;++f)p[f]+=c;hi(s,l.getFlatCoordinates()),a.push(p)}i=r,e=s,n=a}void 0!==i&&n?(this.setFlatCoordinates(i,e),this.endss_=n):this.setCoordinates(e,i)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.appendPolygon=function(t){var e;if(this.flatCoordinates){var i=this.flatCoordinates.length;hi(this.flatCoordinates,t.getFlatCoordinates());for(var n=0,r=(e=t.getEnds().slice()).length;n<r;++n)e[n]+=i}else this.flatCoordinates=t.getFlatCoordinates().slice(),e=t.getEnds().slice(),this.endss_.push();this.endss_.push(e),this.changed()},e.prototype.clone=function(){for(var t=this.endss_.length,i=new Array(t),n=0;n<t;++n)i[n]=this.endss_[n].slice();return new e(this.flatCoordinates.slice(),this.layout,i)},e.prototype.closestPointXY=function(t,e,i,n){return n<Nt(this.getExtent(),t,e)?n:(this.maxDeltaRevision_!=this.getRevision()&&(this.maxDelta_=Math.sqrt(function(t,e,i,n,r){for(var o=0,s=i.length;o<s;++o){var a=i[o];r=ci(t,e,a,n,r),e=a[a.length-1]}return r}(this.flatCoordinates,0,this.endss_,this.stride,0)),this.maxDeltaRevision_=this.getRevision()),function(t,e,i,n,r,o,s,a,h,u,l){for(var c=l||[NaN,NaN],p=0,f=i.length;p<f;++p){var d=i[p];u=fi(t,e,d,n,r,o,s,a,h,u,c),e=d[d.length-1]}return u}(this.getOrientedFlatCoordinates(),0,this.endss_,this.stride,this.maxDelta_,!0,t,e,i,n))},e.prototype.containsXY=function(t,e){return function(t,e,i,n,r,o){if(0===i.length)return!1;for(var s=0,a=i.length;s<a;++s){var h=i[s];if(Ci(t,e,h,n,r,o))return!0;e=h[h.length-1]}return!1}(this.getOrientedFlatCoordinates(),0,this.endss_,this.stride,t,e)},e.prototype.getArea=function(){return function(t,e,i,n){for(var r=0,o=0,s=i.length;o<s;++o){var a=i[o];r+=Gi(t,e,a,n),e=a[a.length-1]}return r}(this.getOrientedFlatCoordinates(),0,this.endss_,this.stride)},e.prototype.getCoordinates=function(t){var e;return void 0!==t?qi(e=this.getOrientedFlatCoordinates().slice(),0,this.endss_,this.stride,t):e=this.flatCoordinates,function(t,e,i,n,r){for(var o=void 0!==r?r:[],s=0,a=0,h=i.length;a<h;++a){var u=i[a];o[s++]=_i(t,e,u,n,o[s]),e=u[u.length-1]}return o.length=s,o}(e,0,this.endss_,this.stride)},e.prototype.getEndss=function(){return this.endss_},e.prototype.getFlatInteriorPoints=function(){if(this.flatInteriorPointsRevision_!=this.getRevision()){var t=function(t,e,i,n){for(var r=[],o=[1/0,1/0,-1/0,-1/0],s=0,a=i.length;s<a;++s){var h=i[s];o=Ht(t,e,h[0],n),r.push((o[0]+o[2])/2,(o[1]+o[3])/2),e=h[h.length-1]}return r}(this.flatCoordinates,0,this.endss_,this.stride);this.flatInteriorPoints_=function(t,e,i,n,r){for(var o=[],s=0,a=i.length;s<a;++s){var h=i[s];o=Li(t,e,h,n,r,2*s,o),e=h[h.length-1]}return o}(this.getOrientedFlatCoordinates(),0,this.endss_,this.stride,t),this.flatInteriorPointsRevision_=this.getRevision()}return this.flatInteriorPoints_},e.prototype.getInteriorPoints=function(){return new Ai(this.getFlatInteriorPoints().slice(),Ze)},e.prototype.getOrientedFlatCoordinates=function(){if(this.orientedRevision_!=this.getRevision()){var t=this.flatCoordinates;!function(t,e,i,n,r){for(var o=0,s=i.length;o<s;++o)if(!ji(t,e,i[o],n,r))return!1;return!0}(t,0,this.endss_,this.stride)?(this.orientedFlatCoordinates_=t.slice(),this.orientedFlatCoordinates_.length=qi(this.orientedFlatCoordinates_,0,this.endss_,this.stride)):this.orientedFlatCoordinates_=t,this.orientedRevision_=this.getRevision()}return this.orientedFlatCoordinates_},e.prototype.getSimplifiedGeometryInternal=function(t){var i=[],n=[];return i.length=function(t,e,i,n,r,o,s,a){for(var h=0,u=i.length;h<u;++h){var l=i[h],c=[];s=Pi(t,e,l,n,r,o,s,c),a.push(c),e=l[l.length-1]}return s}(this.flatCoordinates,0,this.endss_,this.stride,Math.sqrt(t),i,0,n),new e(i,Je,n)},e.prototype.getPolygon=function(t){if(t<0||this.endss_.length<=t)return null;var e;if(0===t)e=0;else{var i=this.endss_[t-1];e=i[i.length-1]}var n=this.endss_[t].slice(),r=n[n.length-1];if(0!==e)for(var o=0,s=n.length;o<s;++o)n[o]-=e;return new Wi(this.flatCoordinates.slice(e,r),this.layout,n)},e.prototype.getPolygons=function(){for(var t=this.layout,e=this.flatCoordinates,i=this.endss_,n=[],r=0,o=0,s=i.length;o<s;++o){var a=i[o].slice(),h=a[a.length-1];if(0!==r)for(var u=0,l=a.length;u<l;++u)a[u]-=r;var c=new Wi(e.slice(r,h),t,a);n.push(c),r=h}return n},e.prototype.getType=function(){return de},e.prototype.intersectsExtent=function(t){return function(t,e,i,n,r){for(var o=0,s=i.length;o<s;++o){var a=i[o];if(Ei(t,e,a,n,r))return!0;e=a[a.length-1]}return!1}(this.getOrientedFlatCoordinates(),0,this.endss_,this.stride,t)},e.prototype.setCoordinates=function(t,e){this.setLayout(e,t,3),this.flatCoordinates||(this.flatCoordinates=[]);var i=function(t,e,i,n,r){for(var o=r||[],s=0,a=0,h=i.length;a<h;++a){var u=ri(t,e,i[a],n,o[s]);o[s++]=u,e=u[u.length-1]}return o.length=s,o}(this.flatCoordinates,0,t,this.stride,this.endss_);if(0===i.length)this.flatCoordinates.length=0;else{var n=i[i.length-1];this.flatCoordinates.length=0===n.length?0:n[n.length-1]}this.changed()},e}(ei),Ki=function(t,e){this.dispatcher=t,this.mapping_=e};Ki.prototype.getEvents=function(){return Object.keys(this.mapping_)},Ki.prototype.getHandlerForEvent=function(t){return this.mapping_[t]};var zi="active";var Ui=function(t){function e(e){var i=e||{};t.call(this,i),i.handleDownEvent&&(this.handleDownEvent=i.handleDownEvent),i.handleDragEvent&&(this.handleDragEvent=i.handleDragEvent),i.handleMoveEvent&&(this.handleMoveEvent=i.handleMoveEvent),i.handleUpEvent&&(this.handleUpEvent=i.handleUpEvent),i.stopDown&&(this.stopDown=i.stopDown),this.handlingDownUpSequence=!1,this.trackedPointers_={},this.targetPointers=[]}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.handleDownEvent=function(t){return!1},e.prototype.handleDragEvent=function(t){},e.prototype.handleEvent=function(t){if(!t.pointerEvent)return!0;var e=!1;if(this.updateTrackedPointers_(t),this.handlingDownUpSequence){if(t.type==Ct)this.handleDragEvent(t);else if(t.type==Et){var i=this.handleUpEvent(t);this.handlingDownUpSequence=i&&this.targetPointers.length>0}}else if(t.type==wt){var n=this.handleDownEvent(t);n&&t.preventDefault(),this.handlingDownUpSequence=n,e=this.stopDown(n)}else t.type==xt&&this.handleMoveEvent(t);return!e},e.prototype.handleMoveEvent=function(t){},e.prototype.handleUpEvent=function(t){return!1},e.prototype.stopDown=function(t){return t},e.prototype.updateTrackedPointers_=function(t){if(function(t){var e=t.type;return e===wt||e===Ct||e===Et}(t)){var e=t.pointerEvent,i=e.pointerId.toString();t.type==Et?delete this.trackedPointers_[i]:(t.type==wt||i in this.trackedPointers_)&&(this.trackedPointers_[i]=e),this.targetPointers=k(this.trackedPointers_)}},e}(function(t){function e(e){t.call(this),e.handleEvent&&(this.handleEvent=e.handleEvent),this.map_=null,this.setActive(!0)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getActive=function(){return this.get(zi)},e.prototype.getMap=function(){return this.map_},e.prototype.handleEvent=function(t){return!0},e.prototype.setActive=function(t){this.set(zi,t)},e.prototype.setMap=function(t){this.map_=t},e}(ft)),Hi="VECTOR",Ji="opacity",$i="visible",Zi="extent",Qi="zIndex",tn="maxResolution",en="minResolution",nn="source",rn="precompose",on="undefined",sn="ready";var an=function(t){function e(e){var i=F({},e);delete i.source,t.call(this,i),this.mapPrecomposeKey_=null,this.mapRenderKey_=null,this.sourceChangeKey_=null,e.map&&this.setMap(e.map),O(this,pt(nn),this.handleSourcePropertyChange_,this);var n=e.source?e.source:null;this.setSource(n)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getLayersArray=function(t){var e=t||[];return e.push(this),e},e.prototype.getLayerStatesArray=function(t){var e=t||[];return e.push(this.getLayerState()),e},e.prototype.getSource=function(){return this.get(nn)||null},e.prototype.getSourceState=function(){var t=this.getSource();return t?t.getState():on},e.prototype.handleSourceChange_=function(){this.changed()},e.prototype.handleSourcePropertyChange_=function(){this.sourceChangeKey_&&(L(this.sourceChangeKey_),this.sourceChangeKey_=null);var t=this.getSource();t&&(this.sourceChangeKey_=O(t,W,this.handleSourceChange_,this)),this.changed()},e.prototype.setMap=function(t){this.mapPrecomposeKey_&&(L(this.mapPrecomposeKey_),this.mapPrecomposeKey_=null),t||this.changed(),this.mapRenderKey_&&(L(this.mapRenderKey_),this.mapRenderKey_=null),t&&(this.mapPrecomposeKey_=O(t,rn,(function(t){var e=t,i=this.getLayerState();i.managed=!1,void 0===this.getZIndex()&&(i.zIndex=1/0),e.frameState.layerStatesArray.push(i),e.frameState.layerStates[o(this)]=i}),this),this.mapRenderKey_=O(this,W,t.render,t),this.changed())},e.prototype.setSource=function(t){this.set(nn,t)},e}(function(t){function e(e){t.call(this);var i=F({},e);i[Ji]=void 0!==e.opacity?e.opacity:1,i[$i]=void 0===e.visible||e.visible,i[Qi]=e.zIndex,i[tn]=void 0!==e.maxResolution?e.maxResolution:1/0,i[en]=void 0!==e.minResolution?e.minResolution:0,this.setProperties(i),this.state_=null,this.type}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getType=function(){return this.type},e.prototype.getLayerState=function(){var t=this.state_||{layer:this,managed:!0};return t.opacity=h(this.getOpacity(),0,1),t.sourceState=this.getSourceState(),t.visible=this.getVisible(),t.extent=this.getExtent(),t.zIndex=this.getZIndex()||0,t.maxResolution=this.getMaxResolution(),t.minResolution=Math.max(this.getMinResolution(),0),this.state_=t,t},e.prototype.getLayersArray=function(t){return n()},e.prototype.getLayerStatesArray=function(t){return n()},e.prototype.getExtent=function(){return this.get(Zi)},e.prototype.getMaxResolution=function(){return this.get(tn)},e.prototype.getMinResolution=function(){return this.get(en)},e.prototype.getOpacity=function(){return this.get(Ji)},e.prototype.getSourceState=function(){return n()},e.prototype.getVisible=function(){return this.get($i)},e.prototype.getZIndex=function(){return this.get(Qi)},e.prototype.setExtent=function(t){this.set(Zi,t)},e.prototype.setMaxResolution=function(t){this.set(tn,t)},e.prototype.setMinResolution=function(t){this.set(en,t)},e.prototype.setOpacity=function(t){this.set(Ji,t)},e.prototype.setVisible=function(t){this.set($i,t)},e.prototype.setZIndex=function(t){this.set(Qi,t)},e}(ft)),hn="vector",un=function(t){var e=t||{};this.color_=void 0!==e.color?e.color:null,this.checksum_=void 0};un.prototype.clone=function(){var t=this.getColor();return new un({color:Array.isArray(t)?t.slice():t||void 0})},un.prototype.getColor=function(){return this.color_},un.prototype.setColor=function(t){this.color_=t,this.checksum_=void 0},un.prototype.getChecksum=function(){if(void 0===this.checksum_){var t=this.color_;t?Array.isArray(t)||"string"==typeof t?this.checksum_="f"+_(t):this.checksum_=o(this.color_):this.checksum_="f-"}return this.checksum_};var ln=un,cn=function(t){var e=t||{};this.color_=void 0!==e.color?e.color:null,this.lineCap_=e.lineCap,this.lineDash_=void 0!==e.lineDash?e.lineDash:null,this.lineDashOffset_=e.lineDashOffset,this.lineJoin_=e.lineJoin,this.miterLimit_=e.miterLimit,this.width_=e.width,this.checksum_=void 0};cn.prototype.clone=function(){var t=this.getColor();return new cn({color:Array.isArray(t)?t.slice():t||void 0,lineCap:this.getLineCap(),lineDash:this.getLineDash()?this.getLineDash().slice():void 0,lineDashOffset:this.getLineDashOffset(),lineJoin:this.getLineJoin(),miterLimit:this.getMiterLimit(),width:this.getWidth()})},cn.prototype.getColor=function(){return this.color_},cn.prototype.getLineCap=function(){return this.lineCap_},cn.prototype.getLineDash=function(){return this.lineDash_},cn.prototype.getLineDashOffset=function(){return this.lineDashOffset_},cn.prototype.getLineJoin=function(){return this.lineJoin_},cn.prototype.getMiterLimit=function(){return this.miterLimit_},cn.prototype.getWidth=function(){return this.width_},cn.prototype.setColor=function(t){this.color_=t,this.checksum_=void 0},cn.prototype.setLineCap=function(t){this.lineCap_=t,this.checksum_=void 0},cn.prototype.setLineDash=function(t){this.lineDash_=t,this.checksum_=void 0},cn.prototype.setLineDashOffset=function(t){this.lineDashOffset_=t,this.checksum_=void 0},cn.prototype.setLineJoin=function(t){this.lineJoin_=t,this.checksum_=void 0},cn.prototype.setMiterLimit=function(t){this.miterLimit_=t,this.checksum_=void 0},cn.prototype.setWidth=function(t){this.width_=t,this.checksum_=void 0},cn.prototype.getChecksum=function(){return void 0===this.checksum_&&(this.checksum_="s",this.color_?"string"==typeof this.color_?this.checksum_+=this.color_:this.checksum_+=o(this.color_):this.checksum_+="-",this.checksum_+=","+(void 0!==this.lineCap_?this.lineCap_.toString():"-")+","+(this.lineDash_?this.lineDash_.toString():"-")+","+(void 0!==this.lineDashOffset_?this.lineDashOffset_:"-")+","+(void 0!==this.lineJoin_?this.lineJoin_:"-")+","+(void 0!==this.miterLimit_?this.miterLimit_.toString():"-")+","+(void 0!==this.width_?this.width_.toString():"-")),this.checksum_};var pn=cn,fn=function(t){var e=t||{};this.geometry_=null,this.geometryFunction_=yn,void 0!==e.geometry&&this.setGeometry(e.geometry),this.fill_=void 0!==e.fill?e.fill:null,this.image_=void 0!==e.image?e.image:null,this.renderer_=void 0!==e.renderer?e.renderer:null,this.stroke_=void 0!==e.stroke?e.stroke:null,this.text_=void 0!==e.text?e.text:null,this.zIndex_=e.zIndex};fn.prototype.clone=function(){var t=this.getGeometry();return t&&"object"==typeof t&&(t=t.clone()),new fn({geometry:t,fill:this.getFill()?this.getFill().clone():void 0,image:this.getImage()?this.getImage().clone():void 0,stroke:this.getStroke()?this.getStroke().clone():void 0,text:this.getText()?this.getText().clone():void 0,zIndex:this.getZIndex()})},fn.prototype.getRenderer=function(){return this.renderer_},fn.prototype.setRenderer=function(t){this.renderer_=t},fn.prototype.getGeometry=function(){return this.geometry_},fn.prototype.getGeometryFunction=function(){return this.geometryFunction_},fn.prototype.getFill=function(){return this.fill_},fn.prototype.setFill=function(t){this.fill_=t},fn.prototype.getImage=function(){return this.image_},fn.prototype.setImage=function(t){this.image_=t},fn.prototype.getStroke=function(){return this.stroke_},fn.prototype.setStroke=function(t){this.stroke_=t},fn.prototype.getText=function(){return this.text_},fn.prototype.setText=function(t){this.text_=t},fn.prototype.getZIndex=function(){return this.zIndex_},fn.prototype.setGeometry=function(t){"function"==typeof t?this.geometryFunction_=t:"string"==typeof t?this.geometryFunction_=function(e){return e.get(t)}:t?void 0!==t&&(this.geometryFunction_=function(){return t}):this.geometryFunction_=yn,this.geometry_=t},fn.prototype.setZIndex=function(t){this.zIndex_=t};var dn=null;function _n(t,e){if(!dn){var i=new ln({color:"rgba(255,255,255,0.4)"}),n=new pn({color:"#3399CC",width:1.25});dn=[new fn({image:new ot({fill:i,stroke:n,radius:5}),fill:i,stroke:n})]}return dn}function gn(){var t={},e=[255,255,255,1],i=[0,153,255,1];return t[ce]=[new fn({fill:new ln({color:[255,255,255,.5]})})],t[de]=t[ce],t[ue]=[new fn({stroke:new pn({color:e,width:5})}),new fn({stroke:new pn({color:i,width:3})})],t[fe]=t[ue],t[ge]=t[ce].concat(t[ue]),t[he]=[new fn({image:new ot({radius:6,fill:new ln({color:i}),stroke:new pn({color:e,width:1.5})}),zIndex:1/0})],t[pe]=t[he],t[_e]=t[ce].concat(t[ue],t[he]),t}function yn(t){return t.getGeometry()}var vn=fn,mn="renderOrder",Cn=function(t){function e(e){var i=e||{},n=F({},i);delete n.style,delete n.renderBuffer,delete n.updateWhileAnimating,delete n.updateWhileInteracting,t.call(this,n),this.declutter_=void 0!==i.declutter&&i.declutter,this.renderBuffer_=void 0!==i.renderBuffer?i.renderBuffer:100,this.style_=null,this.styleFunction_=void 0,this.setStyle(i.style),this.updateWhileAnimating_=void 0!==i.updateWhileAnimating&&i.updateWhileAnimating,this.updateWhileInteracting_=void 0!==i.updateWhileInteracting&&i.updateWhileInteracting,this.renderMode_=i.renderMode||hn,this.type=Hi}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getDeclutter=function(){return this.declutter_},e.prototype.setDeclutter=function(t){this.declutter_=t},e.prototype.getRenderBuffer=function(){return this.renderBuffer_},e.prototype.getRenderOrder=function(){return this.get(mn)},e.prototype.getStyle=function(){return this.style_},e.prototype.getStyleFunction=function(){return this.styleFunction_},e.prototype.getUpdateWhileAnimating=function(){return this.updateWhileAnimating_},e.prototype.getUpdateWhileInteracting=function(){return this.updateWhileInteracting_},e.prototype.setRenderOrder=function(t){this.set(mn,t)},e.prototype.setStyle=function(t){this.style_=void 0!==t?t:_n,this.styleFunction_=null===t?void 0:function(t){var e;if("function"==typeof t)e=t;else{var i;if(Array.isArray(t))i=t;else a("function"==typeof t.getZIndex,41),i=[t];e=function(){return i}}return e}(this.style_),this.changed()},e.prototype.getRenderMode=function(){return this.renderMode_},e}(an);Cn.prototype.getSource;var xn=Cn,wn="arraybuffer",En="json",Fn="text",Sn="xml";function kn(t,e){return function(t,e,i,n){return function(r,o,s){var a=new XMLHttpRequest;a.open("GET","function"==typeof t?t(r,o,s):t,!0),e.getType()==wn&&(a.responseType="arraybuffer"),a.onload=function(t){if(!a.status||a.status>=200&&a.status<300){var r,o=e.getType();o==En||o==Fn?r=a.responseText:o==Sn?(r=a.responseXML)||(r=(new DOMParser).parseFromString(a.responseText,"application/xml")):o==wn&&(r=a.response),r?i.call(this,e.readFeatures(r,{featureProjection:s}),e.readProjection(r),e.getLastExtent()):n.call(this)}else n.call(this)}.bind(this),a.onerror=function(){n.call(this)}.bind(this),a.send()}}(t,e,(function(t,e){"function"==typeof this.addFeatures&&this.addFeatures(t)}),j)}function Mn(t,e){return[[-1/0,-1/0,1/0,1/0]]}function In(t){return t?Array.isArray(t)?function(e){return t}:"function"==typeof t?t:function(e){return[t]}:null}var Pn=function(t){function e(e){t.call(this),this.projection_=Ve(e.projection),this.attributions_=In(e.attributions),this.attributionsCollapsible_=void 0===e.attributionsCollapsible||e.attributionsCollapsible,this.loading=!1,this.state_=void 0!==e.state?e.state:sn,this.wrapX_=void 0!==e.wrapX&&e.wrapX}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getAttributions=function(){return this.attributions_},e.prototype.getAttributionsCollapsible=function(){return this.attributionsCollapsible_},e.prototype.getProjection=function(){return this.projection_},e.prototype.getResolutions=function(){return n()},e.prototype.getState=function(){return this.state_},e.prototype.getWrapX=function(){return this.wrapX_},e.prototype.refresh=function(){this.changed()},e.prototype.setAttributions=function(t){this.attributions_=In(t),this.changed()},e.prototype.setState=function(t){this.state_=t,this.changed()},e}(ft),Rn="addfeature",bn="changefeature",Dn="clear",An="removefeature",On=i("A1R+"),Gn=i.n(On),Tn=function(t){this.rbush_=Gn()(t,void 0),this.items_={}};Tn.prototype.insert=function(t,e){var i={minX:t[0],minY:t[1],maxX:t[2],maxY:t[3],value:e};this.rbush_.insert(i),this.items_[o(e)]=i},Tn.prototype.load=function(t,e){for(var i=new Array(e.length),n=0,r=e.length;n<r;n++){var s=t[n],a=e[n],h={minX:s[0],minY:s[1],maxX:s[2],maxY:s[3],value:a};i[n]=h,this.items_[o(a)]=h}this.rbush_.load(i)},Tn.prototype.remove=function(t){var e=o(t),i=this.items_[e];return delete this.items_[e],null!==this.rbush_.remove(i)},Tn.prototype.update=function(t,e){var i=this.items_[o(e)];Jt([i.minX,i.minY,i.maxX,i.maxY],t)||(this.remove(e),this.insert(t,e))},Tn.prototype.getAll=function(){return this.rbush_.all().map((function(t){return t.value}))},Tn.prototype.getInExtent=function(t){var e={minX:t[0],minY:t[1],maxX:t[2],maxY:t[3]};return this.rbush_.search(e).map((function(t){return t.value}))},Tn.prototype.forEach=function(t,e){return this.forEach_(this.getAll(),t,e)},Tn.prototype.forEachInExtent=function(t,e,i){return this.forEach_(this.getInExtent(t),e,i)},Tn.prototype.forEach_=function(t,e,i){for(var n,r=0,o=t.length;r<o;r++)if(n=e.call(i,t[r]))return n;return n},Tn.prototype.isEmpty=function(){return M(this.items_)},Tn.prototype.clear=function(){this.rbush_.clear(),this.items_={}},Tn.prototype.getExtent=function(t){var e=this.rbush_.toJSON();return Kt(e.minX,e.minY,e.maxX,e.maxY,t)},Tn.prototype.concat=function(t){for(var e in this.rbush_.load(t.rbush_.all()),t.items_)this.items_[e]=t.items_[e]};var Ln=Tn,Xn=function(t){function e(e,i){t.call(this,e),this.feature=i}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(q),Yn=function(t){function e(e){var i=e||{};t.call(this,{attributions:i.attributions,projection:void 0,state:sn,wrapX:void 0===i.wrapX||i.wrapX}),this.loader_=j,this.format_=i.format,this.overlaps_=null==i.overlaps||i.overlaps,this.url_=i.url,void 0!==i.loader?this.loader_=i.loader:void 0!==this.url_&&(a(this.format_,7),this.loader_=kn(this.url_,this.format_)),this.strategy_=void 0!==i.strategy?i.strategy:Mn;var n,r,o=void 0===i.useSpatialIndex||i.useSpatialIndex;this.featuresRtree_=o?new Ln:null,this.loadedExtentsRtree_=new Ln,this.nullGeometryFeatures_={},this.idIndex_={},this.undefIdIndex_={},this.featureChangeKeys_={},this.featuresCollection_=null,Array.isArray(i.features)?r=i.features:i.features&&(r=(n=i.features).getArray()),o||void 0!==n||(n=new gt(r)),void 0!==r&&this.addFeaturesInternal(r),void 0!==n&&this.bindFeaturesCollection_(n)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.addFeature=function(t){this.addFeatureInternal(t),this.changed()},e.prototype.addFeatureInternal=function(t){var e=o(t);if(this.addToIndex_(e,t)){this.setupChangeEvents_(e,t);var i=t.getGeometry();if(i){var n=i.getExtent();this.featuresRtree_&&this.featuresRtree_.insert(n,t)}else this.nullGeometryFeatures_[e]=t;this.dispatchEvent(new Xn(Rn,t))}},e.prototype.setupChangeEvents_=function(t,e){this.featureChangeKeys_[t]=[O(e,W,this.handleFeatureChange_,this),O(e,ht,this.handleFeatureChange_,this)]},e.prototype.addToIndex_=function(t,e){var i=!0,n=e.getId();return void 0!==n?n.toString()in this.idIndex_?i=!1:this.idIndex_[n.toString()]=e:(a(!(t in this.undefIdIndex_),30),this.undefIdIndex_[t]=e),i},e.prototype.addFeatures=function(t){this.addFeaturesInternal(t),this.changed()},e.prototype.addFeaturesInternal=function(t){for(var e=[],i=[],n=[],r=0,s=t.length;r<s;r++){var a=t[r],h=o(a);this.addToIndex_(h,a)&&i.push(a)}for(var u=0,l=i.length;u<l;u++){var c=i[u],p=o(c);this.setupChangeEvents_(p,c);var f=c.getGeometry();if(f){var d=f.getExtent();e.push(d),n.push(c)}else this.nullGeometryFeatures_[p]=c}this.featuresRtree_&&this.featuresRtree_.load(e,n);for(var _=0,g=i.length;_<g;_++)this.dispatchEvent(new Xn(Rn,i[_]))},e.prototype.bindFeaturesCollection_=function(t){var e=!1;O(this,Rn,(function(i){e||(e=!0,t.push(i.feature),e=!1)})),O(this,An,(function(i){e||(e=!0,t.remove(i.feature),e=!1)})),O(t,st,(function(t){e||(e=!0,this.addFeature(t.element),e=!1)}),this),O(t,at,(function(t){e||(e=!0,this.removeFeature(t.element),e=!1)}),this),this.featuresCollection_=t},e.prototype.clear=function(t){if(t){for(var e in this.featureChangeKeys_){this.featureChangeKeys_[e].forEach(L)}this.featuresCollection_||(this.featureChangeKeys_={},this.idIndex_={},this.undefIdIndex_={})}else if(this.featuresRtree_)for(var i in this.featuresRtree_.forEach(this.removeFeatureInternal,this),this.nullGeometryFeatures_)this.removeFeatureInternal(this.nullGeometryFeatures_[i]);this.featuresCollection_&&this.featuresCollection_.clear(),this.featuresRtree_&&this.featuresRtree_.clear(),this.loadedExtentsRtree_.clear(),this.nullGeometryFeatures_={};var n=new Xn(Dn);this.dispatchEvent(n),this.changed()},e.prototype.forEachFeature=function(t){if(this.featuresRtree_)return this.featuresRtree_.forEach(t);this.featuresCollection_&&this.featuresCollection_.forEach(t)},e.prototype.forEachFeatureAtCoordinateDirect=function(t,e){var i=[t[0],t[1],t[0],t[1]];return this.forEachFeatureInExtent(i,(function(i){return i.getGeometry().intersectsCoordinate(t)?e(i):void 0}))},e.prototype.forEachFeatureInExtent=function(t,e){if(this.featuresRtree_)return this.featuresRtree_.forEachInExtent(t,e);this.featuresCollection_&&this.featuresCollection_.forEach(e)},e.prototype.forEachFeatureIntersectingExtent=function(t,e){return this.forEachFeatureInExtent(t,(function(i){if(i.getGeometry().intersectsExtent(t)){var n=e(i);if(n)return n}}))},e.prototype.getFeaturesCollection=function(){return this.featuresCollection_},e.prototype.getFeatures=function(){var t;return this.featuresCollection_?t=this.featuresCollection_.getArray():this.featuresRtree_&&(t=this.featuresRtree_.getAll(),M(this.nullGeometryFeatures_)||hi(t,k(this.nullGeometryFeatures_))),t},e.prototype.getFeaturesAtCoordinate=function(t){var e=[];return this.forEachFeatureAtCoordinateDirect(t,(function(t){e.push(t)})),e},e.prototype.getFeaturesInExtent=function(t){return this.featuresRtree_.getInExtent(t)},e.prototype.getClosestFeatureToCoordinate=function(t,e){var i=t[0],n=t[1],r=null,o=[NaN,NaN],s=1/0,a=[-1/0,-1/0,1/0,1/0],h=e||X;return this.featuresRtree_.forEachInExtent(a,(function(t){if(h(t)){var e=t.getGeometry(),u=s;if((s=e.closestPointXY(i,n,o,s))<u){r=t;var l=Math.sqrt(s);a[0]=i-l,a[1]=n-l,a[2]=i+l,a[3]=n+l}}})),r},e.prototype.getExtent=function(t){return this.featuresRtree_.getExtent(t)},e.prototype.getFeatureById=function(t){var e=this.idIndex_[t.toString()];return void 0!==e?e:null},e.prototype.getFormat=function(){return this.format_},e.prototype.getOverlaps=function(){return this.overlaps_},e.prototype.getUrl=function(){return this.url_},e.prototype.handleFeatureChange_=function(t){var e=t.target,i=o(e),n=e.getGeometry();if(n){var r=n.getExtent();i in this.nullGeometryFeatures_?(delete this.nullGeometryFeatures_[i],this.featuresRtree_&&this.featuresRtree_.insert(r,e)):this.featuresRtree_&&this.featuresRtree_.update(r,e)}else i in this.nullGeometryFeatures_||(this.featuresRtree_&&this.featuresRtree_.remove(e),this.nullGeometryFeatures_[i]=e);var s=e.getId();if(void 0!==s){var a=s.toString();i in this.undefIdIndex_?(delete this.undefIdIndex_[i],this.idIndex_[a]=e):this.idIndex_[a]!==e&&(this.removeFromIdIndex_(e),this.idIndex_[a]=e)}else i in this.undefIdIndex_||(this.removeFromIdIndex_(e),this.undefIdIndex_[i]=e);this.changed(),this.dispatchEvent(new Xn(bn,e))},e.prototype.hasFeature=function(t){var e=t.getId();return void 0!==e?e in this.idIndex_:o(t)in this.undefIdIndex_},e.prototype.isEmpty=function(){return this.featuresRtree_.isEmpty()&&M(this.nullGeometryFeatures_)},e.prototype.loadFeatures=function(t,e,i){var n=this,r=this.loadedExtentsRtree_,o=this.strategy_(t,e);this.loading=!1;for(var s=function(t,s){var a=o[t];r.forEachInExtent(a,(function(t){return qt(t.extent,a)}))||(n.loader_.call(n,a,e,i),r.insert(a,{extent:a.slice()}),n.loading=n.loader_!==j)},a=0,h=o.length;a<h;++a)s(a)},e.prototype.removeLoadedExtent=function(t){var e,i=this.loadedExtentsRtree_;i.forEachInExtent(t,(function(i){if(Jt(i.extent,t))return e=i,!0})),e&&i.remove(e)},e.prototype.removeFeature=function(t){var e=o(t);e in this.nullGeometryFeatures_?delete this.nullGeometryFeatures_[e]:this.featuresRtree_&&this.featuresRtree_.remove(t),this.removeFeatureInternal(t),this.changed()},e.prototype.removeFeatureInternal=function(t){var e=o(t);this.featureChangeKeys_[e].forEach(L),delete this.featureChangeKeys_[e];var i=t.getId();void 0!==i?delete this.idIndex_[i.toString()]:delete this.undefIdIndex_[e],this.dispatchEvent(new Xn(An,t))},e.prototype.removeFromIdIndex_=function(t){var e=!1;for(var i in this.idIndex_)if(this.idIndex_[i]===t){delete this.idIndex_[i],e=!0;break}return e},e.prototype.setLoader=function(t){this.loader_=t},e}(Pn),jn=function(t){function e(){t.apply(this,arguments)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getType=function(){return ye},e.prototype.closestPointXY=function(t,e,i,n){var r,o,s,a=this.flatCoordinates,h=n;i[0]=a[0],i[1]=a[1];for(var u=0,l=a.length;u<l;u+=2)(r=(o=t-a[u])*o+(s=e-a[u+1])*s)<h&&(h=r,i[0]=a[u],i[1]=a[u+1]);return h},e}(Wi),Nn=function(t){function e(){t.apply(this,arguments)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.getType=function(){return ve},e.prototype.closestPointXY=function(t,e,i,n){var r,o,s,a=this.flatCoordinates,h=n;i[0]=a[0],i[1]=a[1];for(var u=0,l=a.length;u<l;u+=2)(r=(o=t-a[u])*o+(s=e-a[u+1])*s)<h&&(h=r,i[0]=a[u],i[1]=a[u+1]);return h},e.prototype.getArea=function(){var t=this.flatCoordinates,e=Math.sqrt(Math.pow(t[0]-t[4],2)+Math.pow(t[1]-t[5],2)),i=Math.sqrt(Math.pow(t[2]-t[6],2)+Math.pow(t[3]-t[7],2));return Math.PI*e*i*.25},e}(Wi),qn="Point",Bn="LineString",Wn="Polygon",Vn="Circle",Kn="Ellipse",zn="drawstart",Un="drawend",Hn=function(t){function e(e,i){t.call(this,e),this.feature=i}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(q);var Jn=function(t){function e(e){var i=e;i.stopDown||(i.stopDown=Y),t.call(this,i),this.shouldHandle_=!1,this.downPx_=null,this.downTimeout_,this.lastDragTime_,this.freehand_=!1,this.source_=e.source?e.source:null,this.features_=e.features?e.features:null,this.snapTolerance_=e.snapTolerance?e.snapTolerance:12,this.type_=e.type,this.mode_=function(t){var e;t===he||t===pe?e=qn:t===ue||t===fe||t===ye?e=Bn:t===ce||t===de?e=Wn:t===ge?e=Vn:t===ve&&(e=Kn);return e}(this.type_),this.stopClick_=!!e.stopClick,this.minPoints_=e.minPoints?e.minPoints:this.mode_===Wn?3:2,this.maxPoints_=e.maxPoints?e.maxPoints:1/0,this.finishCondition_=e.finishCondition?e.finishCondition:X;var n,r=e.geometryFunction;if(!r)if(this.type_===ge)r=function(t,e){var i=e||new si([NaN,NaN]),n=Mt(t[0],t[1]);return i.setCenterAndRadius(t[0],Math.sqrt(n)),i};else if(this.type_===ye)this.minPoints_=3,this.maxPoints_=3,r=function(t,e){if(t.length>2){var i=t[0],n=t[1],r=t[2],o=[n[0]-i[0],n[1]-i[1]];if(0===o[1])t=[[i,i,i,i]];else{var s=[-1*o[1],o[0]],a=o[0]/o[1],h=(r[0]+a*(n[1]-r[1])-n[0])/(s[0]-s[1]*a),u=[h*s[0],h*s[1]];t=[[[i[0]-u[0],i[1]-u[1]],[n[0]-u[0],n[1]-u[1]],[n[0]+u[0],n[1]+u[1]],[i[0]+u[0],i[1]+u[1]]]]}}else t=[t];var l=e;return l?l.setCoordinates(t):l=new jn(t),l};else if(this.type_===ve)this.minPoints_=3,this.maxPoints_=3,r=function(t,e){if(t.length>2){var i=t[0],n=t[1],r=t[2],o=[n[0]-i[0],n[1]-i[1]],s=[i[0]+.5*o[0],i[1]+.5*o[1]];if(0===o[1])t=[[i,i,i,i]];else{var a=[-1*o[1],o[0]],h=o[0]/o[1],u=(r[0]+h*(n[1]-r[1])-n[0])/(a[0]-a[1]*h),l=[u*a[0],u*a[1]];t=[[[i[0],i[1]],[s[0]-l[0],s[1]-l[1]],[n[0],n[1]],[s[0]+l[0],s[1]+l[1]]]]}}else t=[t];var c=e;return c?c.setCoordinates(t):c=new Nn(t),c};else{var o,s=this.mode_;s===qn?o=Di:s===Bn?o=Ri:s===Wn&&(o=Wi),r=function(t,e){var i=e;return i?s===Wn?t[0].length?i.setCoordinates([t[0].concat([t[0][0]])]):i.setCoordinates([]):i.setCoordinates(t):i=new o(t),i}}this.geometryFunction_=r,this.dragVertexDelay_=void 0!==e.dragVertexDelay?e.dragVertexDelay:500,this.finishCoordinate_=null,this.sketchFeature_=null,this.sketchPoint_=null,this.sketchCoords_=null,this.sketchLine_=null,this.sketchLineCoords_=null,this.squaredClickTolerance_=e.clickTolerance?e.clickTolerance*e.clickTolerance:36,this.overlay_=new xn({source:new Yn({useSpatialIndex:!1,wrapX:!!e.wrapX&&e.wrapX}),style:e.style?e.style:(n=gn(),function(t,e){return n[t.getGeometry().getType()]}),updateWhileInteracting:!0}),this.geometryName_=e.geometryName,this.condition_=e.condition?e.condition:bt,this.freehandCondition_,e.freehand?this.freehandCondition_=Pt:this.freehandCondition_=e.freehandCondition?e.freehandCondition:Dt,O(this,pt(zi),this.updateState_,this)}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.setMap=function(e){t.prototype.setMap.call(this,e),this.updateState_()},e.prototype.getOverlay=function(){return this.overlay_},e.prototype.handleEvent=function(e){e.originalEvent.type===K&&e.preventDefault(),this.freehand_=this.mode_!==qn&&this.freehandCondition_(e);var i=e.type===xt,n=!0;!this.freehand_&&this.lastDragTime_&&e.type===Ct&&(Date.now()-this.lastDragTime_>=this.dragVertexDelay_?(this.downPx_=e.pixel,this.shouldHandle_=!this.freehand_,i=!0):this.lastDragTime_=void 0,this.shouldHandle_&&void 0!==this.downTimeout_&&(clearTimeout(this.downTimeout_),this.downTimeout_=void 0));return this.freehand_&&e.type===Ct&&null!==this.sketchFeature_?(this.addToDrawing_(e),n=!1):this.freehand_&&e.type===wt?n=!1:i?(n=e.type===xt)&&this.freehand_?n=this.handlePointerMove_(e):("mouse"==e.pointerEvent.pointerType||e.type===Ct&&void 0===this.downTimeout_)&&this.handlePointerMove_(e):e.type===mt&&(n=!1),t.prototype.handleEvent.call(this,e)&&n},e.prototype.handleDownEvent=function(t){return this.shouldHandle_=!this.freehand_,this.freehand_?(this.downPx_=t.pixel,this.finishCoordinate_||this.startDrawing_(t),!0):!!this.condition_(t)&&(this.lastDragTime_=Date.now(),this.downTimeout_=setTimeout(function(){this.handlePointerMove_(new Ft(xt,t.map,t.pointerEvent,!1,t.frameState))}.bind(this),this.dragVertexDelay_),this.downPx_=t.pixel,!0)},e.prototype.handleUpEvent=function(t){var e=!0;this.downTimeout_&&(clearTimeout(this.downTimeout_),this.downTimeout_=void 0),this.handlePointerMove_(t);var i=this.mode_===Vn;return this.shouldHandle_?(this.finishCoordinate_?this.freehand_||i?this.finishDrawing():this.atFinish_(t)?this.finishCondition_(t)&&this.finishDrawing():this.addToDrawing_(t):(this.startDrawing_(t),this.mode_===qn&&this.finishDrawing()),e=!1):this.freehand_&&(this.finishCoordinate_=null,this.abortDrawing_()),!e&&this.stopClick_&&t.stopPropagation(),e},e.prototype.handlePointerMove_=function(t){if(this.downPx_&&(!this.freehand_&&this.shouldHandle_||this.freehand_&&!this.shouldHandle_)){var e=this.downPx_,i=t.pixel,n=e[0]-i[0],r=e[1]-i[1],o=n*n+r*r;if(this.shouldHandle_=this.freehand_?o>this.squaredClickTolerance_:o<=this.squaredClickTolerance_,!this.shouldHandle_)return!0}return this.finishCoordinate_?this.modifyDrawing_(t):this.createOrUpdateSketchPoint_(t),!0},e.prototype.atFinish_=function(t){var e=!1;if(this.sketchFeature_){var i=!1,n=[this.finishCoordinate_];if(this.mode_===Bn)i=this.sketchCoords_.length>this.minPoints_;else if(this.mode_===Wn){var r=this.sketchCoords_;i=r[0].length>this.minPoints_,n=[r[0][0],r[0][r[0].length-2]]}if(i)for(var o=t.map,s=0,a=n.length;s<a;s++){var h=n[s],u=o.getPixelFromCoordinate(h),l=t.pixel,c=l[0]-u[0],p=l[1]-u[1],f=this.freehand_?1:this.snapTolerance_;if(e=Math.sqrt(c*c+p*p)<=f){this.finishCoordinate_=h;break}}}return e},e.prototype.createOrUpdateSketchPoint_=function(t){var e=t.coordinate.slice();this.sketchPoint_?this.sketchPoint_.getGeometry().setCoordinates(e):(this.sketchPoint_=new yt(new Di(e)),this.updateSketchFeatures_())},e.prototype.startDrawing_=function(t){var e=t.coordinate;this.finishCoordinate_=e,this.mode_===qn?this.sketchCoords_=e.slice():this.mode_===Wn?(this.sketchCoords_=[[e.slice(),e.slice()]],this.sketchLineCoords_=this.sketchCoords_[0]):this.sketchCoords_=[e.slice(),e.slice()],this.sketchLineCoords_&&(this.sketchLine_=new yt(new Ri(this.sketchLineCoords_)));var i=this.geometryFunction_(this.sketchCoords_);this.sketchFeature_=new yt,this.geometryName_&&this.sketchFeature_.setGeometryName(this.geometryName_),this.sketchFeature_.setGeometry(i),this.updateSketchFeatures_(),this.dispatchEvent(new Hn(zn,this.sketchFeature_))},e.prototype.modifyDrawing_=function(t){var e,i,n,r=t.coordinate,o=this.sketchFeature_.getGeometry();(this.mode_===qn?i=this.sketchCoords_:this.mode_===Wn?(i=(e=this.sketchCoords_[0])[e.length-1],this.atFinish_(t)&&(r=this.finishCoordinate_.slice())):i=(e=this.sketchCoords_)[e.length-1],i[0]=r[0],i[1]=r[1],this.geometryFunction_(this.sketchCoords_,o),this.sketchPoint_)&&this.sketchPoint_.getGeometry().setCoordinates(r);if(this.mode_===Kn)if(this.sketchLine_||(this.sketchLine_=new yt(new Ri([0,0]))),n=this.sketchLine_.getGeometry(),this.sketchCoords_.length<3)n.setCoordinates(this.sketchCoords_);else{var s=[((e=o.getCoordinates()[0])[0][0]+e[2][0])/2,(e[0][1]+e[2][1])/2];n.setCoordinates([e[0],s,e[3]])}else if(o.getType()==ce&&this.mode_!==Wn){this.sketchLine_||(this.sketchLine_=new yt);var a=o.getLinearRing(0);(n=this.sketchLine_.getGeometry())?(n.setFlatCoordinates(a.getLayout(),a.getFlatCoordinates()),n.changed()):(n=new Ri(a.getFlatCoordinates(),a.getLayout()),this.sketchLine_.setGeometry(n))}else this.sketchLineCoords_&&(n=this.sketchLine_.getGeometry()).setCoordinates(this.sketchLineCoords_);this.updateSketchFeatures_()},e.prototype.addToDrawing_=function(t){var e,i,n=t.coordinate,r=this.sketchFeature_.getGeometry();this.mode_===Bn||this.mode_===Kn?(this.finishCoordinate_=n.slice(),(i=this.sketchCoords_).length>=this.maxPoints_&&(this.freehand_?i.pop():e=!0),i.push(n.slice()),this.geometryFunction_(i,r)):this.mode_===Wn&&((i=this.sketchCoords_[0]).length>=this.maxPoints_&&(this.freehand_?i.pop():e=!0),i.push(n.slice()),e&&(this.finishCoordinate_=i[0]),this.geometryFunction_(this.sketchCoords_,r)),this.updateSketchFeatures_(),e&&this.finishDrawing()},e.prototype.removeLastPoint=function(){if(this.sketchFeature_){var t,e=this.sketchFeature_.getGeometry();this.mode_===Bn?((t=this.sketchCoords_).splice(-2,1),this.geometryFunction_(t,e),t.length>=2&&(this.finishCoordinate_=t[t.length-2].slice())):this.mode_===Wn&&((t=this.sketchCoords_[0]).splice(-2,1),this.sketchLine_.getGeometry().setCoordinates(t),this.geometryFunction_(this.sketchCoords_,e)),0===t.length&&(this.finishCoordinate_=null),this.updateSketchFeatures_()}},e.prototype.finishDrawing=function(){var t=this.abortDrawing_();if(t){var e=this.sketchCoords_,i=t.getGeometry();this.mode_===Bn?(e.pop(),this.geometryFunction_(e,i)):this.mode_===Wn&&(e[0].pop(),this.geometryFunction_(e,i),e=i.getCoordinates()),this.type_===pe?t.setGeometry(new Ai([e])):this.type_===fe?t.setGeometry(new bi([e])):this.type_===de&&t.setGeometry(new Vi([e])),this.dispatchEvent(new Hn(Un,t)),this.features_&&this.features_.push(t),this.source_&&this.source_.addFeature(t)}},e.prototype.abortDrawing_=function(){this.finishCoordinate_=null;var t=this.sketchFeature_;return t&&(this.sketchFeature_=null,this.sketchPoint_=null,this.sketchLine_=null,this.overlay_.getSource().clear(!0)),t},e.prototype.extend=function(t){var e=t.getGeometry();this.sketchFeature_=t,this.sketchCoords_=e.getCoordinates();var i=this.sketchCoords_[this.sketchCoords_.length-1];this.finishCoordinate_=i.slice(),this.sketchCoords_.push(i.slice()),this.updateSketchFeatures_(),this.dispatchEvent(new Hn(zn,this.sketchFeature_))},e.prototype.updateSketchFeatures_=function(){var t=[];this.sketchFeature_&&t.push(this.sketchFeature_),this.sketchLine_&&t.push(this.sketchLine_),this.sketchPoint_&&t.push(this.sketchPoint_);var e=this.overlay_.getSource();e.clear(!0),e.addFeatures(t)},e.prototype.updateState_=function(){var t=this.getMap(),e=this.getActive();t&&e||this.abortDrawing_(),this.overlay_.setMap(e?t:null)},e}(Ui),$n=Vue.resource("/api/v1/volumes{/id}/export-area"),Zn="modifystart",Qn="modifyend",tr=function(t){function e(e,i,n){t.call(this,e),this.features=i,this.mapBrowserEvent=n}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e}(q);function er(t,e){return t.index-e.index}function ir(t,e){var i,n,r=e.geometry;if(r.getType()===ge){var o=r;if(1===e.index){var s=Mt(o.getCenter(),t),a=Math.sqrt(s)-o.getRadius();return a*a}}return i=t,n=e.segment,Mt(i,St(i,n))}function nr(t,e){var i=e.geometry;return i.getType()===ge&&1===e.index||i.getType()===ve||i.getType()===ye?i.getClosestPoint(t):St(t,e.segment)}var rr=function(t){function e(e){var i,n;if(t.call(this,e),this.condition_=e.condition?e.condition:At,this.defaultDeleteCondition_=function(t){return function(t){var e=t.originalEvent;return e.altKey&&!(e.metaKey||e.ctrlKey)&&!e.shiftKey}(t)&&function(t){return t.type==vt}(t)},this.deleteCondition_=e.deleteCondition?e.deleteCondition:this.defaultDeleteCondition_,this.insertVertexCondition_=e.insertVertexCondition?e.insertVertexCondition:Pt,this.vertexFeature_=null,this.vertexSegments_=null,this.lastPixel_=[0,0],this.ignoreNextSingleClick_=!1,this.modified_=!1,this.rBush_=new Ln,this.pixelTolerance_=void 0!==e.pixelTolerance?e.pixelTolerance:10,this.snappedToVertex_=!1,this.changingFeature_=!1,this.dragSegments_=[],this.overlay_=new xn({source:new Yn({useSpatialIndex:!1,wrapX:!!e.wrapX}),style:e.style?e.style:(i=gn(),function(t,e){return i[he]}),updateWhileAnimating:!0,updateWhileInteracting:!0}),this.SEGMENT_WRITERS_={Point:this.writePointGeometry_,LineString:this.writeLineStringGeometry_,LinearRing:this.writeLineStringGeometry_,Polygon:this.writePolygonGeometry_,MultiPoint:this.writeMultiPointGeometry_,MultiLineString:this.writeMultiLineStringGeometry_,MultiPolygon:this.writeMultiPolygonGeometry_,Rectangle:this.writePolygonGeometry_,Ellipse:this.writePolygonGeometry_,Circle:this.writeCircleGeometry_,GeometryCollection:this.writeGeometryCollectionGeometry_},this.source_=null,e.source?(this.source_=e.source,n=new gt(this.source_.getFeatures()),O(this.source_,Rn,this.handleSourceAdd_,this),O(this.source_,An,this.handleSourceRemove_,this)):n=e.features,!n)throw new Error("The modify interaction requires features or a source");this.features_=n,this.features_.forEach(this.addFeature_.bind(this)),O(this.features_,st,this.handleFeatureAdd_,this),O(this.features_,at,this.handleFeatureRemove_,this),this.lastPointerEvent_=null}return t&&(e.__proto__=t),e.prototype=Object.create(t&&t.prototype),e.prototype.constructor=e,e.prototype.addFeature_=function(t){var e=t.getGeometry();e&&e.getType()in this.SEGMENT_WRITERS_&&this.SEGMENT_WRITERS_[e.getType()].call(this,t,e);var i=this.getMap();i&&i.isRendered()&&this.getActive()&&this.handlePointerAtPixel_(this.lastPixel_,i),O(t,W,this.handleFeatureChange_,this)},e.prototype.willModifyFeatures_=function(t){this.modified_||(this.modified_=!0,this.dispatchEvent(new tr(Zn,this.features_,t)))},e.prototype.removeFeature_=function(t){this.removeFeatureSegmentData_(t),this.vertexFeature_&&0===this.features_.getLength()&&(this.overlay_.getSource().removeFeature(this.vertexFeature_),this.vertexFeature_=null),T(t,W,this.handleFeatureChange_,this)},e.prototype.removeFeatureSegmentData_=function(t){var e=this.rBush_,i=[];e.forEach((function(e){t===e.feature&&i.push(e)}));for(var n=i.length-1;n>=0;--n)e.remove(i[n])},e.prototype.setActive=function(e){this.vertexFeature_&&!e&&(this.overlay_.getSource().removeFeature(this.vertexFeature_),this.vertexFeature_=null),t.prototype.setActive.call(this,e)},e.prototype.setMap=function(e){this.overlay_.setMap(e),t.prototype.setMap.call(this,e)},e.prototype.getOverlay=function(){return this.overlay_},e.prototype.handleSourceAdd_=function(t){t.feature&&this.features_.push(t.feature)},e.prototype.handleSourceRemove_=function(t){t.feature&&this.features_.remove(t.feature)},e.prototype.handleFeatureAdd_=function(t){this.addFeature_(t.element)},e.prototype.handleFeatureChange_=function(t){if(!this.changingFeature_){var e=t.target;this.removeFeature_(e),this.addFeature_(e)}},e.prototype.handleFeatureRemove_=function(t){var e=t.element;this.removeFeature_(e)},e.prototype.writePointGeometry_=function(t,e){var i=e.getCoordinates(),n={feature:t,geometry:e,segment:[i,i]};this.rBush_.insert(e.getExtent(),n)},e.prototype.writeMultiPointGeometry_=function(t,e){for(var i=e.getCoordinates(),n=0,r=i.length;n<r;++n){var o=i[n],s={feature:t,geometry:e,depth:[n],index:n,segment:[o,o]};this.rBush_.insert(e.getExtent(),s)}},e.prototype.writeLineStringGeometry_=function(t,e){for(var i=e.getCoordinates(),n=0,r=i.length-1;n<r;++n){var o=i.slice(n,n+2),s={feature:t,geometry:e,index:n,segment:o};this.rBush_.insert(jt(o),s)}},e.prototype.writeMultiLineStringGeometry_=function(t,e){for(var i=e.getCoordinates(),n=0,r=i.length;n<r;++n)for(var o=i[n],s=0,a=o.length-1;s<a;++s){var h=o.slice(s,s+2),u={feature:t,geometry:e,depth:[n],index:s,segment:h};this.rBush_.insert(jt(h),u)}},e.prototype.writePolygonGeometry_=function(t,e){for(var i=e.getCoordinates(),n=0,r=i.length;n<r;++n)for(var o=i[n],s=0,a=o.length-1;s<a;++s){var h=o.slice(s,s+2),u={feature:t,geometry:e,depth:[n],index:s,segment:h};this.rBush_.insert(jt(h),u)}},e.prototype.writeMultiPolygonGeometry_=function(t,e){for(var i=e.getCoordinates(),n=0,r=i.length;n<r;++n)for(var o=i[n],s=0,a=o.length;s<a;++s)for(var h=o[s],u=0,l=h.length-1;u<l;++u){var c=h.slice(u,u+2),p={feature:t,geometry:e,depth:[s,n],index:u,segment:c};this.rBush_.insert(jt(c),p)}},e.prototype.writeCircleGeometry_=function(t,e){var i=e.getCenter(),n={feature:t,geometry:e,index:0,segment:[i,i]},r={feature:t,geometry:e,index:1,segment:[i,i]},o=[n,r];n.featureSegments=r.featureSegments=o,this.rBush_.insert(Ut(i),n),this.rBush_.insert(e.getExtent(),r)},e.prototype.writeGeometryCollectionGeometry_=function(t,e){for(var i=e.getGeometriesArray(),n=0;n<i.length;++n)this.SEGMENT_WRITERS_[i[n].getType()].call(this,t,i[n])},e.prototype.createOrUpdateVertexFeature_=function(t){var e=this.vertexFeature_;e?e.getGeometry().setCoordinates(t):(e=new yt(new Di(t)),this.vertexFeature_=e,this.overlay_.getSource().addFeature(e));return e},e.prototype.handleEvent=function(e){return!e.pointerEvent||(this.lastPointerEvent_=e,e.map.getView().getInteracting()||e.type!=xt||this.handlingDownUpSequence||this.handlePointerMove_(e),this.vertexFeature_&&this.deleteCondition_(e)&&(i=!(e.type!=vt||!this.ignoreNextSingleClick_)||this.removePoint()),e.type==vt&&(this.ignoreNextSingleClick_=!1),t.prototype.handleEvent.call(this,e)&&!i);var i},e.prototype.handleDragEvent=function(t){this.ignoreNextSingleClick_=!1,this.willModifyFeatures_(t);for(var e=t.coordinate,i=0,n=this.dragSegments_.length;i<n;++i){for(var r=this.dragSegments_[i],o=r[0],s=o.depth,a=o.geometry,h=void 0,u=o.segment,l=r[1];e.length<a.getStride();)e.push(u[l][e.length]);switch(a.getType()){case he:h=e,u[0]=u[1]=e;break;case pe:(h=a.getCoordinates())[o.index]=e,u[0]=u[1]=e;break;case ue:(h=a.getCoordinates())[o.index+l]=e,u[l]=e;break;case fe:case ce:(h=a.getCoordinates())[s[0]][o.index+l]=e,u[l]=e;break;case de:(h=a.getCoordinates())[s[1]][s[0]][o.index+l]=e,u[l]=e;break;case ye:var c=(y=(h=a.getCoordinates())[s[0]])[(o.index+l+2)%4],p=[e[0]-c[0],e[1]-c[1]],f=[(v=y[(o.index+l+1)%4])[0]-c[0],v[1]-c[1]],d=Math.sqrt(f[0]*f[0]+f[1]*f[1]);f[0]=f[0]/d,f[1]=f[1]/d;var _=p[0]*f[0]+p[1]*f[1];y[(o.index+l+1)%4]=[c[0]+_*f[0],c[1]+_*f[1]];var g=[(m=y[(o.index+l+3)%4])[0]-c[0],m[1]-c[1]];d=Math.sqrt(g[0]*g[0]+g[1]*g[1]),g[0]=g[0]/d,g[1]=g[1]/d,_=p[0]*g[0]+p[1]*g[1],y[(o.index+l+3)%4]=[c[0]+_*g[0],c[1]+_*g[1]],y[o.index+l]=e;break;case ve:var y,v=(y=(h=a.getCoordinates())[s[0]])[(o.index+l+1)%4],m=(c=y[(o.index+l+2)%4],y[(o.index+l+3)%4]),C=Math.sqrt(Math.pow(v[0]-m[0],2)+Math.pow(v[1]-m[1],2))/2,x=[e[0]-c[0],e[1]-c[1]],w=[-1*x[1],x[0]];d=Math.sqrt(w[0]*w[0]+w[1]*w[1]);w[0]=w[0]/d,w[1]=w[1]/d;var E=[c[0]+x[0]/2,c[1]+x[1]/2];y[(o.index+l+1)%4]=[E[0]+w[0]*C,E[1]+w[1]*C],y[(o.index+l+3)%4]=[E[0]-w[0]*C,E[1]-w[1]*C],y[o.index+l]=e;break;case ge:u[0]=u[1]=e,0===o.index?(this.changingFeature_=!0,a.setCenter(e),this.changingFeature_=!1):(this.changingFeature_=!0,a.setRadius(It(a.getCenter(),e)),this.changingFeature_=!1)}h&&this.setGeometryCoordinates_(a,h)}this.createOrUpdateVertexFeature_(e)},e.prototype.handleDownEvent=function(t){if(!this.condition_(t))return!1;this.handlePointerAtPixel_(t.pixel,t.map);var e=t.map.getCoordinateFromPixel(t.pixel);this.dragSegments_.length=0,this.modified_=!1;var i=this.vertexFeature_;if(i){var n=[],r=i.getGeometry().getCoordinates(),s=jt([r]),a=this.rBush_.getInExtent(s),h={};a.sort(er);for(var u=0,l=a.length;u<l;++u){var c=a[u],p=c.segment,f=o(c.feature),d=c.depth;if(d&&(f+="-"+d.join("-")),h[f]||(h[f]=new Array(2)),c.geometry.getType()===ge&&1===c.index)kt(nr(e,c),r)&&!h[f][0]&&(this.dragSegments_.push([c,0]),h[f][0]=c);else if(kt(p[0],r)&&!h[f][0])this.dragSegments_.push([c,0]),h[f][0]=c;else if(kt(p[1],r)&&!h[f][1]){if((c.geometry.getType()===ue||c.geometry.getType()===fe)&&h[f][0]&&0===h[f][0].index)continue;this.dragSegments_.push([c,1]),h[f][1]=c}else this.insertVertexCondition_(t)&&o(p)in this.vertexSegments_&&!h[f][0]&&!h[f][1]&&n.push([c,r])}n.length&&this.willModifyFeatures_(t);for(var _=n.length-1;_>=0;--_)this.insertVertex_.apply(this,n[_])}return!!this.vertexFeature_},e.prototype.handleUpEvent=function(t){for(var e=this.dragSegments_.length-1;e>=0;--e){var i=this.dragSegments_[e][0],n=i.geometry;if(n.getType()===ye||n.getType()===ve)this.features_.remove(i.feature),this.features_.push(i.feature);else if(n.getType()===ge){var r=n.getCenter(),o=i.featureSegments[0],s=i.featureSegments[1];o.segment[0]=o.segment[1]=r,s.segment[0]=s.segment[1]=r,this.rBush_.update(Ut(r),o),this.rBush_.update(n.getExtent(),s)}else this.rBush_.update(jt(i.segment),i)}return this.modified_&&(this.dispatchEvent(new tr(Qn,this.features_,t)),this.modified_=!1),!1},e.prototype.handlePointerMove_=function(t){this.lastPixel_=t.pixel,this.handlePointerAtPixel_(t.pixel,t.map)},e.prototype.handlePointerAtPixel_=function(t,e){var i,n,r,s=e.getCoordinateFromPixel(t),a=(i=Ut(s),n=e.getView().getResolution()*this.pixelTolerance_,r?(r[0]=i[0]-n,r[1]=i[1]-n,r[2]=i[2]+n,r[3]=i[3]+n,r):[i[0]-n,i[1]-n,i[2]+n,i[3]+n]),h=this.rBush_.getInExtent(a);if(h.length>0){h.sort((function(t,e){return ir(s,t)-ir(s,e)}));var u=h[0],l=u.segment,c=nr(s,u),p=e.getPixelFromCoordinate(c),f=It(t,p);if(f<=this.pixelTolerance_){var d={};if(u.geometry.getType()===ge&&1===u.index)this.snappedToVertex_=!0,this.createOrUpdateVertexFeature_(c);else{var _=e.getPixelFromCoordinate(l[0]),g=e.getPixelFromCoordinate(l[1]),y=Mt(p,_),v=Mt(p,g);f=Math.sqrt(Math.min(y,v)),this.snappedToVertex_=f<=this.pixelTolerance_,this.snappedToVertex_&&(c=y>v?l[1]:l[0]),this.createOrUpdateVertexFeature_(c);for(var m=1,C=h.length;m<C;++m){var x=h[m].segment;if(!(kt(l[0],x[0])&&kt(l[1],x[1])||kt(l[0],x[1])&&kt(l[1],x[0])))break;d[o(x)]=!0}}return d[o(l)]=!0,void(this.vertexSegments_=d)}}this.vertexFeature_&&(this.overlay_.getSource().removeFeature(this.vertexFeature_),this.vertexFeature_=null)},e.prototype.insertVertex_=function(t,e){for(var i,n=t.segment,r=t.feature,o=t.geometry,s=t.depth,a=t.index;e.length<o.getStride();)e.push(0);switch(o.getType()){case fe:case ce:(i=o.getCoordinates())[s[0]].splice(a+1,0,e);break;case de:(i=o.getCoordinates())[s[1]][s[0]].splice(a+1,0,e);break;case ue:(i=o.getCoordinates()).splice(a+1,0,e);break;default:return}this.setGeometryCoordinates_(o,i);var h=this.rBush_;h.remove(t),this.updateSegmentIndices_(o,a,s,1);var u={segment:[n[0],e],feature:r,geometry:o,depth:s,index:a};h.insert(jt(u.segment),u),this.dragSegments_.push([u,1]);var l={segment:[e,n[1]],feature:r,geometry:o,depth:s,index:a+1};h.insert(jt(l.segment),l),this.dragSegments_.push([l,0]),this.ignoreNextSingleClick_=!0},e.prototype.removePoint=function(){if(this.lastPointerEvent_&&this.lastPointerEvent_.type!=Ct){var t=this.lastPointerEvent_;return this.willModifyFeatures_(t),this.removeVertex_(),this.dispatchEvent(new tr(Qn,this.features_,t)),this.modified_=!1,!0}return!1},e.prototype.removeVertex_=function(){var t,e,i,n,r,s,a,h,u,l,c,p=this.dragSegments_,f={},d=!1;for(r=p.length-1;r>=0;--r)c=o((l=(i=p[r])[0]).feature),l.depth&&(c+="-"+l.depth.join("-")),c in f||(f[c]={}),0===i[1]?(f[c].right=l,f[c].index=l.index):1==i[1]&&(f[c].left=l,f[c].index=l.index+1);for(c in f){switch(u=f[c].right,a=f[c].left,(h=(s=f[c].index)-1)<0&&(h=0),t=e=(n=(l=void 0!==a?a:u).geometry).getCoordinates(),d=!1,n.getType()){case fe:e[l.depth[0]].length>2&&(e[l.depth[0]].splice(s,1),d=!0);break;case ue:e.length>2&&(e.splice(s,1),d=!0);break;case de:t=t[l.depth[1]];case ce:(t=t[l.depth[0]]).length>4&&(s==t.length-1&&(s=0),t.splice(s,1),d=!0,0===s&&(t.pop(),t.push(t[0]),h=t.length-1))}if(d){this.setGeometryCoordinates_(n,e);var _=[];if(void 0!==a&&(this.rBush_.remove(a),_.push(a.segment[0])),void 0!==u&&(this.rBush_.remove(u),_.push(u.segment[1])),void 0!==a&&void 0!==u){var g={depth:l.depth,feature:l.feature,geometry:l.geometry,index:h,segment:_};this.rBush_.insert(jt(g.segment),g)}this.updateSegmentIndices_(n,s,l.depth,-1),this.vertexFeature_&&(this.overlay_.getSource().removeFeature(this.vertexFeature_),this.vertexFeature_=null),p.length=0}}return d},e.prototype.setGeometryCoordinates_=function(t,e){this.changingFeature_=!0,t.setCoordinates(e),this.changingFeature_=!1},e.prototype.updateSegmentIndices_=function(t,e,i,n){this.rBush_.forEachInExtent(t.getExtent(),(function(r){r.geometry===t&&(void 0===i||void 0===r.depth||function(t,e){var i=t.length;if(i!==e.length)return!1;for(var n=0;n<i;n++)if(t[n]!==e[n])return!1;return!0}(r.depth,i))&&r.index>e&&(r.index+=n)}))},e}(Ui),or=biigle.$require("events"),sr=biigle.$require("messages").handleErrorResponse,ar=biigle.$require("annotations.components.settingsTabPlugins"),hr={props:{settings:{type:Object,required:!0}},data:function(){return{opacityValue:"1",currentImage:null,isEditing:!1,exportArea:null,volumeId:null}},computed:{opacity:function(){return parseFloat(this.opacityValue)},shown:function(){return this.opacity>0},height:function(){return this.currentImage?this.currentImage.height:0},hasExportArea:function(){return null!==this.exportArea},layer:function(){return new xn({source:new Yn({features:new gt}),style:[new vn({stroke:new pn({color:"white",width:4}),image:new ot({radius:6,fill:new ln({color:"#666666"}),stroke:new pn({color:"white",width:2,lineDash:[2]})})}),new vn({stroke:new pn({color:"#666666",width:1,lineDash:[2]})})],zIndex:4,updateWhileAnimating:!0,updateWhileInteracting:!0})},drawInteraction:function(){return new Jn({source:this.layer.getSource(),type:"Rectangle",style:this.layer.getStyle(),minPoints:2,maxPoints:2,geometryFunction:function(t,e){t.length>1&&(t=[t[0],[t[0][0],t[1][1]],t[1],[t[1][0],t[0][1]]]);var i=e;return i?i.setCoordinates([t]):i=new jn([t]),i}})},modifyInteraction:function(){return new rr({features:this.layer.getSource().getFeaturesCollection(),style:this.layer.getStyle(),deleteCondition:Rt})}},methods:{toggleEditing:function(){this.isEditing=!this.isEditing,this.isEditing?(this.drawInteraction.setActive(!0),this.modifyInteraction.setActive(!0)):(this.drawInteraction.setActive(!1),this.modifyInteraction.setActive(!1))},deleteArea:function(){var t=this;if(this.hasExportArea&&confirm("Do you really want to delete the export area?")){var e=this.layer.getSource(),i=e.getFeatures()[0];e.clear(),$n.delete({id:this.volumeId}).then((function(){return t.exportArea=null})).catch((function(t){e.addFeature(i),sr(t)}))}},updateCurrentImage:function(t,e){this.currentImage=e},maybeDrawArea:function(){if(this.clearSource(),this.exportArea&&this.height>0){var t=this.currentImage.tiled?0:this.height,e=new jn([[[this.exportArea[0],t-this.exportArea[1]],[this.exportArea[0],t-this.exportArea[3]],[this.exportArea[2],t-this.exportArea[3]],[this.exportArea[2],t-this.exportArea[1]]]]);this.layer.getSource().addFeature(new yt({geometry:e}))}},handleModifyend:function(t){this.updateExportArea(t.features.item(0))},clearSource:function(){this.layer.getSource().clear()},handleDrawend:function(t){var e=this.layer.getSource(),i=e.getFeatures()[0];e.clear(),this.updateExportArea(t.feature).catch((function(){e.clear(),i&&e.addFeature(i)}))},updateExportArea:function(t){var e=this,i=t.getGeometry().getCoordinates()[0],n=this.currentImage.tiled?0:this.height;i=[i[0][0],n-i[0][1],i[2][0],n-i[2][1]].map(Math.round);var r=$n.save({id:this.volumeId},{coordinates:i}).then((function(){return e.exportArea=i}));return r.catch(sr),r},extendMap:function(t){t.addLayer(this.layer),t.addInteraction(this.drawInteraction),t.addInteraction(this.modifyInteraction)}},watch:{opacity:function(t){t<1?this.settings.set("exportAreaOpacity",t):this.settings.delete("exportAreaOpacity"),this.layer.setOpacity(t)},exportArea:function(){this.maybeDrawArea()},height:function(){this.maybeDrawArea()}},created:function(){this.volumeId=biigle.$require("annotations.volumeId"),this.settings.has("exportAreaOpacity")&&(this.opacityValue=this.settings.get("exportAreaOpacity")),this.exportArea=biigle.$require("annotations.exportArea"),this.drawInteraction.setActive(!1),this.modifyInteraction.setActive(!1),this.drawInteraction.on("drawend",this.handleDrawend),this.modifyInteraction.on("modifyend",this.handleModifyend),or.$on("images.change",this.updateCurrentImage),or.$on("annotations.map.init",this.extendMap)}},ur=i("KHd+"),lr=Object(ur.a)(hr,void 0,void 0,!1,null,null,null).exports;ar&&(ar.exportArea=lr)},"A1R+":function(t,e,i){"use strict";t.exports=r,t.exports.default=r;var n=i("YcpW");function r(t,e){if(!(this instanceof r))return new r(t,e);this._maxEntries=Math.max(4,t||9),this._minEntries=Math.max(2,Math.ceil(.4*this._maxEntries)),e&&this._initFormat(e),this.clear()}function o(t,e,i){if(!i)return e.indexOf(t);for(var n=0;n<e.length;n++)if(i(t,e[n]))return n;return-1}function s(t,e){a(t,0,t.children.length,e,t)}function a(t,e,i,n,r){r||(r=_(null)),r.minX=1/0,r.minY=1/0,r.maxX=-1/0,r.maxY=-1/0;for(var o,s=e;s<i;s++)o=t.children[s],h(r,t.leaf?n(o):o);return r}function h(t,e){return t.minX=Math.min(t.minX,e.minX),t.minY=Math.min(t.minY,e.minY),t.maxX=Math.max(t.maxX,e.maxX),t.maxY=Math.max(t.maxY,e.maxY),t}function u(t,e){return t.minX-e.minX}function l(t,e){return t.minY-e.minY}function c(t){return(t.maxX-t.minX)*(t.maxY-t.minY)}function p(t){return t.maxX-t.minX+(t.maxY-t.minY)}function f(t,e){return t.minX<=e.minX&&t.minY<=e.minY&&e.maxX<=t.maxX&&e.maxY<=t.maxY}function d(t,e){return e.minX<=t.maxX&&e.minY<=t.maxY&&e.maxX>=t.minX&&e.maxY>=t.minY}function _(t){return{children:t,height:1,leaf:!0,minX:1/0,minY:1/0,maxX:-1/0,maxY:-1/0}}function g(t,e,i,r,o){for(var s,a=[e,i];a.length;)(i=a.pop())-(e=a.pop())<=r||(s=e+Math.ceil((i-e)/r/2)*r,n(t,s,e,i,o),a.push(e,s,s,i))}r.prototype={all:function(){return this._all(this.data,[])},search:function(t){var e=this.data,i=[],n=this.toBBox;if(!d(t,e))return i;for(var r,o,s,a,h=[];e;){for(r=0,o=e.children.length;r<o;r++)s=e.children[r],d(t,a=e.leaf?n(s):s)&&(e.leaf?i.push(s):f(t,a)?this._all(s,i):h.push(s));e=h.pop()}return i},collides:function(t){var e=this.data,i=this.toBBox;if(!d(t,e))return!1;for(var n,r,o,s,a=[];e;){for(n=0,r=e.children.length;n<r;n++)if(o=e.children[n],d(t,s=e.leaf?i(o):o)){if(e.leaf||f(t,s))return!0;a.push(o)}e=a.pop()}return!1},load:function(t){if(!t||!t.length)return this;if(t.length<this._minEntries){for(var e=0,i=t.length;e<i;e++)this.insert(t[e]);return this}var n=this._build(t.slice(),0,t.length-1,0);if(this.data.children.length)if(this.data.height===n.height)this._splitRoot(this.data,n);else{if(this.data.height<n.height){var r=this.data;this.data=n,n=r}this._insert(n,this.data.height-n.height-1,!0)}else this.data=n;return this},insert:function(t){return t&&this._insert(t,this.data.height-1),this},clear:function(){return this.data=_([]),this},remove:function(t,e){if(!t)return this;for(var i,n,r,s,a=this.data,h=this.toBBox(t),u=[],l=[];a||u.length;){if(a||(a=u.pop(),n=u[u.length-1],i=l.pop(),s=!0),a.leaf&&-1!==(r=o(t,a.children,e)))return a.children.splice(r,1),u.push(a),this._condense(u),this;s||a.leaf||!f(a,h)?n?(i++,a=n.children[i],s=!1):a=null:(u.push(a),l.push(i),i=0,n=a,a=a.children[0])}return this},toBBox:function(t){return t},compareMinX:u,compareMinY:l,toJSON:function(){return this.data},fromJSON:function(t){return this.data=t,this},_all:function(t,e){for(var i=[];t;)t.leaf?e.push.apply(e,t.children):i.push.apply(i,t.children),t=i.pop();return e},_build:function(t,e,i,n){var r,o=i-e+1,a=this._maxEntries;if(o<=a)return s(r=_(t.slice(e,i+1)),this.toBBox),r;n||(n=Math.ceil(Math.log(o)/Math.log(a)),a=Math.ceil(o/Math.pow(a,n-1))),(r=_([])).leaf=!1,r.height=n;var h,u,l,c,p=Math.ceil(o/a),f=p*Math.ceil(Math.sqrt(a));for(g(t,e,i,f,this.compareMinX),h=e;h<=i;h+=f)for(g(t,h,l=Math.min(h+f-1,i),p,this.compareMinY),u=h;u<=l;u+=p)c=Math.min(u+p-1,l),r.children.push(this._build(t,u,c,n-1));return s(r,this.toBBox),r},_chooseSubtree:function(t,e,i,n){for(var r,o,s,a,h,u,l,p,f,d;n.push(e),!e.leaf&&n.length-1!==i;){for(l=p=1/0,r=0,o=e.children.length;r<o;r++)h=c(s=e.children[r]),f=t,d=s,(u=(Math.max(d.maxX,f.maxX)-Math.min(d.minX,f.minX))*(Math.max(d.maxY,f.maxY)-Math.min(d.minY,f.minY))-h)<p?(p=u,l=h<l?h:l,a=s):u===p&&h<l&&(l=h,a=s);e=a||e.children[0]}return e},_insert:function(t,e,i){var n=this.toBBox,r=i?t:n(t),o=[],s=this._chooseSubtree(r,this.data,e,o);for(s.children.push(t),h(s,r);e>=0&&o[e].children.length>this._maxEntries;)this._split(o,e),e--;this._adjustParentBBoxes(r,o,e)},_split:function(t,e){var i=t[e],n=i.children.length,r=this._minEntries;this._chooseSplitAxis(i,r,n);var o=this._chooseSplitIndex(i,r,n),a=_(i.children.splice(o,i.children.length-o));a.height=i.height,a.leaf=i.leaf,s(i,this.toBBox),s(a,this.toBBox),e?t[e-1].children.push(a):this._splitRoot(i,a)},_splitRoot:function(t,e){this.data=_([t,e]),this.data.height=t.height+1,this.data.leaf=!1,s(this.data,this.toBBox)},_chooseSplitIndex:function(t,e,i){var n,r,o,s,h,u,l,p,f,d,_,g,y,v;for(u=l=1/0,n=e;n<=i-e;n++)r=a(t,0,n,this.toBBox),o=a(t,n,i,this.toBBox),f=r,d=o,_=void 0,g=void 0,y=void 0,v=void 0,_=Math.max(f.minX,d.minX),g=Math.max(f.minY,d.minY),y=Math.min(f.maxX,d.maxX),v=Math.min(f.maxY,d.maxY),s=Math.max(0,y-_)*Math.max(0,v-g),h=c(r)+c(o),s<u?(u=s,p=n,l=h<l?h:l):s===u&&h<l&&(l=h,p=n);return p},_chooseSplitAxis:function(t,e,i){var n=t.leaf?this.compareMinX:u,r=t.leaf?this.compareMinY:l;this._allDistMargin(t,e,i,n)<this._allDistMargin(t,e,i,r)&&t.children.sort(n)},_allDistMargin:function(t,e,i,n){t.children.sort(n);var r,o,s=this.toBBox,u=a(t,0,e,s),l=a(t,i-e,i,s),c=p(u)+p(l);for(r=e;r<i-e;r++)o=t.children[r],h(u,t.leaf?s(o):o),c+=p(u);for(r=i-e-1;r>=e;r--)o=t.children[r],h(l,t.leaf?s(o):o),c+=p(l);return c},_adjustParentBBoxes:function(t,e,i){for(var n=i;n>=0;n--)h(e[n],t)},_condense:function(t){for(var e,i=t.length-1;i>=0;i--)0===t[i].children.length?i>0?(e=t[i-1].children).splice(e.indexOf(t[i]),1):this.clear():s(t[i],this.toBBox)},_initFormat:function(t){var e=["return a"," - b",";"];this.compareMinX=new Function("a","b",e.join(t[0])),this.compareMinY=new Function("a","b",e.join(t[1])),this.toBBox=new Function("a","return {minX: a"+t[0]+", minY: a"+t[1]+", maxX: a"+t[2]+", maxY: a"+t[3]+"};")}}},"KHd+":function(t,e,i){"use strict";function n(t,e,i,n,r,o,s,a){var h,u="function"==typeof t?t.options:t;if(e&&(u.render=e,u.staticRenderFns=i,u._compiled=!0),n&&(u.functional=!0),o&&(u._scopeId="data-v-"+o),s?(h=function(t){(t=t||this.$vnode&&this.$vnode.ssrContext||this.parent&&this.parent.$vnode&&this.parent.$vnode.ssrContext)||"undefined"==typeof __VUE_SSR_CONTEXT__||(t=__VUE_SSR_CONTEXT__),r&&r.call(this,t),t&&t._registeredComponents&&t._registeredComponents.add(s)},u._ssrRegister=h):r&&(h=a?function(){r.call(this,(u.functional?this.parent:this).$root.$options.shadowRoot)}:r),h)if(u.functional){u._injectStyles=h;var l=u.render;u.render=function(t,e){return h.call(e),l(t,e)}}else{var c=u.beforeCreate;u.beforeCreate=c?[].concat(c,h):[h]}return{exports:t,options:u}}i.d(e,"a",(function(){return n}))},YcpW:function(t,e,i){t.exports=function(){"use strict";function t(t,e,i){var n=t[e];t[e]=t[i],t[i]=n}function e(t,e){return t<e?-1:t>e?1:0}return function(i,n,r,o,s){!function e(i,n,r,o,s){for(;o>r;){if(o-r>600){var a=o-r+1,h=n-r+1,u=Math.log(a),l=.5*Math.exp(2*u/3),c=.5*Math.sqrt(u*l*(a-l)/a)*(h-a/2<0?-1:1),p=Math.max(r,Math.floor(n-h*l/a+c)),f=Math.min(o,Math.floor(n+(a-h)*l/a+c));e(i,n,p,f,s)}var d=i[n],_=r,g=o;for(t(i,r,n),s(i[o],d)>0&&t(i,r,o);_<g;){for(t(i,_,g),_++,g--;s(i[_],d)<0;)_++;for(;s(i[g],d)>0;)g--}0===s(i[r],d)?t(i,r,g):(g++,t(i,g,o)),g<=n&&(r=g+1),n<=g&&(o=g-1)}}(i,n,r||0,o||i.length-1,s||e)}}()}});
