+++ src/public/assets/scripts/main.js
@@ -1,1 +1,1 @@
-!function(e){var t={};function r(s){if(t[s])return t[s].exports;var i=t[s]={i:s,l:!1,exports:{}};return e[s].call(i.exports,i,i.exports,r),i.l=!0,i.exports}r.m=e,r.c=t,r.d=function(e,t,s){r.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:s})},r.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},r.t=function(e,t){if(1&t&&(e=r(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var s=Object.create(null);if(r.r(s),Object.defineProperty(s,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var i in e)r.d(s,i,function(t){return e[t]}.bind(null,i));return s},r.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return r.d(t,"a",t),t},r.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},r.p="/",r(r.s=0)}({0:function(e,t,r){r("Or9D"),e.exports=r("zcrr")},Or9D:function(e,t,r){"use strict";r.r(t);var s=biigle.$require("messages").handleErrorResponse,i=biigle.$require("labelTrees.components.labelTrees"),n={mixins:[biigle.$require("core.mixins.loader")],components:{labelTrees:i},data:{allowedOptions:{},selectedType:"",selectedVariant:"",reportTypes:[],labelTrees:[],hasOnlyLabels:!1,success:!1,errors:{},options:{export_area:!1,newest_label:!1,separate_label_trees:!1,only_labels:[],aggregate_child_labels:!1}},computed:{flatLabels:function(){var e=[];return this.labelTrees.forEach((function(t){Array.prototype.push.apply(e,t.labels)})),e},selectedLabels:function(){return this.flatLabels.filter((function(e){return e.selected}))},selectedLabelsCount:function(){return this.selectedLabels.length},variants:function(){var e={};return this.reportTypes.forEach((function(t){var r=t.name.split("\\");e.hasOwnProperty(r[0])||(e[r[0]]=[]),e[r[0]].push(r[1])})),e},availableReportTypes:function(){var e={};return this.reportTypes.forEach((function(t){e[t.name]=t.id})),e},selectedReportTypeId:function(){return this.availableReportTypes[this.selectedType+"\\"+this.selectedVariant]},availableVariants:function(){return this.variants[this.selectedType]},onlyOneAvailableVariant:function(){return 1===this.availableVariants.length},selectedOptions:function(){var e=this,t={};return this.allowedOptions[this.selectedType].forEach((function(r){t[r]=e.options[r]})),t.type_id=this.selectedReportTypeId,t}},methods:{request:function(e,t){this.loading||(this.success=!1,this.startLoading(),t.save({id:e},this.selectedOptions).then(this.submitted,this.handleError).finally(this.finishLoading))},submitted:function(){this.success=!0,this.errors={}},handleError:function(e){422===e.status?this.errors=e.data:s(e)},selectType:function(e){this.selectedType=e,-1===this.availableVariants.indexOf(this.selectedVariant)&&(this.selectedVariant=this.availableVariants[0])},wantsType:function(e){return this.selectedType===e},wantsVariant:function(e){return Array.isArray(e)?-1!==e.indexOf(this.selectedVariant):this.selectedVariant===e},hasError:function(e){return this.errors.hasOwnProperty(e)},getError:function(e){return this.errors[e]?this.errors[e].join(" "):""},wantsCombination:function(e,t){return this.wantsType(e)&&this.wantsVariant(t)}},watch:{selectedLabels:function(e){this.options.only_labels=e.map((function(e){return e.id}))},hasOnlyLabels:function(e){e||this.flatLabels.forEach((function(e){e.selected=!1}))}},created:function(){this.reportTypes=biigle.$require("reports.reportTypes"),this.selectedType=Object.keys(this.variants)[0],this.selectedVariant=this.availableVariants[0],this.labelTrees=biigle.$require("reports.labelTrees")}},a=Vue.resource("/api/v1/projects{/id}/reports"),o={mixins:[n],data:{projectId:null,allowedOptions:{Annotations:["export_area","newest_label","separate_label_trees","only_labels","aggregate_child_labels"],ImageLabels:["separate_label_trees","only_labels"],VideoAnnotations:["separate_label_trees","only_labels"]}},methods:{submit:function(){this.request(this.projectId,a)}},created:function(){this.projectId=biigle.$require("reports.projectId")}},l=Vue.resource("/api/v1/videos{/id}/reports"),u={mixins:[n],data:{videoId:null,selectedType:"VideoAnnotations",selectedVariant:"Csv",allowedOptions:{VideoAnnotations:["separate_label_trees","only_labels"]}},methods:{submit:function(){this.request(this.videoId,l)}},created:function(){this.videoId=biigle.$require("reports.videoId")}},c=Vue.resource("/api/v1/volumes{/id}/reports"),d={mixins:[n],data:{allowedOptions:{Annotations:["export_area","newest_label","separate_label_trees","annotation_session_id","only_labels","aggregate_child_labels"],ImageLabels:["separate_label_trees","annotation_session_id","only_labels"]},options:{annotation_session_id:null}},methods:{submit:function(){this.request(this.volumeId,c)}},created:function(){this.volumeId=biigle.$require("reports.volumeId")}};biigle.$mount("project-report-form",o),biigle.$mount("video-report-form",u),biigle.$mount("volume-report-form",d)},zcrr:function(e,t){}});
+!function(e){var t={};function n(r){if(t[r])return t[r].exports;var i=t[r]={i:r,l:!1,exports:{}};return e[r].call(i.exports,i,i.exports,n),i.l=!0,i.exports}n.m=e,n.c=t,n.d=function(e,t,r){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:r})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var i in e)n.d(r,i,function(t){return e[t]}.bind(null,i));return r},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="/",n(n.s=0)}({0:function(e,t,n){n("Or9D"),e.exports=n("zcrr")},"KHd+":function(e,t,n){"use strict";function r(e,t,n,r,i,s,o,a){var l,u="function"==typeof e?e.options:e;if(t&&(u.render=t,u.staticRenderFns=n,u._compiled=!0),r&&(u.functional=!0),s&&(u._scopeId="data-v-"+s),o?(l=function(e){(e=e||this.$vnode&&this.$vnode.ssrContext||this.parent&&this.parent.$vnode&&this.parent.$vnode.ssrContext)||"undefined"==typeof __VUE_SSR_CONTEXT__||(e=__VUE_SSR_CONTEXT__),i&&i.call(this,e),e&&e._registeredComponents&&e._registeredComponents.add(o)},u._ssrRegister=l):i&&(l=a?function(){i.call(this,(u.functional?this.parent:this).$root.$options.shadowRoot)}:i),l)if(u.functional){u._injectStyles=l;var c=u.render;u.render=function(e,t){return l.call(t),c(e,t)}}else{var d=u.beforeCreate;u.beforeCreate=d?[].concat(d,l):[l]}return{exports:e,options:u}}n.d(t,"a",(function(){return r}))},Or9D:function(e,t,n){"use strict";n.r(t);var r=biigle.$require("messages").handleErrorResponse,i=biigle.$require("labelTrees.components.labelTrees"),s={mixins:[biigle.$require("core.mixins.loader")],components:{labelTrees:i},data:function(){return{allowedOptions:{},selectedType:"",selectedVariant:"",reportTypes:[],labelTrees:[],hasOnlyLabels:!1,success:!1,errors:{},options:{export_area:!1,newest_label:!1,separate_label_trees:!1,only_labels:[],aggregate_child_labels:!1}}},computed:{flatLabels:function(){var e=[];return this.labelTrees.forEach((function(t){Array.prototype.push.apply(e,t.labels)})),e},selectedLabels:function(){return this.flatLabels.filter((function(e){return e.selected}))},selectedLabelsCount:function(){return this.selectedLabels.length},variants:function(){var e={};return this.reportTypes.forEach((function(t){var n=t.name.split("\\");e.hasOwnProperty(n[0])||(e[n[0]]=[]),e[n[0]].push(n[1])})),e},availableReportTypes:function(){var e={};return this.reportTypes.forEach((function(t){e[t.name]=t.id})),e},selectedReportTypeId:function(){return this.availableReportTypes[this.selectedType+"\\"+this.selectedVariant]},availableVariants:function(){return this.variants[this.selectedType]},onlyOneAvailableVariant:function(){return 1===this.availableVariants.length},selectedOptions:function(){var e=this,t={};return this.allowedOptions[this.selectedType].forEach((function(n){t[n]=e.options[n]})),t.type_id=this.selectedReportTypeId,t}},methods:{request:function(e,t){this.loading||(this.success=!1,this.startLoading(),t.save({id:e},this.selectedOptions).then(this.submitted,this.handleError).finally(this.finishLoading))},submitted:function(){this.success=!0,this.errors={}},handleError:function(e){422===e.status?this.errors=e.data:r(e)},selectType:function(e){this.selectedType=e,-1===this.availableVariants.indexOf(this.selectedVariant)&&(this.selectedVariant=this.availableVariants[0])},wantsType:function(e){return this.selectedType===e},wantsVariant:function(e){return Array.isArray(e)?-1!==e.indexOf(this.selectedVariant):this.selectedVariant===e},hasError:function(e){return this.errors.hasOwnProperty(e)},getError:function(e){return this.errors[e]?this.errors[e].join(" "):""},wantsCombination:function(e,t){return this.wantsType(e)&&this.wantsVariant(t)}},watch:{selectedLabels:function(e){this.options.only_labels=e.map((function(e){return e.id}))},hasOnlyLabels:function(e){e||this.flatLabels.forEach((function(e){e.selected=!1}))}},created:function(){this.reportTypes=biigle.$require("reports.reportTypes"),this.selectedType=Object.keys(this.variants)[0],this.selectedVariant=this.availableVariants[0],this.labelTrees=biigle.$require("reports.labelTrees")}},o=n("KHd+"),a=Object(o.a)(s,void 0,void 0,!1,null,null,null).exports,l=Vue.resource("/api/v1/projects{/id}/reports"),u={mixins:[a],data:function(){return{projectId:null,allowedOptions:{Annotations:["export_area","newest_label","separate_label_trees","only_labels","aggregate_child_labels"],ImageLabels:["separate_label_trees","only_labels"],VideoAnnotations:["separate_label_trees","only_labels"]}}},methods:{submit:function(){this.request(this.projectId,l)}},created:function(){this.projectId=biigle.$require("reports.projectId")}},c=Object(o.a)(u,void 0,void 0,!1,null,null,null).exports,d=Vue.resource("/api/v1/videos{/id}/reports"),p={mixins:[a],data:function(){return{videoId:null,selectedType:"VideoAnnotations",selectedVariant:"Csv",allowedOptions:{VideoAnnotations:["separate_label_trees","only_labels"]}}},methods:{submit:function(){this.request(this.videoId,d)}},created:function(){this.videoId=biigle.$require("reports.videoId")}},f=Object(o.a)(p,void 0,void 0,!1,null,null,null).exports,h=Vue.resource("/api/v1/volumes{/id}/reports"),b={mixins:[a],data:function(){return{allowedOptions:{Annotations:["export_area","newest_label","separate_label_trees","annotation_session_id","only_labels","aggregate_child_labels"],ImageLabels:["separate_label_trees","annotation_session_id","only_labels"]},options:{annotation_session_id:null}}},methods:{submit:function(){this.request(this.volumeId,h)}},created:function(){this.volumeId=biigle.$require("reports.volumeId")}},_=Object(o.a)(b,void 0,void 0,!1,null,null,null).exports;biigle.$mount("project-report-form",c),biigle.$mount("video-report-form",f),biigle.$mount("volume-report-form",_)},zcrr:function(e,t){}});
