+++ src/public/mix-manifest.json
@@ -1,5 +1,5 @@
 {
-    "/assets/scripts/annotations.js": "/assets/scripts/annotations.js?id=5d076cdaafa31ad5b50d",
-    "/assets/scripts/main.js": "/assets/scripts/main.js?id=0b411999e5e7c5498cbf",
+    "/assets/scripts/annotations.js": "/assets/scripts/annotations.js?id=30a9a1a72b6179458798",
+    "/assets/scripts/main.js": "/assets/scripts/main.js?id=8b01b655e6b98fe7e685",
     "/assets/styles/main.css": "/assets/styles/main.css?id=9d45536e01d0a0b3c70b"
 }
