+++ src/resources/assets/js/annotations/annotations.js
@@ -1,1 +1,1 @@
-import './settingsTabPlugin';
+import './settingsTabPlugins';
